% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factset.R
\name{check_schema}
\alias{check_schema}
\title{Check Schema}
\usage{
check_schema(schema = NULL)
}
\arguments{
\item{schema}{is a list containing the entity and datetime column information and optionally the data types for the dataframe}
}
\description{
Check Schema
}

