\name{EKs}
\alias{EKs}
\docType{data}
\title{
Swiss forest vegetation data base 1972, site information
}
\description{
Swiss forest vegetation data base 1972, site information. Vegetation data is in data frame EKv.
}
\usage{data(EKs)}
\format{
  A data frame with 2533 observations on the following 11 variables.
  \describe{
    \item{\code{Autor}}{a factor with levels \code{AMBERG_} \code{ANTONI_} \code{AUER_C_} \code{BARDET_} \code{BEGER} \code{BIJLEV_} \code{BRAUN-_} \code{CLOSUI_} \code{DUTOIT_} \code{ELLENB_} \code{ESKUCH_} \code{ETTER} \code{FREHNE_} \code{FRITSC_} \code{FURRER_} \code{GAMS_H_} \code{GLAVAC_} \code{GRUENI_} \code{HEINIS_} \code{HELLER_} \code{HOEHN} \code{KLOETZ_} \code{KOCH_W_} \code{KUHN_N_} \code{KUOCH} \code{KURTH} \code{LINKOL_} \code{LUEDI} \code{MEYER} \code{MOOR_M_} \code{PALLMA_} \code{REHDER_} \code{RICHAR_} \code{RUEBEL_} \code{SCHLAE_} \code{SCHNYD_} \code{SIMAK} \code{STAMM} \code{TREPP} \code{VILLAR_} \code{VOLK_O_} \code{WINTEL_} \code{WIRZ-L_} \code{ZELLER_} \code{ZOLLER_}}
    \item{\code{Jahr}}{a numeric vector}
    \item{\code{Tabellennr.}}{a numeric vector}
    \item{\code{Laufnr.}}{a numeric vector}
    \item{\code{Gesellschaftsname}}{a factor with levels \code{ABIETETUM_ALBAE_} \code{ABIETETUM_ELYMETOSUM} \code{ABIETETUM_FESTUCETOS} \code{ABIETETUM_GALIETOSUM} \code{ABIETETUM_ISOTHECIET} \code{ABIETETUM_MELAMPYRET} \code{ABIETI-FAGETUM} \code{ABIETI-FAGETUM_DENTA} \code{ABIETI-FAGETUM_LUZUL} \code{ABIETI-FAGETUM_PETAS} \code{ABIETI-FAGETUM_VALER} \code{ABIETO-FAGETUM} \code{ABIETO-FAGETUM_ADENO} \code{ABIETO-FAGETUM_ALLIE} \code{ABIETO-FAGETUM_ELYME} \code{ABIETO-FAGETUM_FESTU} \code{ABIETO-FAGETUM_LUZUL} \code{ACER_CAMPSTRE-MISCHW} \code{ACER_OPALUS-REICHER_} \code{ACER_PSEUDOPLATANUS-} \code{ACERETO-FAGETUM_} \code{ACERETO-FAGETUM_ARUN} \code{ACERETO-FAGETUM_CIRS} \code{ACERETO-FAGETUM_PREN} \code{ACERETO-FAGETUM_RUMI} \code{ACERETO-FRAXINETUM} \code{ACERETO-FRAXINETUM_A} \code{ACERETO-FRAXINETUM_C} \code{ACERETUM} \code{ACERETUM_PSEUDOPLATA} \code{ACERI-FRAXINETUM} \code{ACERI-FRAXINETUM_ALL} \code{ACERI-FRAXINETUM_VER} \code{ADENOSTYLO-ABIETETUM} \code{AHORN-ESCHENWALD} \code{ALNETUM_ACERETOSUM} \code{ALNETUM_INCANAE_} \code{ARUNCO-ACERETUM_} \code{ASPLENIO-PICEETUM_} \code{ASPLENIO-PICEETUM_CA} \code{ASPLENIO-PICEETUM_TY} \code{ASTRAGALO-PINETUM_} \code{BERGBUSCHWALD_} \code{BRAUNERDE-BUCHENWALD} \code{BUCHEN-HANGFOEHRENWA} \code{BUCHENWALD} \code{BUXETUM_} \code{CALAMAGROSTI-ALNETUM} \code{CALLUNO-PINETUM_} \code{CALLUNO-PINETUM_PICE} \code{CARICETO-FAGETUM} \code{CARICI_ELONGATAE-ALN} \code{CARICI_REMOTAE-FRAXI} \code{CARICI-FAGETUM} \code{CARICI-FAGETUM_CARIC} \code{CARICI-FAGETUM_FINIC} \code{CARICI-FRAXINETUM_CH} \code{CHAMAEBUXO-PINETUM_P} \code{CHENAIE_A_CHARMES_FO} \code{EQUISETO-ABIETETUM} \code{EQUISETO-ABIETETUM_H} \code{EQUISETO-ALNETUM_INC} \code{EQUISETUM_SILVATICUM} \code{ERICETO-PINETUM_} \code{ERICETO-PINETUM_CARI} \code{ERICETO-PINETUM_PEUC} \code{ERICETO-PINETUM_SILV} \code{ERICETO-PINETUM_VACC} \code{ERICO-PINETUM_} \code{ERICO-PINETUM_FAGETO} \code{ERICO-PINETUM_SILVES} \code{ERISITHALO-ULMETUM_A} \code{ERISITHALO-ULMETUM_F} \code{ERLENWALD_} \code{ESCHEN-SCHWARZERLENW} \code{EUPHRASIETO-PINETUM_} \code{FAGETUM_} \code{FAGETUM_ABIETOSUM_} \code{FAGETUM_ADENOSTYLETO} \code{FAGETUM_CARICETOSUM_} \code{FAGETUM_DENTARIETOSU} \code{FAGETUM_FINICOLA} \code{FAGETUM_MAJANTHEMETO} \code{FAGETUM_PRAEALPINO-J} \code{FAGETUM_SESLERIETOSU} \code{FAGETUM_SILVATICAE} \code{FAGETUM_SILVATICAE_A} \code{FAGETUM_SILVATICAE_P} \code{FAGETUM_SILVATICAE_T} \code{FAGION_SILVATICAE_} \code{FELSENARVENWALD_} \code{FICHTENEINPFLANZUNGE} \code{FICHTENKULTURBESTAND} \code{FOEHREN-ASSOZIATION_} \code{FOEHRENWALD_} \code{FRAXINETUM} \code{FRAXINETUM_ACERETOSU} \code{FRAXINO-ULMETUM_} \code{FRISCHER_BUCHENMISCH} \code{HANG-AHORN-ESCHENWAL} \code{HANGAHORNESCHENWALD_} \code{HANGBUCHENWALD} \code{HELLEBORO-ORNETUM_AS} \code{HELLEBORO-ORNETUM_GA} \code{HELLEBORO-ORNETUM_TY} \code{KALK-BUCHENWALD_} \code{KEIN_SYNTAXON_} \code{LA_FORET_DES_PENTES_} \code{LAERCHENWALD} \code{LARICETUM_CEMBRETOSU} \code{LATHYRO-QUERCETUM_CA} \code{LATHYRO-QUERCETUM_LU} \code{LINDEN-BUCHENWALD_} \code{LINDENMISCHBESTAND_(} \code{LUZULO-FAGETUM} \code{LUZULO-FAGION_} \code{LYCOPODIO-BETULETUM_} \code{LYCOPODIO-MUGETUM_TY} \code{MACROPHORBIO-ALNETUM} \code{MASTIGOBRYETO-PICEET} \code{MEHLBEERENHANGBUCHEN} \code{MELAMPYRO-FAGETUM_} \code{MELAMPYRO-FAGETUM_AS} \code{MELAMPYRO-FAGETUM_BL} \code{MELAMPYRO-FAGETUM_LE} \code{MELAMPYRO-FAGETUM_LU} \code{MELAMPYRO-FAGETUM_TY} \code{MELICO-FAGETUM} \code{MELICO-FAGETUM_ASPER} \code{MELICO-FAGETUM_BLECH} \code{MELICO-FAGETUM_CARIC} \code{MELICO-FAGETUM_CORNE} \code{MELICO-FAGETUM_LUZUL} \code{MELICO-FAGETUM_PULMO} \code{MILIO-FAGETUM_} \code{MILIO-FAGETUM_BLECHN} \code{MILIO-FAGETUM_DRYOPT} \code{MILIO-FAGETUM_FESTUC} \code{MILIO-FAGETUM_LUZULE} \code{MILIO-FAGETUM_MAJANT} \code{MILIO-FAGETUM_VAR._E} \code{MOLINIETO_LITORALIS-} \code{MOLINIO_LITORALIS-PI} \code{MUGET0-ERICETOSUM_CA} \code{MUGETO-ERICETOSUM_HY} \code{MUGETO-ERICETUM_} \code{MUGETO-ERICETUMCLA} \code{MUGETO-RHODORETUM_HI} \code{MYRTILLO-ABIETETUM} \code{MYRTILLO-ABIETETUM_P} \code{MYRTILLO-ABIETETUM_T} \code{NORDLAND-BUCHENWALD_} \code{OBERER_HANGWALD_} \code{ONONIDO-PINETUM_} \code{OSTRYA-NIEDERWALD_} \code{PFEIFENGRAS-FOEHREN-} \code{PFEIFENGRAS-FOEHRENW} \code{PFEIFENGRAS-HANGFOEH} \code{PHEGOPTERI-PICEETUM_} \code{PHYLLITIDO-ACERETUM_} \code{PICEETUM} \code{PICEETUM_EXCELSAE_} \code{PICEETUM_EXCELSAE_MY} \code{PICEETUM_FILICETOSUM} \code{PICEETUM_MONTANUM_CA} \code{PICEETUM_MONTANUM_GA} \code{PICEETUM_MONTANUM_ME} \code{PICEETUM_MONTANUM_SU} \code{PICEETUM_SUBALPINUM_} \code{PICEETUM_TRANSALPINU} \code{PICEO_EXCELSAE-PINET} \code{PICETO-FAGETUM} \code{PICION_EXCELSAE_} \code{PINETO-CARICETUM_HUM} \code{PINETO-CYTISETUM} \code{PINETUM_} \code{PINETUM_SILICICOLUM_} \code{PINETUM_SILVESTRIS} \code{PINETUM_SILVESTRIS_A} \code{PRUNO-FRAXINETUM} \code{PRUNO-FRAXINETUM_EQU} \code{PRUNO-FRAXINETUM_TYP} \code{PULMONARIO-FAGETUM} \code{PULMONARIO-FAGETUM_A} \code{PULMONARIO-FAGETUM_C} \code{PULMONARIO-FAGETUM_T} \code{PYROLO-PINETUM} \code{QUERCETO-CARPINETUM_} \code{QUERCETO-LITHOSPERME} \code{QUERCETUM_SESSILIFLO} \code{QUERCO-ABIETETUM} \code{QUERCO-ABIETETUM_LUZ} \code{QUERCO-ABIETETUM_SPH} \code{QUERCO-BETULETUM} \code{QUERCO-CARPINETUM_} \code{QUERCO-CARPINETUM_AE} \code{QUERCO-CARPINETUM_AR} \code{QUERCO-CARPINETUM_MO} \code{QUERCO-FRAXINETUM_SA} \code{QUERCO-FRAXINETUM_TY} \code{RHODODENDRO_HIRSUTI-} \code{RHODODENDRO-ABIETETU} \code{RHODORETO-VACCINIETU} \code{SALICETUM_ALBO-FRAGI} \code{SALICI-BETULETUM_HYL} \code{SESLERIETO-FAGETUM_A} \code{SESLERIETO-FAGETUM_H} \code{SORBETO-ACERETUM_DRY} \code{SORBETO-ACERETUM_MEL} \code{SPHAGNO-MUGETUM_} \code{SPHAGNO-MUGETUM_BETU} \code{SPHAGNO-MUGETUM_ERIO} \code{SPHAGNO-MUGETUM_JUNI} \code{SPHAGNO-PICEETUM} \code{SPHAGNO-PICEETUM_BET} \code{SPHAGNO-PICEETUM_BLE} \code{SPHAGNO-PICEETUM_THU} \code{STIELEICHEN-HAGEBUCH} \code{TAXETO-FAGETUM} \code{TAXETO-FAGETUM_FESTU} \code{TAXETO-FAGETUM_FRAXI} \code{TILIETO-ASPERULETUM_} \code{TILIETO-FAGETUM_} \code{TILIO-ACERETUM} \code{TILIO-FAGETUM_} \code{TRAUBENEICHEN-BUCHEN} \code{TYP._MEHLBEEREN-HANG} \code{UEBERGANG_QUER.CARP.} \code{UNTERER_HANGBUCHENWA} \code{VACCINIO-PINETUM} \code{WEISSTANNENWALD_}}
    \item{\code{Hoehe_u.M.}}{a numeric vector}
    \item{\code{Neigung_in_Prozent}}{a numeric vector}
    \item{\code{Exposition}}{a factor with levels \code{_?} \code{_E} \code{_ENE} \code{_ESE} \code{_N} \code{_NNE} \code{_NNW} \code{_S} \code{_SSE} \code{_SSW} \code{_W} \code{_WNW} \code{_WSW} \code{--} \code{NE} \code{NW} \code{SE} \code{SW}}
    \item{\code{Y.Koordinate}}{a numeric vector}
    \item{\code{X.Koordinate}}{a numeric vector}
    \item{\code{EK.Gesellschaftsnr}}{a numeric vector}
  }
}
\details{
Classification used in 1972 is in variable "EK.Gesellschaftsnr"
}
\source{
Ellenberg, H. & Kloetzli, F. 1972.
Waldgesellschaften und Waldstandorte der Schweiz. Mitt. Eidgenoess.
Forsch. anst. Wald Schnee Landsch. 48(4): 587--930.

Keller, W., Wohlgemuth, T., Kuhn, N., Schuetz, M. &
Wildi, O. 1998. Waldgesellschaften der Schweiz auf floristischer
Grundlage. Mitteilungen der Eidgenoessischen Forschungsanstalt fuer
Wald, Schnee und Landschaft (WSL) 73, Vol. 2.
}
\references{
Wildi, O. 2013. Data Analysis in Vegetation Ecology. 2nd ed. Wiley-Blackwell,
Chichester.
}
\examples{
summary(EKs)
}
\keyword{datasets}
