\name{EKv}
\alias{EKv}
\docType{data}
\title{
Swiss forest vegetation data base 1972, vegetation information
}
\description{
Swiss forest vegetation data base 1972, vegetation  information. Site data is in data frame \code{\link{EKs}}.
}
\usage{data(EKv)}
\format{
A data frame with 2533 observations on 1259 species, the variables on a numerical scale from 0 to 7.
}
\details{
Old taxonomy, not updated.
}
\source{
Ellenberg, H. & Kloetzli, F. 1972.
Waldgesellschaften und Waldstandorte der Schweiz. Mitt. Eidgenoess.
Forsch. anst. Wald Schnee Landsch. 48(4): 587--930.

Keller, W., Wohlgemuth, T., Kuhn, N., Schuetz, M. &
Wildi, O. 1998. Waldgesellschaften der Schweiz auf floristischer
Grundlage. Mitteilungen der Eidgenoessischen Forschungsanstalt fuer
Wald, Schnee und Landschaft (WSL) 73, Vol. 2.
}
\references{
Wildi, O. 2013. Data Analysis in Vegetation Ecology. 2nd ed. Wiley-Blackwell,
Chichester.
}
\examples{
summary(EKv)
}
\keyword{datasets}
