% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factset.R
\name{export_factsets}
\alias{export_factsets}
\title{Export factsets exports the dataframe into a jsonlines file format.}
\usage{
export_factsets(factset, path = NULL)
}
\arguments{
\item{factset}{is the data frame to be exported}

\item{path}{path is the optional path which the file is to be saved}
}
\description{
Export factsets exports the dataframe into a jsonlines file format.
}

