% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factset.R
\name{flatten_df}
\alias{flatten_df}
\title{Flattens a dataframe given the entity and date fields}
\usage{
flatten_df(factsets, schema = NULL)
}
\arguments{
\item{factsets}{is the dataframe to be flattened}

\item{schema}{is the provided schema of the dataframe}
}
\description{
Flattens a dataframe given the entity and date fields
}

