% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureset.R
\name{get_filtered.character}
\alias{get_filtered.character}
\title{Gets the gradient related information for a period of time}
\usage{
\method{get_filtered}{character}(df, schema = NULL, date = NULL,
  method = "snapshot")
}
\arguments{
\item{df}{is the path of dataframe object of interest}

\item{schema}{is the schema of the object}

\item{date}{is the date which we want to filter on.
This can be either a vector, empty, or single date. If it is a vector
it will filter based on this vector, if it is a single date it will take latest entry based
on the date. If it is NULL it will consider the whole time period}

\item{method}{is the approach on which the data will be subsetted. Can be either snapshot  or left out (will be assumed to be non-snapshot)}
}
\description{
Gets the gradient related information for a period of time
}
\seealso{
\code{\link{get_filtered}}
}

