% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureset.R
\name{get_snapshot}
\alias{get_snapshot}
\title{Convenience function for get filtered}
\usage{
get_snapshot(df, schema = NULL, date = NULL)
}
\arguments{
\item{df}{is the path of dataframe object of interest}

\item{schema}{is the schema of the object}

\item{date}{is the date which we want to filter on.
This can be either a vector, empty, or single date. If it is a vector
it will filter based on this vector, if it is a single date it will take latest entry based
on the date. If it is NULL it will consider the whole time period}
}
\description{
Convenience function for get filtered
}
\seealso{
\code{\link{get_filtered}}
}

