% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factset.R
\name{load_factsets}
\alias{load_factsets}
\title{Load factsets is the key function which takes in a jsonline files and flattens, applying
all relevant transformations based on the schema.}
\usage{
load_factsets(path, schema = NULL)
}
\arguments{
\item{path}{is the path of the jsonlines file}

\item{schema}{is the schema information}
}
\description{
Load factsets is the key function which takes in a jsonline files and flattens, applying
all relevant transformations based on the schema.
}

