% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factset.R
\name{load_jsonlines}
\alias{load_jsonlines}
\title{Load jsonlines file into a data frame object.}
\usage{
load_jsonlines(path)
}
\arguments{
\item{path}{is the location of the jsonlines file}
}
\description{
Load jsonlines file into a data frame object.
}

