\name{ltim}
\alias{ltim}
\docType{data}
\title{
Lippe et al. 1985 data set, yr of observation
}
\description{
Lippe et al. 1985 data set, yr of observation. Vegetation data is in data frame \code{\link{lveg}}.
}
\usage{data(ltim)}
\format{
A data frame with 19 observations on the following variable.
  \describe{
    \item{\code{Year}}{a numeric vector containing the year of sampling}
  }
}
\source{
Lippe, E., De Smitt, J.T. & Glenn-Lewin, D.C.
1985. Markov models and succession: a test from a heathland in the
Netherlands. Journal of Ecology 73: 775--791.

Orloci, L., Anand, M. & He, X. 1993. Markov chain: a
realistic model for temporal coenosere? Biom. Praxim 33: 7--26.
}
\references{
Wildi, O. 2013. Data Analysis in Vegetation Ecology. 2nd ed. Wiley-Blackwell,
Chichester.
}
\examples{
summary(ltim)
}
\keyword{datasets}
