\name{lveg}
\alias{lveg}
\docType{data}
\title{
Lippe et al. 1985 data set, vegetation data 
}
\description{
Lippe et al. 1985 data set, vegetation data. Year of observation is in data frame ltim.
}
\usage{data(lveg)}
\format{
  A data frame with 19 observations on the following 9 variables.
  \describe{
    \item{\code{open.soil}}{a numeric vector}
    \item{\code{Empetrum.nigrum}}{a numeric vector}
    \item{\code{Calluna.vulgaris}}{a numeric vector}
    \item{\code{Erica.tetralix}}{a numeric vector}
    \item{\code{Molinia.coerulea}}{a numeric vector}
    \item{\code{Carex.pilulifera}}{a numeric vector}
    \item{\code{Juncus.squarrosus}}{a numeric vector}
    \item{\code{Rumex.acetosella}}{a numeric vector}
    \item{\code{other.species}}{a numeric vector}
  }
}
\source{
Lippe, E., De Smitt, J.T. & Glenn-Lewin, D.C.
1985. Markov models and succession: a test from a heathland in the
Netherlands. Journal of Ecology 73: 775--791.

Orloci, L., Anand, M. & He, X. 1993. Markov chain: a
realistic model for temporal coenosere? Biom. Praxim 33: 7--26.
}
\references{
Wildi, O. 2013. Data Analysis in Vegetation Ecology. 2nd ed. Wiley-Blackwell,
Chichester.
}
\examples{
summary(lveg)
}
\keyword{datasets}
