\name{mveg}
\alias{mveg}
\docType{data}
\title{
Ellenberg's 1956 meadow data
}
\description{
Ellenberg's 1956 meadow data. No site factors available.
}
\usage{data(mveg)}
\format{
  A data frame with 25 observations on the following 94 variables.
  \describe{
    \item{\code{ACHI.MILL.}}{a numeric vector}
    \item{\code{AJUG.REPT.}}{a numeric vector}
    \item{\code{ALCH.VULG.}}{a numeric vector}
    \item{\code{ALOP.PRAT.}}{a numeric vector}
    \item{\code{ANGE.SILV.}}{a numeric vector}
    \item{\code{ANTH.SILV.}}{a numeric vector}
    \item{\code{ARRH.ELAT.}}{a numeric vector}
    \item{\code{BELL.PERE.}}{a numeric vector}
    \item{\code{BRAC.PINN.}}{a numeric vector}
    \item{\code{BRIZ.MEDI.}}{a numeric vector}
    \item{\code{BROM.EREC.}}{a numeric vector}
    \item{\code{CAMP.GLOM.}}{a numeric vector}
    \item{\code{CAMP.ROTU.}}{a numeric vector}
    \item{\code{CARD.PRAT.}}{a numeric vector}
    \item{\code{CARE.ACUT.}}{a numeric vector}
    \item{\code{CARE.FLAC.}}{a numeric vector}
    \item{\code{CARE.GRAC.}}{a numeric vector}
    \item{\code{CAR.EHIRT.}}{a numeric vector}
    \item{\code{CARE.PANI.}}{a numeric vector}
    \item{\code{CENT.JACE.}}{a numeric vector}
    \item{\code{CERA.CAES.}}{a numeric vector}
    \item{\code{CHEN.ALBU.}}{a numeric vector}
    \item{\code{CHRY.LEUC.}}{a numeric vector}
    \item{\code{CIRS.OLER.}}{a numeric vector}
    \item{\code{CREP.BIEN.}}{a numeric vector}
    \item{\code{DACT.GLOM.}}{a numeric vector}
    \item{\code{DAUC.CARO.}}{a numeric vector}
    \item{\code{DESC.CAES.}}{a numeric vector}
    \item{\code{DIAN.SUPE.}}{a numeric vector}
    \item{\code{EUPH.ODON.}}{a numeric vector}
    \item{\code{FEST.ARUN.}}{a numeric vector}
    \item{\code{FEST.OVIN.}}{a numeric vector}
    \item{\code{FEST.PRAT.}}{a numeric vector}
    \item{\code{FEST.RUBR.}}{a numeric vector}
    \item{\code{FILI.ULMA.}}{a numeric vector}
    \item{\code{GALI.BORE.}}{a numeric vector}
    \item{\code{GALI.MOLL.}}{a numeric vector}
    \item{\code{GALI.ULIG.}}{a numeric vector}
    \item{\code{GALI.VERU.}}{a numeric vector}
    \item{\code{GERA.PRAT.}}{a numeric vector}
    \item{\code{GEUM.RIVA.}}{a numeric vector}
    \item{\code{GLEC.HEDE.}}{a numeric vector}
    \item{\code{GLYC.FLUI.}}{a numeric vector}
    \item{\code{HELI.PUBE.}}{a numeric vector}
    \item{\code{HERA.SPHO.}}{a numeric vector}
    \item{\code{HOLC.LANA.}}{a numeric vector}
    \item{\code{KNAU.ARVE.}}{a numeric vector}
    \item{\code{KOEL.PYRA.}}{a numeric vector}
    \item{\code{LAMI.ALBU.}}{a numeric vector}
    \item{\code{LATH.PRAT.}}{a numeric vector}
    \item{\code{LEON.HISP.}}{a numeric vector}
    \item{\code{LINU.CATH.}}{a numeric vector}
    \item{\code{LOLI.PERE.}}{a numeric vector}
    \item{\code{LOTU.CORN.}}{a numeric vector}
    \item{\code{LYCH.FLOS.}}{a numeric vector}
    \item{\code{LYSI.NUMM.}}{a numeric vector}
    \item{\code{MEDI.LUPU.}}{a numeric vector}
    \item{\code{MELA.DIUR.}}{a numeric vector}
    \item{\code{MYOS.ARVE.}}{a numeric vector}
    \item{\code{MYOS.PALU.}}{a numeric vector}
    \item{\code{PAST.SATI.}}{a numeric vector}
    \item{\code{PHAL.ARUN.}}{a numeric vector}
    \item{\code{PHLE.PRAT.}}{a numeric vector}
    \item{\code{PHRA.COMM.}}{a numeric vector}
    \item{\code{PIMP.MAGN.}}{a numeric vector}
    \item{\code{PIMP.SAXI.}}{a numeric vector}
    \item{\code{PLAN.LANC.}}{a numeric vector}
    \item{\code{PLAN.MEDI.}}{a numeric vector}
    \item{\code{POAPRAT.}}{a numeric vector}
    \item{\code{POATRIV.}}{a numeric vector}
    \item{\code{POLY.BIST.}}{a numeric vector}
    \item{\code{POLY.CONV.}}{a numeric vector}
    \item{\code{POTE.REPT.}}{a numeric vector}
    \item{\code{PRUN.VULG.}}{a numeric vector}
    \item{\code{RANU.ACER.}}{a numeric vector}
    \item{\code{RANU.REPE.}}{a numeric vector}
    \item{\code{RUME.ACET.}}{a numeric vector}
    \item{\code{RUME.CRIS.}}{a numeric vector}
    \item{\code{SALV.PRAT.}}{a numeric vector}
    \item{\code{SANG.OFFI.}}{a numeric vector}
    \item{\code{SCAB.COLU.}}{a numeric vector}
    \item{\code{SENE.JACO.}}{a numeric vector}
    \item{\code{SILA.PRAT.}}{a numeric vector}
    \item{\code{SILE.INFL.}}{a numeric vector}
    \item{\code{TARA.OFFI.}}{a numeric vector}
    \item{\code{THYM.SERP.}}{a numeric vector}
    \item{\code{TRAG.PRAT.}}{a numeric vector}
    \item{\code{TRIF.PRAT.}}{a numeric vector}
    \item{\code{TRIF.REPE.}}{a numeric vector}
    \item{\code{TRIS.FLAV.}}{a numeric vector}
    \item{\code{VERO.CHAM.}}{a numeric vector}
    \item{\code{VICI.CRAC.}}{a numeric vector}
    \item{\code{VICI.SEPI.}}{a numeric vector}
    \item{\code{VIOL.HIRT.}}{a numeric vector}
  }
}
\details{
No site factors available for this data frame.
}
\source{
Mueller-Dombois, D. & Ellenberg, H. 1974. Aims and
Methods of Vegetation Ecology. John Wiley & Sons, New York, Chichester,
Brisbane, Toronto.
}
\references{
Wildi, O. 2013. Data Analysis in Vegetation Ecology. 2nd ed. Wiley-Blackwell,
Chichester.
}
\examples{
summary(mveg)
}
\keyword{datasets}
