\name{nveg}
\alias{nveg}
\docType{data}
\title{
European beach forest data, vegetation
}
\description{
European beach forest data, vegetation. Site factors are in data frame sveg.
}
\usage{data(nveg)}
\format{
  A data frame with 11 observations on the following 21 variables.
  \describe{
    \item{\code{Fagus.silvatica}}{a numeric vector}
    \item{\code{Quercus.petraea}}{a numeric vector}
    \item{\code{Acer.pseudoplatanus}}{a numeric vector}
    \item{\code{Fraxinus.excelsior}}{a numeric vector}
    \item{\code{Lonicera.xylosteum}}{a numeric vector}
    \item{\code{Sambucus.racemosa}}{a numeric vector}
    \item{\code{Sambucus.nigra}}{a numeric vector}
    \item{\code{Vaccinium.myrtillus}}{a numeric vector}
    \item{\code{Carex.silvatica}}{a numeric vector}
    \item{\code{Oxalis.acetosella}}{a numeric vector}
    \item{\code{Viola.silvestris}}{a numeric vector}
    \item{\code{Luzula.nemorosa}}{a numeric vector}
    \item{\code{Veronica.officinalis}}{a numeric vector}
    \item{\code{Galium.odoratum}}{a numeric vector}
    \item{\code{Lamium.galeobdolon}}{a numeric vector}
    \item{\code{Primula.elatior}}{a numeric vector}
    \item{\code{Allium.ursinum}}{a numeric vector}
    \item{\code{Arum.maculatum}}{a numeric vector}
    \item{\code{Ranunculus.ficaria}}{a numeric vector}
    \item{\code{Eurhynchium.striatum}}{a numeric vector}
    \item{\code{Polytrichum.formosum}}{a numeric vector}
  }
}
\details{
Artificial data
}
\source{
Wildi, O. & Orloci, L. 1996. Numerical Exploration
of Community Patterns. 2nd ed. SPB Academic Publishing, The Hague.
}
\references{
Wildi, O. 2013. Data Analysis in Vegetation Ecology. 2nd ed. Wiley-Blackwell,
Chichester.
}
\examples{
summary(nveg)
}
\keyword{datasets}
