\name{psit}
\alias{psit}
\docType{data}
\title{
Time scale (yr) for Soppensee pollen data
}
\description{
Time scale (yr) for Soppensee pollen data. See details below. Vegetation is in data frame \code{\link{pveg}}.
}
\usage{data(psit)}
\format{
  A data frame with 145 observations on the following variable.
  \describe{
    \item{\code{Years.B.P}}{a numeric vector}
  }
}
\details{
Time scale (no corrections applied for revised 14C calibration).
}
\source{
Lotter, A.F. 1999. Late-glacial and Holocene
vegetation history and dynamics as shown by pollen macrofossil
analyses in annually laminated sediments from Soppensee, central
Switzerland. Vegetation History and Archaeobotany 8: 165-184.
}
\references{
Wildi, O. 2013. Data Analysis in Vegetation Ecology. 2nd Edition, Wiley-Blackwell,
Chichester.
}
\examples{
summary(psit)
}
\keyword{datasets}
