\name{pveg}
\alias{pveg}
\docType{data}
\title{
Soppensee pollen data
}
\description{
Soppensee pollen data. Vegetation only (tree species). The age (year B.P) is in \code{\link{psit}}.
}
\usage{data(pveg)}
\format{
  A data frame with 145 observations on the following 14 variables.
  \describe{
    \item{\code{X1.Abies}}{a numeric vector}
    \item{\code{X2.Pinus}}{a numeric vector}
    \item{\code{X3.Fagus}}{a numeric vector}
    \item{\code{X4.Quercus}}{a numeric vector}
    \item{\code{X5.Acer}}{a numeric vector}
    \item{\code{X6.Fraxinus}}{a numeric vector}
    \item{\code{X7.Ulmus}}{a numeric vector}
    \item{\code{X8.Tilia}}{a numeric vector}
    \item{\code{X9.Betula}}{a numeric vector}
    \item{\code{X10.Alnus}}{a numeric vector}
    \item{\code{X11.Populus}}{a numeric vector}
    \item{\code{X12.Salix}}{a numeric vector}
    \item{\code{X13.Sorbus}}{a numeric vector}
    \item{\code{X14.Picea}}{a numeric vector}
  }
}
\source{
Lotter, A.F. 1999. Late-glacial and Holocene
vegetation history and dynamics as shown by pollen macrofossil
analyses in annually laminated sediments from Soppensee, central
Switzerland. Vegetation History and Archaeobotany 8: 165-184.
}
\references{
Wildi, O. 2013. Data Analysis in Vegetation Ecology. 2nd ed. Wiley-Blackwell,
Chichester.
}
\examples{
summary(pveg)
}
\keyword{datasets}
