\name{ssit}
\alias{ssit}
\docType{data}
\title{
Schlaenggli site data
}
\description{
Schlaenggli site data. 63 plots arranged in a square grid. Vegetation in data frame sveg.
}
\usage{data(ssit)}
\format{
  A data frame with 63 observations on the following 20 variables.
  \describe{
    \item{\code{pH.peat}}{a numeric vector}
    \item{\code{log.ash.perc}}{a numeric vector}
    \item{\code{Ca_peat}}{a numeric vector}
    \item{\code{Mg_peat}}{a numeric vector}
    \item{\code{Na_peat}}{a numeric vector}
    \item{\code{K_peat}}{a numeric vector}
    \item{\code{Acidity.peat}}{a numeric vector}
    \item{\code{CEC.peat}}{a numeric vector}
    \item{\code{Base.sat.perc}}{a numeric vector}
    \item{\code{P.peat}}{a numeric vector}
    \item{\code{Waterlev.max}}{a numeric vector}
    \item{\code{Waterlev.av}}{a numeric vector}
    \item{\code{Waterlev.min}}{a numeric vector}
    \item{\code{log.peat.lev}}{a numeric vector}
    \item{\code{log.slope.deg}}{a numeric vector}
    \item{\code{pH.water}}{a numeric vector}
    \item{\code{log.cond.water}}{a numeric vector}
    \item{\code{log.Ca.water}}{a numeric vector}
    \item{\code{x.axis}}{a numeric vector}
    \item{\code{y.axis}}{a numeric vector}
  }
}
\source{
Wildi, O. 1977. Beschreibung exzentrischer Hochmoore
mit Hilfe quantitativer Methoden. Veroeff. Geobot. Inst. ETH, Stiftung Ruebel 60: 128S.
}
\references{
Wildi, O. 2013. Data Analysis in Vegetation Ecology. 2nd ed. Wiley-Blackwell,
Chichester.
}
\examples{
summary(ssit)
}
\keyword{datasets}
