\name{sveg}
\alias{sveg}
\docType{data}
\title{
Schlaenggli vegetation data 
}
\description{
Schlaenggli vegetation data. 63 plots arranged in a square grid. Site factors in data frame ssit.
}
\usage{data(sveg)}
\format{
  A data frame with 63 observations on the following 119 variables.
  \describe{
    \item{\code{Vaccinium.myrtillus}}{a numeric vector}
    \item{\code{Vaccinium.uliginosum}}{a numeric vector}
    \item{\code{Vaccinium.oxycoccos}}{a numeric vector}
    \item{\code{Calluna.vulgaris}}{a numeric vector}
    \item{\code{Carex.rostrata}}{a numeric vector}
    \item{\code{Carex.lasiocarpa}}{a numeric vector}
    \item{\code{Carex.nigra}}{a numeric vector}
    \item{\code{Carex.davalliana}}{a numeric vector}
    \item{\code{Carex.hostiana}}{a numeric vector}
    \item{\code{Carex.pauciflora}}{a numeric vector}
    \item{\code{Carex.echinata}}{a numeric vector}
    \item{\code{Carex.panicea}}{a numeric vector}
    \item{\code{Trichophorum.caespitosum}}{a numeric vector}
    \item{\code{Trichophorum.alpinum}}{a numeric vector}
    \item{\code{Eriophorum.vaginatum}}{a numeric vector}
    \item{\code{Eriophorum.latifolium}}{a numeric vector}
    \item{\code{Eriophorum.angustifolium}}{a numeric vector}
    \item{\code{Tofieldia.calyculata}}{a numeric vector}
    \item{\code{Pinguicula.vulgaris}}{a numeric vector}
    \item{\code{Parnassia.palustris}}{a numeric vector}
    \item{\code{Rhynchospora.alba}}{a numeric vector}
    \item{\code{Drosera.rotundifolia}}{a numeric vector}
    \item{\code{Arnica.montana}}{a numeric vector}
    \item{\code{Dactylorhiza.maculata}}{a numeric vector}
    \item{\code{Polytrichum.strictum}}{a numeric vector}
    \item{\code{Succisa.pratensis}}{a numeric vector}
    \item{\code{Anthoxanthum.odoratum}}{a numeric vector}
    \item{\code{Molinia.caerulea}}{a numeric vector}
    \item{\code{Gentiana.asclepiadea}}{a numeric vector}
    \item{\code{Luzula.multiflora}}{a numeric vector}
    \item{\code{Festuca.rubra}}{a numeric vector}
    \item{\code{Potentilla.erecta}}{a numeric vector}
    \item{\code{Sphagnum.magellanicum}}{a numeric vector}
    \item{\code{Sphagnum.papillosum}}{a numeric vector}
    \item{\code{Equisetum.palustre}}{a numeric vector}
    \item{\code{Aulacomnium.palustre}}{a numeric vector}
    \item{\code{Euphrasia.rostkoviana}}{a numeric vector}
    \item{\code{Nardus.stricta}}{a numeric vector}
    \item{\code{Carex.pulicaris}}{a numeric vector}
    \item{\code{Climacium.dendroides}}{a numeric vector}
    \item{\code{Carex.pallescens}}{a numeric vector}
    \item{\code{Carex.lepidocarpa}}{a numeric vector}
    \item{\code{Briza.media}}{a numeric vector}
    \item{\code{Sphagnum.platyphyllum}}{a numeric vector}
    \item{\code{Sphagnum.tenellum}}{a numeric vector}
    \item{\code{Sphagnum.recurvum}}{a numeric vector}
    \item{\code{Sphagnum.rubellum}}{a numeric vector}
    \item{\code{Sphagnum.centrale}}{a numeric vector}
    \item{\code{Sphagnum.subsecundum}}{a numeric vector}
    \item{\code{Juncus.alpinoarticulatus}}{a numeric vector}
    \item{\code{Solidago.virgaurea}}{a numeric vector}
    \item{\code{Agrostis.canina}}{a numeric vector}
    \item{\code{Juncus.effusus}}{a numeric vector}
    \item{\code{Frangula.alnus}}{a numeric vector}
    \item{\code{Anemone.nemorosa}}{a numeric vector}
    \item{\code{Holcus.lanatus}}{a numeric vector}
    \item{\code{Drepanocladus.exannulatus}}{a numeric vector}
    \item{\code{Homogyne.alpina}}{a numeric vector}
    \item{\code{Polygala.vulgaris}}{a numeric vector}
    \item{\code{Trollius.europaeus}}{a numeric vector}
    \item{\code{Polygonum.bistorta}}{a numeric vector}
    \item{\code{Lotus.pedunculatus}}{a numeric vector}
    \item{\code{Trifolium.pratense}}{a numeric vector}
    \item{\code{Taraxacum.palustre}}{a numeric vector}
    \item{\code{Campylium.stellatum}}{a numeric vector}
    \item{\code{Cirsium.palustre}}{a numeric vector}
    \item{\code{Lysimachia.vulgaris}}{a numeric vector}
    \item{\code{Maianthemum.bifolium}}{a numeric vector}
    \item{\code{Equisetum.sylvaticum}}{a numeric vector}
    \item{\code{Centaurea.jacea}}{a numeric vector}
    \item{\code{Prunella.grandiflora}}{a numeric vector}
    \item{\code{Viola.palustris}}{a numeric vector}
    \item{\code{Angelica.sylvestris}}{a numeric vector}
    \item{\code{Filipendula.ulmaria}}{a numeric vector}
    \item{\code{Thuidium.delicatulum}}{a numeric vector}
    \item{\code{Sanguisorba.officinalis}}{a numeric vector}
    \item{\code{Galium.uliginosum}}{a numeric vector}
    \item{\code{Epipactis.palustris}}{a numeric vector}
    \item{\code{Pleurozium.schreberi}}{a numeric vector}
    \item{\code{Calliergon.stramineum}}{a numeric vector}
    \item{\code{Aster.bellidiastrum}}{a numeric vector}
    \item{\code{Hypnum.bambergeri}}{a numeric vector}
    \item{\code{Drepanocladus.revolvens}}{a numeric vector}
    \item{\code{Dicranum.bonjeanii}}{a numeric vector}
    \item{\code{Dicranum.scoparium}}{a numeric vector}
    \item{\code{Hylocomium.splendens}}{a numeric vector}
    \item{\code{Rhytidiadelphus.squarrosus}}{a numeric vector}
    \item{\code{Calliergonella.cuspidata}}{a numeric vector}
    \item{\code{Fissidens.adianthoides}}{a numeric vector}
    \item{\code{Primula.veris}}{a numeric vector}
    \item{\code{Primula.farinosa}}{a numeric vector}
    \item{\code{Hieracium.lachenalii}}{a numeric vector}
    \item{\code{Crepis.paludosa}}{a numeric vector}
    \item{\code{Leontodon.hispidus}}{a numeric vector}
    \item{\code{Lotus.corniculatus}}{a numeric vector}
    \item{\code{Stachys.officinalis}}{a numeric vector}
    \item{\code{Carex.flacca}}{a numeric vector}
    \item{\code{Gentiana.verna}}{a numeric vector}
    \item{\code{Geranium.sylvaticum}}{a numeric vector}
    \item{\code{Ranunculus.nemorosus}}{a numeric vector}
    \item{\code{Linum.catharticum}}{a numeric vector}
    \item{\code{Cirsium.oleraceum}}{a numeric vector}
    \item{\code{Colchicum.autumnale}}{a numeric vector}
    \item{\code{Polygala.amarella}}{a numeric vector}
    \item{\code{Polygala.serpyllifolia}}{a numeric vector}
    \item{\code{Equisetum.arvense}}{a numeric vector}
    \item{\code{Deschampsia.cespitosa}}{a numeric vector}
    \item{\code{Rhinanthus.minor}}{a numeric vector}
    \item{\code{Gymnadenia.conopsea}}{a numeric vector}
    \item{\code{Melampyrum.pratense}}{a numeric vector}
    \item{\code{Salix.spec.}}{a numeric vector}
    \item{\code{Caltha.palustris}}{a numeric vector}
    \item{\code{Dactylorhiza.fistulosa}}{a numeric vector}
    \item{\code{Platanthera.bifolia}}{a numeric vector}
    \item{\code{Acer.pseudoplatanus}}{a numeric vector}
    \item{\code{Danthonia.decumbens}}{a numeric vector}
    \item{\code{Drepanocladus.spec.}}{a numeric vector}
    \item{\code{Tomentypnum.nitens}}{a numeric vector}
    \item{\code{Ctenidium.molluscum}}{a numeric vector}
  }
}
\source{
Wildi, O. 1977. Beschreibung exzentrischer Hochmoore
mit Hilfe quantitativer Methoden. Veroeff. Geobot. Inst. ETH, Stiftung Ruebel 60: 128S.
}
\references{
Wildi, O. 2013. Data Analysis in Vegetation Ecology. 2nd ed. Wiley-Blackwell,
Chichester.
}
\examples{
summary(sveg)
}
\keyword{datasets}
