% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_datetime}
\alias{to_datetime}
\title{converts date string vector from iso8601 format}
\usage{
to_datetime(dates)
}
\arguments{
\item{dates}{is a vector of date strings}
}
\description{
converts date string vector from iso8601 format
}

