\name{wetveg}
\alias{wetveg}
\docType{data}
\title{
Random sample of Swiss wetland vegetation, vegetation information.
}
\description{
Random sample of Swiss wetland vegetation, vegetation information. Site data is in \code{\link{wetsit}}.
Note: this differs from the same used in the 2nd edition in that 4 observations and 1 variable are erased.
}
\usage{data(wetveg)}
\format{
A data frame with 1496 observations on 1163 variables, the species. A 0 to 4 step scales is used.
}
\source{
Graf, U., Wildi, O., Feldmeyer-Christe, E. &
Kuechler, M. 2010. A phytosociological classification of Swiss mire
vegetation. Botanica Helvetica 120: 1-13.
}
\references{
Wildi, O. 2017. Data Analysis in Vegetation Ecology. 3rd ed. CABI, Oxfordshire, Boston.
}
\examples{
summary(wetveg)
}
\keyword{datasets}
