\name{ws30}
\alias{ws30}
\docType{data}
\title{
Four kilometre grid forest data of Switzerland, 30m2 plots
}
\description{
Four kilometre grid forest data of Switzerland, 30m2 plots. Also see ws200, ws500. Site information is in \code{\link{wssit}}.
}
\usage{data(ws30)}
\format{
  A data frame with 726 observations on 1262 variables, the species.
 }
\details{
See object wssit for corresponding site information
}
\source{
Wohlgemuth, T., Moser B., Braendli, U.-B., Kull, P. & Schuetz, M. 2008. 
Diversity of forest plant species at the community and landscape scales in Switzerland. 
Plant Biosystems 142: 604-613.
}
\references{
Wildi, O. 2017. Data Analysis in Vegetation Ecology. 3rd ed. CABI, Oxfordshire, Boston.
}
\examples{
summary(ws30)
}
\keyword{datasets}
