\name{ws500}
\alias{ws500}
\docType{data}
\title{
Four kilometre grid forest data of Switzerland, 500m2 plots
}
\description{
Four kilometre grid forest data of Switzerland, 500m2 plots. Also see ws30, ws200.
}
\usage{data(ws500)}
\format{
  A data frame with 726 observations on the following 1262 variables.
  \describe{
    \item{\code{K.Huperzia.selago.}}{a numeric vector}
    \item{\code{K.Equisetum.palustre}}{a numeric vector}
    \item{\code{K.Equisetum.sylvaticum}}{a numeric vector}
    \item{\code{K.Equisetum.arvense.}}{a numeric vector}
    \item{\code{K.Equisetum.telmateia.}}{a numeric vector}
    \item{\code{K.Pteridium.aquilinum.}}{a numeric vector}
    \item{\code{K.Oreopteris.limbosperma}}{a numeric vector}
    \item{\code{K.Phegopteris.connectilis.}}{a numeric vector}
    \item{\code{K.Asplenium.trichomanes.}}{a numeric vector}
    \item{\code{K.Asplenium.viride}}{a numeric vector}
    \item{\code{K.Asplenium.adiantum.nigrum.}}{a numeric vector}
    \item{\code{K.Phyllitis.scolopendrium.}}{a numeric vector}
    \item{\code{K.Athyrium.filix.femina.}}{a numeric vector}
    \item{\code{K.Athyrium.distentifolium.}}{a numeric vector}
    \item{\code{K.Cystopteris.fragilis}}{a numeric vector}
    \item{\code{K.Polystichum.lonchitis.}}{a numeric vector}
    \item{\code{K.Polystichum.aculeatum.}}{a numeric vector}
    \item{\code{K.Dryopteris.filix.mas}}{a numeric vector}
    \item{\code{K.Dryopteris.carthusiana}}{a numeric vector}
    \item{\code{K.Dryopteris.dilatata.}}{a numeric vector}
    \item{\code{K.Gymnocarpium.dryopteris.}}{a numeric vector}
    \item{\code{K.Gymnocarpium.robertianum}}{a numeric vector}
    \item{\code{K.Blechnum.spicant}}{a numeric vector}
    \item{\code{K.Polypodium.vulgare}}{a numeric vector}
    \item{\code{B.Abies.alba}}{a numeric vector}
    \item{\code{S.Abies.alba}}{a numeric vector}
    \item{\code{K.Abies.alba}}{a numeric vector}
    \item{\code{B.Picea.abies.}}{a numeric vector}
    \item{\code{S.Picea.abies.}}{a numeric vector}
    \item{\code{K.Picea.abies.}}{a numeric vector}
    \item{\code{B.Larix.decidua.}}{a numeric vector}
    \item{\code{S.Larix.decidua.}}{a numeric vector}
    \item{\code{K.Larix.decidua.}}{a numeric vector}
    \item{\code{B.Pinus.sylvestris}}{a numeric vector}
    \item{\code{S.Pinus.sylvestris}}{a numeric vector}
    \item{\code{K.Pinus.sylvestris}}{a numeric vector}
    \item{\code{S.Juniperus.communis}}{a numeric vector}
    \item{\code{K.Juniperus.communis}}{a numeric vector}
    \item{\code{B.Taxus.baccata.}}{a numeric vector}
    \item{\code{S.Taxus.baccata.}}{a numeric vector}
    \item{\code{K.Taxus.baccata.}}{a numeric vector}
    \item{\code{K.Ephedra.helvetica.}}{a numeric vector}
    \item{\code{S.Salix.alba}}{a numeric vector}
    \item{\code{K.Salix.alba}}{a numeric vector}
    \item{\code{S.Salix.appendiculata.}}{a numeric vector}
    \item{\code{B.Salix.caprea}}{a numeric vector}
    \item{\code{S.Salix.caprea}}{a numeric vector}
    \item{\code{K.Salix.caprea}}{a numeric vector}
    \item{\code{S.Salix.waldsteiniana.}}{a numeric vector}
    \item{\code{K.Salix.waldsteiniana.}}{a numeric vector}
    \item{\code{S.Salix.purpurea}}{a numeric vector}
    \item{\code{K.Salix.purpurea}}{a numeric vector}
    \item{\code{B.Populus.tremula.}}{a numeric vector}
    \item{\code{S.Populus.tremula.}}{a numeric vector}
    \item{\code{K.Populus.tremula.}}{a numeric vector}
    \item{\code{B.Populus.nigra.}}{a numeric vector}
    \item{\code{S.Populus.nigra.}}{a numeric vector}
    \item{\code{K.Populus.nigra.}}{a numeric vector}
    \item{\code{B.Juglans.regia.}}{a numeric vector}
    \item{\code{S.Juglans.regia.}}{a numeric vector}
    \item{\code{K.Juglans.regia.}}{a numeric vector}
    \item{\code{B.Betula.pendula}}{a numeric vector}
    \item{\code{S.Betula.pendula}}{a numeric vector}
    \item{\code{K.Betula.pendula}}{a numeric vector}
    \item{\code{S.Alnus.viridis.}}{a numeric vector}
    \item{\code{B.Alnus.glutinosa.}}{a numeric vector}
    \item{\code{S.Alnus.glutinosa.}}{a numeric vector}
    \item{\code{K.Alnus.glutinosa.}}{a numeric vector}
    \item{\code{B.Alnus.incana}}{a numeric vector}
    \item{\code{K.Alnus.incana}}{a numeric vector}
    \item{\code{B.Carpinus.betulus}}{a numeric vector}
    \item{\code{S.Carpinus.betulus}}{a numeric vector}
    \item{\code{K.Carpinus.betulus}}{a numeric vector}
    \item{\code{B.Corylus.avellana}}{a numeric vector}
    \item{\code{S.Corylus.avellana}}{a numeric vector}
    \item{\code{K.Corylus.avellana}}{a numeric vector}
    \item{\code{B.Fagus.sylvatica.}}{a numeric vector}
    \item{\code{S.Fagus.sylvatica.}}{a numeric vector}
    \item{\code{K.Fagus.sylvatica.}}{a numeric vector}
    \item{\code{B.Quercus.petraea.}}{a numeric vector}
    \item{\code{S.Quercus.petraea.}}{a numeric vector}
    \item{\code{K.Quercus.petraea.}}{a numeric vector}
    \item{\code{B.Quercus.robur.}}{a numeric vector}
    \item{\code{S.Quercus.robur.}}{a numeric vector}
    \item{\code{K.Quercus.robur.}}{a numeric vector}
    \item{\code{B.Ulmus.glabra}}{a numeric vector}
    \item{\code{S.Ulmus.glabra}}{a numeric vector}
    \item{\code{K.Ulmus.glabra}}{a numeric vector}
    \item{\code{B.Ulmus.minor.}}{a numeric vector}
    \item{\code{S.Ulmus.minor.}}{a numeric vector}
    \item{\code{K.Ulmus.minor.}}{a numeric vector}
    \item{\code{K.Urtica.dioeca.}}{a numeric vector}
    \item{\code{K.Thesium.alpinum.}}{a numeric vector}
    \item{\code{K.Thesium.pyrenaicum}}{a numeric vector}
    \item{\code{K.Asarum.europaeum}}{a numeric vector}
    \item{\code{K.Polygonum.mite}}{a numeric vector}
    \item{\code{K.Polygonum.hydropiper}}{a numeric vector}
    \item{\code{K.Polygonum.bistorta}}{a numeric vector}
    \item{\code{K.Reynoutria.japonica.}}{a numeric vector}
    \item{\code{K.Rumex.acetosella.s.l}}{a numeric vector}
    \item{\code{K.Rumex.alpestris.}}{a numeric vector}
    \item{\code{K.Rumex.acetosa.}}{a numeric vector}
    \item{\code{K.Rumex.sanguineus}}{a numeric vector}
    \item{\code{K.Rumex.obtusifolius}}{a numeric vector}
    \item{\code{K.Moehringia.trinervia}}{a numeric vector}
    \item{\code{K.Moehringia.muscosa}}{a numeric vector}
    \item{\code{K.Stellaria.nemorum.}}{a numeric vector}
    \item{\code{K.Stellaria.media.}}{a numeric vector}
    \item{\code{K.Stellaria.neglecta}}{a numeric vector}
    \item{\code{K.Stellaria.uliginosa.}}{a numeric vector}
    \item{\code{K.Stellaria.graminea}}{a numeric vector}
    \item{\code{K.Cerastium.holosteoides}}{a numeric vector}
    \item{\code{K.Cerastium.fontanum}}{a numeric vector}
    \item{\code{K.Myosoton.aquaticum}}{a numeric vector}
    \item{\code{K.Lychnis.flos.cuculi.}}{a numeric vector}
    \item{\code{K.Silene.nutans.}}{a numeric vector}
    \item{\code{K.Silene.vulgaris.}}{a numeric vector}
    \item{\code{K.Silene.dioica.}}{a numeric vector}
    \item{\code{K.Helleborus.foetidus.}}{a numeric vector}
    \item{\code{K.Trollius.europaeus}}{a numeric vector}
    \item{\code{K.Actaea.spicata}}{a numeric vector}
    \item{\code{K.Caltha.palustris}}{a numeric vector}
    \item{\code{K.Aconitum.vulparia.}}{a numeric vector}
    \item{\code{K.Aconitum.napellus.s.l.}}{a numeric vector}
    \item{\code{K.Anemone.nemorosa}}{a numeric vector}
    \item{\code{K.Hepatica.nobilis}}{a numeric vector}
    \item{\code{B.Clematis.vitalba}}{a numeric vector}
    \item{\code{S.Clematis.vitalba}}{a numeric vector}
    \item{\code{K.Clematis.vitalba}}{a numeric vector}
    \item{\code{K.Ranunculus.nemorosus}}{a numeric vector}
    \item{\code{K.Ranunculus.repens.}}{a numeric vector}
    \item{\code{K.Ranunculus.lanuginosus}}{a numeric vector}
    \item{\code{K.Ranunculus.acris}}{a numeric vector}
    \item{\code{K.Ranunculus.acris.sspfriesianus}}{a numeric vector}
    \item{\code{K.Ranunculus.montanus.}}{a numeric vector}
    \item{\code{K.Ranunculus.auricomus.s.l}}{a numeric vector}
    \item{\code{K.Ranunculus.ficaria}}{a numeric vector}
    \item{\code{K.Ranunculus.aconitifolius}}{a numeric vector}
    \item{\code{K.Ranunculus.reptans}}{a numeric vector}
    \item{\code{K.Aquilegia.vulgaris}}{a numeric vector}
    \item{\code{K.Aquilegia.atrata}}{a numeric vector}
    \item{\code{K.Thalictrum.aquilegiifolium}}{a numeric vector}
    \item{\code{S.Berberis.vulgaris.}}{a numeric vector}
    \item{\code{K.Berberis.vulgaris.}}{a numeric vector}
    \item{\code{K.Corydalis.cava}}{a numeric vector}
    \item{\code{K.Alliaria.petiolata}}{a numeric vector}
    \item{\code{K.Nasturtium.officinale.}}{a numeric vector}
    \item{\code{K.Dentaria.bulbifera}}{a numeric vector}
    \item{\code{K.Dentaria.heptaphylla}}{a numeric vector}
    \item{\code{K.Dentaria.pentaphyllos.}}{a numeric vector}
    \item{\code{K.Dentaria.polyphylla.}}{a numeric vector}
    \item{\code{K.Cardamine.pratensis.}}{a numeric vector}
    \item{\code{K.Cardamine.flexuosa}}{a numeric vector}
    \item{\code{K.Cardamine.hirsuta.}}{a numeric vector}
    \item{\code{K.Cardaminopsis.arenosa.}}{a numeric vector}
    \item{\code{K.Arabis.hirsuta}}{a numeric vector}
    \item{\code{K.Arabis.turrita}}{a numeric vector}
    \item{\code{K.Arabis.alpina.}}{a numeric vector}
    \item{\code{K.Capsella.bursa.pastoris.}}{a numeric vector}
    \item{\code{K.Thlaspi.caerulescens}}{a numeric vector}
    \item{\code{K.Thlaspi.montanum}}{a numeric vector}
    \item{\code{K.Sinapis.arvensis}}{a numeric vector}
    \item{\code{K.Sedum.acre}}{a numeric vector}
    \item{\code{K.Saxifraga.rotundifolia}}{a numeric vector}
    \item{\code{K.Chrysosplenium.alternifolium}}{a numeric vector}
    \item{\code{S.Ribes.uva.crispa}}{a numeric vector}
    \item{\code{K.Ribes.uva.crispa}}{a numeric vector}
    \item{\code{S.Ribes.alpinum.}}{a numeric vector}
    \item{\code{K.Ribes.alpinum.}}{a numeric vector}
    \item{\code{K.Aruncus.dioicus.}}{a numeric vector}
    \item{\code{K.Filipendula.ulmaria.}}{a numeric vector}
    \item{\code{K.Rubus.saxatilis.}}{a numeric vector}
    \item{\code{S.Rubus.idaeus}}{a numeric vector}
    \item{\code{K.Rubus.idaeus}}{a numeric vector}
    \item{\code{S.Rubus.fruticosus.s.l}}{a numeric vector}
    \item{\code{K.Rubus.fruticosus.s.l}}{a numeric vector}
    \item{\code{K.Rubus.caesius.}}{a numeric vector}
    \item{\code{S.Rosa.arvensis.}}{a numeric vector}
    \item{\code{K.Rosa.arvensis.}}{a numeric vector}
    \item{\code{K.Rosa.pendulina}}{a numeric vector}
    \item{\code{S.Rosa.canina.s.l.}}{a numeric vector}
    \item{\code{K.Rosa.canina.s.l.}}{a numeric vector}
    \item{\code{K.Sanguisorba.minor.}}{a numeric vector}
    \item{\code{K.Geum.montanum.}}{a numeric vector}
    \item{\code{K.Geum.rivale.}}{a numeric vector}
    \item{\code{K.Geum.urbanum}}{a numeric vector}
    \item{\code{K.Potentilla.recta}}{a numeric vector}
    \item{\code{K.Potentilla.crantzii.}}{a numeric vector}
    \item{\code{K.Potentilla.aurea}}{a numeric vector}
    \item{\code{K.Potentilla.erecta.}}{a numeric vector}
    \item{\code{K.Potentilla.reptans}}{a numeric vector}
    \item{\code{K.Potentilla.sterilis.}}{a numeric vector}
    \item{\code{K.Fragaria.vesca}}{a numeric vector}
    \item{\code{K.Alchemilla.conjuncta.s.l}}{a numeric vector}
    \item{\code{K.Alchemilla.vulgaris.s.l.}}{a numeric vector}
    \item{\code{B.Malus.sylvestris}}{a numeric vector}
    \item{\code{S.Malus.sylvestris}}{a numeric vector}
    \item{\code{K.Malus.sylvestris}}{a numeric vector}
    \item{\code{B.Sorbus.aucuparia}}{a numeric vector}
    \item{\code{S.Sorbus.aucuparia}}{a numeric vector}
    \item{\code{K.Sorbus.aucuparia}}{a numeric vector}
    \item{\code{B.Sorbus.torminalis.}}{a numeric vector}
    \item{\code{S.Sorbus.torminalis.}}{a numeric vector}
    \item{\code{K.Sorbus.torminalis.}}{a numeric vector}
    \item{\code{B.Sorbus.aria.}}{a numeric vector}
    \item{\code{S.Sorbus.aria.}}{a numeric vector}
    \item{\code{K.Sorbus.aria.}}{a numeric vector}
    \item{\code{B.Sorbus.mougeotii}}{a numeric vector}
    \item{\code{S.Sorbus.mougeotii}}{a numeric vector}
    \item{\code{K.Sorbus.mougeotii}}{a numeric vector}
    \item{\code{K.Cotoneaster.integerrimus}}{a numeric vector}
    \item{\code{S.Crataegus.laevigata.}}{a numeric vector}
    \item{\code{K.Crataegus.laevigata.}}{a numeric vector}
    \item{\code{B.Crataegus.monogyna}}{a numeric vector}
    \item{\code{S.Crataegus.monogyna}}{a numeric vector}
    \item{\code{K.Crataegus.monogyna}}{a numeric vector}
    \item{\code{S.Prunus.spinosa}}{a numeric vector}
    \item{\code{K.Prunus.spinosa}}{a numeric vector}
    \item{\code{B.Prunus.avium}}{a numeric vector}
    \item{\code{S.Prunus.avium}}{a numeric vector}
    \item{\code{K.Prunus.avium}}{a numeric vector}
    \item{\code{S.Prunus.mahaleb}}{a numeric vector}
    \item{\code{K.Prunus.mahaleb}}{a numeric vector}
    \item{\code{B.Prunus.padus.s.l}}{a numeric vector}
    \item{\code{S.Prunus.padus.s.l}}{a numeric vector}
    \item{\code{K.Prunus.padus.s.l}}{a numeric vector}
    \item{\code{K.Chamaespartium.sagittale}}{a numeric vector}
    \item{\code{B.Robinia.pseudoacacia}}{a numeric vector}
    \item{\code{K.Vicia.cracca}}{a numeric vector}
    \item{\code{K.Vicia.sylvatica.}}{a numeric vector}
    \item{\code{K.Vicia.sepium}}{a numeric vector}
    \item{\code{K.Lathyrus.vernus.}}{a numeric vector}
    \item{\code{K.Lathyrus.pratensis}}{a numeric vector}
    \item{\code{K.Ononis.repens.sspprocurrens.}}{a numeric vector}
    \item{\code{K.Medicago.lupulina.}}{a numeric vector}
    \item{\code{K.Medicago.sativa.}}{a numeric vector}
    \item{\code{K.Trifolium.repens}}{a numeric vector}
    \item{\code{K.Trifolium.pratense}}{a numeric vector}
    \item{\code{K.Trifolium.medium}}{a numeric vector}
    \item{\code{K.Trifolium.rubens}}{a numeric vector}
    \item{\code{K.Lotus.corniculatus}}{a numeric vector}
    \item{\code{K.Lotus.alpinus.}}{a numeric vector}
    \item{\code{K.Anthyllis.vulneraria}}{a numeric vector}
    \item{\code{K.Anthyllis.alpestris.}}{a numeric vector}
    \item{\code{S.Coronilla.emerus}}{a numeric vector}
    \item{\code{K.Coronilla.emerus}}{a numeric vector}
    \item{\code{K.Hippocrepis.comosa}}{a numeric vector}
    \item{\code{K.Oxalis.acetosella.}}{a numeric vector}
    \item{\code{K.Geranium.sylvaticum.}}{a numeric vector}
    \item{\code{K.Geranium.dissectum}}{a numeric vector}
    \item{\code{K.Geranium.robertianum}}{a numeric vector}
    \item{\code{K.Linum.catharticum.}}{a numeric vector}
    \item{\code{K.Mercurialis.perennis}}{a numeric vector}
    \item{\code{K.Euphorbia.dulcis}}{a numeric vector}
    \item{\code{K.Euphorbia.cyparissias.}}{a numeric vector}
    \item{\code{K.Euphorbia.amygdaloides}}{a numeric vector}
    \item{\code{K.Polygala.chamaebuxus}}{a numeric vector}
    \item{\code{K.Polygala.vulgaris.}}{a numeric vector}
    \item{\code{K.Polygala.alpestris}}{a numeric vector}
    \item{\code{B.Acer.platanoides}}{a numeric vector}
    \item{\code{S.Acer.platanoides}}{a numeric vector}
    \item{\code{K.Acer.platanoides}}{a numeric vector}
    \item{\code{B.Acer.campestre}}{a numeric vector}
    \item{\code{S.Acer.campestre}}{a numeric vector}
    \item{\code{K.Acer.campestre}}{a numeric vector}
    \item{\code{B.Acer.pseudoplatanus.}}{a numeric vector}
    \item{\code{S.Acer.pseudoplatanus.}}{a numeric vector}
    \item{\code{K.Acer.pseudoplatanus.}}{a numeric vector}
    \item{\code{B.Acer.opalus.}}{a numeric vector}
    \item{\code{S.Acer.opalus.}}{a numeric vector}
    \item{\code{K.Acer.opalus.}}{a numeric vector}
    \item{\code{K.Impatiens.noli.tangere}}{a numeric vector}
    \item{\code{K.Impatiens.parviflora}}{a numeric vector}
    \item{\code{B.Ilex.aquifolium.}}{a numeric vector}
    \item{\code{S.Ilex.aquifolium.}}{a numeric vector}
    \item{\code{K.Ilex.aquifolium.}}{a numeric vector}
    \item{\code{S.Euonymus.europaeus}}{a numeric vector}
    \item{\code{K.Euonymus.europaeus}}{a numeric vector}
    \item{\code{K.Euonymus.latifolius.}}{a numeric vector}
    \item{\code{K.Rhamnus.catharticus.}}{a numeric vector}
    \item{\code{S.Rhamnus.alpinus.}}{a numeric vector}
    \item{\code{K.Rhamnus.alpinus.}}{a numeric vector}
    \item{\code{S.Frangula.alnus}}{a numeric vector}
    \item{\code{K.Frangula.alnus}}{a numeric vector}
    \item{\code{B.Tilia.platyphyllos}}{a numeric vector}
    \item{\code{S.Tilia.platyphyllos}}{a numeric vector}
    \item{\code{K.Tilia.platyphyllos}}{a numeric vector}
    \item{\code{B.Tilia.cordata.}}{a numeric vector}
    \item{\code{S.Tilia.cordata.}}{a numeric vector}
    \item{\code{K.Tilia.cordata.}}{a numeric vector}
    \item{\code{S.Daphne.mezereum.}}{a numeric vector}
    \item{\code{K.Daphne.mezereum.}}{a numeric vector}
    \item{\code{S.Daphne.laureola.}}{a numeric vector}
    \item{\code{K.Daphne.laureola.}}{a numeric vector}
    \item{\code{S.Hippophae.rhamnoides}}{a numeric vector}
    \item{\code{K.Hypericum.hirsutum}}{a numeric vector}
    \item{\code{K.Hypericum.montanum}}{a numeric vector}
    \item{\code{K.Hypericum.maculatum.}}{a numeric vector}
    \item{\code{K.Hypericum.perforatum}}{a numeric vector}
    \item{\code{K.Viola.hirta.}}{a numeric vector}
    \item{\code{K.Viola.mirabilis.}}{a numeric vector}
    \item{\code{K.Viola.reichenbachiana.}}{a numeric vector}
    \item{\code{K.Viola.biflora.}}{a numeric vector}
    \item{\code{K.Helianthemum.nummularium}}{a numeric vector}
    \item{\code{K.Circaea.lutetiana.}}{a numeric vector}
    \item{\code{K.Epilobium.angustifolium.}}{a numeric vector}
    \item{\code{K.Epilobium.hirsutum}}{a numeric vector}
    \item{\code{K.Epilobium.parviflorum.}}{a numeric vector}
    \item{\code{K.Epilobium.montanum}}{a numeric vector}
    \item{\code{K.Epilobium.alpestre}}{a numeric vector}
    \item{\code{S.Cornus.sanguinea}}{a numeric vector}
    \item{\code{K.Cornus.sanguinea}}{a numeric vector}
    \item{\code{B.Hedera.helix}}{a numeric vector}
    \item{\code{S.Hedera.helix}}{a numeric vector}
    \item{\code{K.Hedera.helix}}{a numeric vector}
    \item{\code{K.Sanicula.europaea.}}{a numeric vector}
    \item{\code{K.Astrantia.major.}}{a numeric vector}
    \item{\code{K.Chaerophyllum.hirsutum}}{a numeric vector}
    \item{\code{K.Chaerophyllum.villarsii.}}{a numeric vector}
    \item{\code{K.Anthriscus.sylvestris.}}{a numeric vector}
    \item{\code{K.Pimpinella.major}}{a numeric vector}
    \item{\code{K.Pimpinella.saxifraga}}{a numeric vector}
    \item{\code{K.Aegopodium.podagraria.}}{a numeric vector}
    \item{\code{K.Bupleurum.falcatum}}{a numeric vector}
    \item{\code{K.Carum.carvi.}}{a numeric vector}
    \item{\code{K.Angelica.sylvestris.}}{a numeric vector}
    \item{\code{K.Peucedanum.ostruthium.}}{a numeric vector}
    \item{\code{K.Heracleum.sphondylium.sspalpinum}}{a numeric vector}
    \item{\code{K.Heracleum.sphondylium.}}{a numeric vector}
    \item{\code{K.Laserpitium.latifolium}}{a numeric vector}
    \item{\code{K.Torilis.japonica}}{a numeric vector}
    \item{\code{K.Daucus.carota.}}{a numeric vector}
    \item{\code{K.Orthilia.secunda}}{a numeric vector}
    \item{\code{K.Monotropa.hypopitys.}}{a numeric vector}
    \item{\code{K.Calluna.vulgaris}}{a numeric vector}
    \item{\code{K.Arctostaphylos.uva.ursi.}}{a numeric vector}
    \item{\code{K.Vaccinium.microcarpum.}}{a numeric vector}
    \item{\code{K.Vaccinium.vitis.idaea.}}{a numeric vector}
    \item{\code{S.Vaccinium.myrtillus.}}{a numeric vector}
    \item{\code{K.Vaccinium.myrtillus.}}{a numeric vector}
    \item{\code{K.Primula.elatior.}}{a numeric vector}
    \item{\code{K.Primula.veris.sspsuaveolens.}}{a numeric vector}
    \item{\code{K.Soldanella.alpina.}}{a numeric vector}
    \item{\code{K.Lysimachia.nemorum}}{a numeric vector}
    \item{\code{K.Lysimachia.vulgaris.}}{a numeric vector}
    \item{\code{B.Fraxinus.excelsior}}{a numeric vector}
    \item{\code{S.Fraxinus.excelsior}}{a numeric vector}
    \item{\code{K.Fraxinus.excelsior}}{a numeric vector}
    \item{\code{S.Ligustrum.vulgare.}}{a numeric vector}
    \item{\code{K.Ligustrum.vulgare.}}{a numeric vector}
    \item{\code{K.Gentiana.lutea}}{a numeric vector}
    \item{\code{K.Gentiana.purpurea.}}{a numeric vector}
    \item{\code{K.Gentiana.asclepiadea}}{a numeric vector}
    \item{\code{K.Vinca.minor.}}{a numeric vector}
    \item{\code{K.Vincetoxicum.hirudinaria}}{a numeric vector}
    \item{\code{K.Calystegia.sepium.}}{a numeric vector}
    \item{\code{K.Pulmonaria.obscura}}{a numeric vector}
    \item{\code{K.Pulmonaria.officinalis}}{a numeric vector}
    \item{\code{K.Symphytum.officinale}}{a numeric vector}
    \item{\code{K.Myosotis.arvensis.}}{a numeric vector}
    \item{\code{K.Myosotis.sylvatica}}{a numeric vector}
    \item{\code{K.Myosotis.scorpioides}}{a numeric vector}
    \item{\code{K.Verbena.officinalis.}}{a numeric vector}
    \item{\code{K.Ajuga.reptans.}}{a numeric vector}
    \item{\code{K.Teucrium.scorodonia.}}{a numeric vector}
    \item{\code{K.Teucrium.chamaedrys.}}{a numeric vector}
    \item{\code{K.Melittis.melissophyllum.}}{a numeric vector}
    \item{\code{K.Galeopsis.tetrahit}}{a numeric vector}
    \item{\code{K.Lamium.maculatum}}{a numeric vector}
    \item{\code{K.Lamiastrum.montanum.}}{a numeric vector}
    \item{\code{K.Stachys.alpina}}{a numeric vector}
    \item{\code{K.Stachys.sylvatica.}}{a numeric vector}
    \item{\code{K.Stachys.recta.}}{a numeric vector}
    \item{\code{K.Glechoma.hederacea}}{a numeric vector}
    \item{\code{K.Dracocephalum.austriacum}}{a numeric vector}
    \item{\code{K.Prunella.vulgaris.}}{a numeric vector}
    \item{\code{K.Acinos.arvensis.}}{a numeric vector}
    \item{\code{K.Clinopodium.vulgare.}}{a numeric vector}
    \item{\code{K.Origanum.vulgare}}{a numeric vector}
    \item{\code{K.Thymus.oenipontanus.sspdecipiens}}{a numeric vector}
    \item{\code{K.Thymus.serpyllum}}{a numeric vector}
    \item{\code{K.Mentha.arvensis.}}{a numeric vector}
    \item{\code{K.Mentha.longifolia.}}{a numeric vector}
    \item{\code{K.Salvia.glutinosa}}{a numeric vector}
    \item{\code{K.Salvia.pratensis}}{a numeric vector}
    \item{\code{K.Atropa.belladonna.}}{a numeric vector}
    \item{\code{K.Verbascum.nigrum}}{a numeric vector}
    \item{\code{K.Scrophularia.nodosa.}}{a numeric vector}
    \item{\code{K.Chaenorrhinum.minus.}}{a numeric vector}
    \item{\code{K.Digitalis.grandiflora.}}{a numeric vector}
    \item{\code{K.Digitalis.lutea.}}{a numeric vector}
    \item{\code{K.Veronica.serpyllifolia.s.l}}{a numeric vector}
    \item{\code{K.Veronica.urticifolia}}{a numeric vector}
    \item{\code{K.Veronica.teucrium.}}{a numeric vector}
    \item{\code{K.Veronica.officinalis}}{a numeric vector}
    \item{\code{K.Veronica.chamaedrys.}}{a numeric vector}
    \item{\code{K.Veronica.montana}}{a numeric vector}
    \item{\code{K.Veronica.beccabunga.}}{a numeric vector}
    \item{\code{K.Veronica.hederifolia}}{a numeric vector}
    \item{\code{K.Melampyrum.arvense}}{a numeric vector}
    \item{\code{K.Melampyrum.sylvaticum.}}{a numeric vector}
    \item{\code{K.Euphrasia.rostkoviana.}}{a numeric vector}
    \item{\code{K.Rhinanthus.minor}}{a numeric vector}
    \item{\code{K.Rhinanthus.alectorolophus.}}{a numeric vector}
    \item{\code{K.Asperula.taurina}}{a numeric vector}
    \item{\code{K.Galium.rotundifolium}}{a numeric vector}
    \item{\code{K.Galium.odoratum.}}{a numeric vector}
    \item{\code{K.Galium.palustre.}}{a numeric vector}
    \item{\code{K.Galium.mollugo.s.l}}{a numeric vector}
    \item{\code{K.Galium.sylvaticum.}}{a numeric vector}
    \item{\code{K.Galium.pumilum}}{a numeric vector}
    \item{\code{K.Galium.anisophyllon.}}{a numeric vector}
    \item{\code{K.Galium.aparine}}{a numeric vector}
    \item{\code{K.Cruciata.laevipes.}}{a numeric vector}
    \item{\code{K.Plantago.major}}{a numeric vector}
    \item{\code{K.Plantago.media}}{a numeric vector}
    \item{\code{K.Plantago.lanceolata.}}{a numeric vector}
    \item{\code{S.Sambucus.ebulus.}}{a numeric vector}
    \item{\code{K.Sambucus.ebulus.}}{a numeric vector}
    \item{\code{S.Sambucus.nigra}}{a numeric vector}
    \item{\code{K.Sambucus.nigra}}{a numeric vector}
    \item{\code{S.Sambucus.racemosa.}}{a numeric vector}
    \item{\code{K.Sambucus.racemosa.}}{a numeric vector}
    \item{\code{S.Viburnum.opulus.}}{a numeric vector}
    \item{\code{K.Viburnum.opulus.}}{a numeric vector}
    \item{\code{S.Viburnum.lantana}}{a numeric vector}
    \item{\code{K.Viburnum.lantana}}{a numeric vector}
    \item{\code{S.Lonicera.alpigena.}}{a numeric vector}
    \item{\code{K.Lonicera.alpigena.}}{a numeric vector}
    \item{\code{S.Lonicera.nigra}}{a numeric vector}
    \item{\code{K.Lonicera.nigra}}{a numeric vector}
    \item{\code{S.Lonicera.xylosteum}}{a numeric vector}
    \item{\code{K.Lonicera.xylosteum}}{a numeric vector}
    \item{\code{K.Adoxa.moschatellina.}}{a numeric vector}
    \item{\code{K.Valeriana.officinalis.s.l.}}{a numeric vector}
    \item{\code{K.Valeriana.dioica}}{a numeric vector}
    \item{\code{K.Valeriana.tripteris.}}{a numeric vector}
    \item{\code{K.Valeriana.montana.}}{a numeric vector}
    \item{\code{K.Dipsacus.pilosus}}{a numeric vector}
    \item{\code{K.Succisa.pratensis.}}{a numeric vector}
    \item{\code{K.Knautia.dipsacifolia}}{a numeric vector}
    \item{\code{K.Knautia.arvensis}}{a numeric vector}
    \item{\code{K.Scabiosa.columbaria.}}{a numeric vector}
    \item{\code{K.Campanula.trachelium}}{a numeric vector}
    \item{\code{K.Campanula.cochleariifolia.}}{a numeric vector}
    \item{\code{K.Campanula.scheuchzeri.}}{a numeric vector}
    \item{\code{K.Campanula.rotundifolia}}{a numeric vector}
    \item{\code{K.Phyteuma.spicatum.}}{a numeric vector}
    \item{\code{K.Phyteuma.ovatum.}}{a numeric vector}
    \item{\code{K.Phyteuma.orbiculare.}}{a numeric vector}
    \item{\code{K.Eupatorium.cannabinum.}}{a numeric vector}
    \item{\code{K.Solidago.virgaurea}}{a numeric vector}
    \item{\code{K.Solidago.canadensis.}}{a numeric vector}
    \item{\code{K.Bellis.perennis.}}{a numeric vector}
    \item{\code{K.Aster.bellidiastrum.}}{a numeric vector}
    \item{\code{K.Achillea.macrophylla}}{a numeric vector}
    \item{\code{K.Achillea.millefolium}}{a numeric vector}
    \item{\code{K.Leucanthemum.vulgare}}{a numeric vector}
    \item{\code{K.Tussilago.farfara.}}{a numeric vector}
    \item{\code{K.Petasites.albus.}}{a numeric vector}
    \item{\code{K.Homogyne.alpina.}}{a numeric vector}
    \item{\code{K.Adenostyles.alliariae.}}{a numeric vector}
    \item{\code{K.Adenostyles.glabra}}{a numeric vector}
    \item{\code{K.Senecio.fuchsii.}}{a numeric vector}
    \item{\code{K.Senecio.alpinus.}}{a numeric vector}
    \item{\code{K.Senecio.jacobaea}}{a numeric vector}
    \item{\code{K.Carlina.vulgaris.s.l}}{a numeric vector}
    \item{\code{K.Carlina.acaulis.}}{a numeric vector}
    \item{\code{K.Arctium.nemorosum.}}{a numeric vector}
    \item{\code{K.Carduus.nutans.s.l}}{a numeric vector}
    \item{\code{K.Carduus.personata.}}{a numeric vector}
    \item{\code{K.Carduus.defloratus}}{a numeric vector}
    \item{\code{K.Carduus.defloratus.sspcrassifolius}}{a numeric vector}
    \item{\code{K.Cirsium.vulgare.}}{a numeric vector}
    \item{\code{K.Cirsium.oleraceum.}}{a numeric vector}
    \item{\code{K.Cirsium.palustre}}{a numeric vector}
    \item{\code{K.Cirsium.arvense.}}{a numeric vector}
    \item{\code{K.Centaurea.scabiosa}}{a numeric vector}
    \item{\code{K.Centaurea.jacea.}}{a numeric vector}
    \item{\code{K.Centaurea.montana.}}{a numeric vector}
    \item{\code{K.Hypochoeris.radicata}}{a numeric vector}
    \item{\code{K.Leontodon.helveticus}}{a numeric vector}
    \item{\code{K.Leontodon.autumnalis}}{a numeric vector}
    \item{\code{K.Leontodon.hispidus}}{a numeric vector}
    \item{\code{K.Picris.hieracioides.s.l.}}{a numeric vector}
    \item{\code{K.Sonchus.asper.}}{a numeric vector}
    \item{\code{K.Sonchus.oleraceus.}}{a numeric vector}
    \item{\code{K.Cicerbita.alpina}}{a numeric vector}
    \item{\code{K.Prenanthes.purpurea.}}{a numeric vector}
    \item{\code{K.Mycelis.muralis.}}{a numeric vector}
    \item{\code{K.Taraxacum.officinale}}{a numeric vector}
    \item{\code{K.Lapsana.communis}}{a numeric vector}
    \item{\code{K.Crepis.paludosa.}}{a numeric vector}
    \item{\code{K.Crepis.mollis.}}{a numeric vector}
    \item{\code{K.Crepis.alpestris}}{a numeric vector}
    \item{\code{K.Crepis.biennis}}{a numeric vector}
    \item{\code{K.Hieracium.pilosella.}}{a numeric vector}
    \item{\code{K.Hieracium.murorum.}}{a numeric vector}
    \item{\code{K.Hieracium.prenanthoides.}}{a numeric vector}
    \item{\code{K.Veratrum.album}}{a numeric vector}
    \item{\code{K.Anthericum.liliago}}{a numeric vector}
    \item{\code{K.Colchicum.autumnale.}}{a numeric vector}
    \item{\code{K.Gagea.lutea.}}{a numeric vector}
    \item{\code{K.Lilium.martagon.}}{a numeric vector}
    \item{\code{K.Allium.ursinum}}{a numeric vector}
    \item{\code{K.Allium.carinatum}}{a numeric vector}
    \item{\code{K.Convallaria.majalis.}}{a numeric vector}
    \item{\code{K.Maianthemum.bifolium}}{a numeric vector}
    \item{\code{K.Polygonatum.verticillatum.}}{a numeric vector}
    \item{\code{K.Polygonatum.multiflorum.}}{a numeric vector}
    \item{\code{K.Polygonatum.odoratum}}{a numeric vector}
    \item{\code{K.Paris.quadrifolia.}}{a numeric vector}
    \item{\code{K.Leucojum.vernum.}}{a numeric vector}
    \item{\code{K.Narcissus.pseudo.narcissus}}{a numeric vector}
    \item{\code{S.Tamus.communis}}{a numeric vector}
    \item{\code{K.Tamus.communis}}{a numeric vector}
    \item{\code{K.Crocus.albiflorus.}}{a numeric vector}
    \item{\code{K.Juncus.inflexus.}}{a numeric vector}
    \item{\code{K.Juncus.effusus}}{a numeric vector}
    \item{\code{K.Juncus.bufonius.}}{a numeric vector}
    \item{\code{K.Luzula.multiflora.}}{a numeric vector}
    \item{\code{K.Luzula.sylvatica.s.l}}{a numeric vector}
    \item{\code{K.Luzula.nivea}}{a numeric vector}
    \item{\code{K.Luzula.luzuloides.}}{a numeric vector}
    \item{\code{K.Luzula.pilosa.}}{a numeric vector}
    \item{\code{K.Luzula.luzulina.}}{a numeric vector}
    \item{\code{K.Festuca.altissima.}}{a numeric vector}
    \item{\code{K.Festuca.gigantea}}{a numeric vector}
    \item{\code{K.Festuca.pratensis.}}{a numeric vector}
    \item{\code{K.Festuca.heterophylla}}{a numeric vector}
    \item{\code{K.Festuca.rubra.s.l.}}{a numeric vector}
    \item{\code{K.Festuca.ovina.s.l.}}{a numeric vector}
    \item{\code{K.Lolium.perenne}}{a numeric vector}
    \item{\code{K.Poa.annua.}}{a numeric vector}
    \item{\code{K.Poa.trivialis.}}{a numeric vector}
    \item{\code{K.Poa.pratensis.s.l.}}{a numeric vector}
    \item{\code{K.Poa.chaixii.}}{a numeric vector}
    \item{\code{K.Poa.compressa.}}{a numeric vector}
    \item{\code{K.Poa.nemoralis.}}{a numeric vector}
    \item{\code{K.Poa.alpina}}{a numeric vector}
    \item{\code{K.Dactylis.glomerata}}{a numeric vector}
    \item{\code{K.Cynosurus.cristatus.}}{a numeric vector}
    \item{\code{K.Briza.media.}}{a numeric vector}
    \item{\code{K.Sesleria.caerulea.}}{a numeric vector}
    \item{\code{K.Melica.nutans.}}{a numeric vector}
    \item{\code{K.Melica.uniflora.}}{a numeric vector}
    \item{\code{K.Glyceria.fluitans.}}{a numeric vector}
    \item{\code{K.Glyceria.plicata}}{a numeric vector}
    \item{\code{K.Bromus.sterilis.}}{a numeric vector}
    \item{\code{K.Bromus.tectorum.}}{a numeric vector}
    \item{\code{K.Bromus.inermis}}{a numeric vector}
    \item{\code{K.Bromus.ramosus}}{a numeric vector}
    \item{\code{K.Bromus.benekenii}}{a numeric vector}
    \item{\code{K.Bromus.erectus}}{a numeric vector}
    \item{\code{K.Bromus.hordeaceus.}}{a numeric vector}
    \item{\code{K.Brachypodium.sylvaticum.}}{a numeric vector}
    \item{\code{K.Brachypodium.pinnatum.}}{a numeric vector}
    \item{\code{K.Agropyron.caninum.}}{a numeric vector}
    \item{\code{K.Agropyron.pungens.}}{a numeric vector}
    \item{\code{K.Hordelymus.europaeus}}{a numeric vector}
    \item{\code{K.Avenula.pubescens.}}{a numeric vector}
    \item{\code{K.Arrhenatherum.elatius.}}{a numeric vector}
    \item{\code{K.Koeleria.pyramidata.}}{a numeric vector}
    \item{\code{K.Trisetum.flavescens.}}{a numeric vector}
    \item{\code{K.Deschampsia.cespitosa.}}{a numeric vector}
    \item{\code{K.Avenella.flexuosa.}}{a numeric vector}
    \item{\code{K.Anthoxanthum.odoratum.}}{a numeric vector}
    \item{\code{K.Anthoxanthum.alpinum}}{a numeric vector}
    \item{\code{K.Holcus.lanatus}}{a numeric vector}
    \item{\code{K.Agrostis.tenuis.}}{a numeric vector}
    \item{\code{K.Agrostis.gigantea.}}{a numeric vector}
    \item{\code{K.Agrostis.stolonifera}}{a numeric vector}
    \item{\code{K.Calamagrostis.epigeios}}{a numeric vector}
    \item{\code{K.Calamagrostis.villosa.}}{a numeric vector}
    \item{\code{K.Calamagrostis.arundinacea.}}{a numeric vector}
    \item{\code{K.Calamagrostis.varia.}}{a numeric vector}
    \item{\code{K.Phleum.pratense.}}{a numeric vector}
    \item{\code{K.Phalaris.arundinacea}}{a numeric vector}
    \item{\code{K.Milium.effusum}}{a numeric vector}
    \item{\code{K.Danthonia.decumbens.}}{a numeric vector}
    \item{\code{K.Molinia.arundinacea.}}{a numeric vector}
    \item{\code{K.Nardus.stricta}}{a numeric vector}
    \item{\code{K.Arum.maculatum}}{a numeric vector}
    \item{\code{K.Lemna.minor.}}{a numeric vector}
    \item{\code{K.Typha.latifolia.}}{a numeric vector}
    \item{\code{K.Scirpus.sylvaticus}}{a numeric vector}
    \item{\code{K.Carex.diandra.}}{a numeric vector}
    \item{\code{K.Carex.spicata.}}{a numeric vector}
    \item{\code{K.Carex.pairae}}{a numeric vector}
    \item{\code{K.Carex.divulsa.}}{a numeric vector}
    \item{\code{K.Carex.leersii.}}{a numeric vector}
    \item{\code{K.Carex.brizoides.}}{a numeric vector}
    \item{\code{K.Carex.remota}}{a numeric vector}
    \item{\code{K.Carex.leporina}}{a numeric vector}
    \item{\code{K.Carex.echinata}}{a numeric vector}
    \item{\code{K.Carex.elongata}}{a numeric vector}
    \item{\code{K.Carex.hirta.}}{a numeric vector}
    \item{\code{K.Carex.acutiformis.}}{a numeric vector}
    \item{\code{K.Carex.pendula.}}{a numeric vector}
    \item{\code{K.Carex.sylvatica.}}{a numeric vector}
    \item{\code{K.Carex.flacca}}{a numeric vector}
    \item{\code{K.Carex.pilosa}}{a numeric vector}
    \item{\code{K.Carex.pallescens}}{a numeric vector}
    \item{\code{K.Carex.digitata}}{a numeric vector}
    \item{\code{K.Carex.ornithopoda.}}{a numeric vector}
    \item{\code{K.Carex.montana.}}{a numeric vector}
    \item{\code{K.Carex.pilulifera}}{a numeric vector}
    \item{\code{K.Carex.alba}}{a numeric vector}
    \item{\code{K.Carex.mucronata.}}{a numeric vector}
    \item{\code{K.Carex.ferruginea}}{a numeric vector}
    \item{\code{K.Carex.nigra.}}{a numeric vector}
    \item{\code{K.Epipactis.palustris.}}{a numeric vector}
    \item{\code{K.Epipactis.helleborine.}}{a numeric vector}
    \item{\code{K.Epipactis.atrorubens}}{a numeric vector}
    \item{\code{K.Cephalanthera.damasonium}}{a numeric vector}
    \item{\code{K.Neottia.nidus.avis}}{a numeric vector}
    \item{\code{K.Listera.ovata.}}{a numeric vector}
    \item{\code{K.Platanthera.bifolia.}}{a numeric vector}
    \item{\code{K.Gymnadenia.conopsea.}}{a numeric vector}
    \item{\code{K.Pseudorchis.albida}}{a numeric vector}
    \item{\code{K.Coeloglossum.viride.}}{a numeric vector}
    \item{\code{K.Dactylorhiza.maculata.}}{a numeric vector}
    \item{\code{K.Orchis.ustulata.}}{a numeric vector}
    \item{\code{S.Ribes.rubrum}}{a numeric vector}
    \item{\code{K.Ribes.rubrum}}{a numeric vector}
    \item{\code{S.Prunus.laurocerasus.}}{a numeric vector}
    \item{\code{K.Prunus.laurocerasus.}}{a numeric vector}
    \item{\code{B.Quercus.rubra.}}{a numeric vector}
    \item{\code{K.Quercus.rubra.}}{a numeric vector}
    \item{\code{B.Pseudotsuga.menziesii.}}{a numeric vector}
    \item{\code{S.Pseudotsuga.menziesii.}}{a numeric vector}
    \item{\code{K.Pseudotsuga.menziesii.}}{a numeric vector}
    \item{\code{B.Pinus.strobus.}}{a numeric vector}
    \item{\code{S.Pinus.strobus.}}{a numeric vector}
    \item{\code{K.Pinus.strobus.}}{a numeric vector}
    \item{\code{K.Orchis.spec.}}{a numeric vector}
    \item{\code{K.Compositae.spec.}}{a numeric vector}
    \item{\code{S.Cotoneaster.horizontalis}}{a numeric vector}
    \item{\code{K.Cotoneaster.horizontalis}}{a numeric vector}
    \item{\code{K.Rubus.spec}}{a numeric vector}
    \item{\code{S.Thuja.plicata.}}{a numeric vector}
    \item{\code{M.Atrichum.undulatum}}{a numeric vector}
    \item{\code{M.Cirriphyllum.piliferum}}{a numeric vector}
    \item{\code{M.Ctenidium.molluscum.}}{a numeric vector}
    \item{\code{M.Dicranella.heteromalla}}{a numeric vector}
    \item{\code{M.Dicranum.scoparium}}{a numeric vector}
    \item{\code{M.Eurhynchium.striatum}}{a numeric vector}
    \item{\code{M.Eurhynchium.swartzii}}{a numeric vector}
    \item{\code{M.Fissidens.taxifolius}}{a numeric vector}
    \item{\code{M.Hylocomium.splendens}}{a numeric vector}
    \item{\code{M.Leucobryum.glaucum}}{a numeric vector}
    \item{\code{M.Mnium.undulatum.}}{a numeric vector}
    \item{\code{M.Pleurozium.schreberi}}{a numeric vector}
    \item{\code{M.Polytrichum.formosum}}{a numeric vector}
    \item{\code{M.Ptilium.crista.castrensis.}}{a numeric vector}
    \item{\code{M.Rhytidiadelphus.loreus}}{a numeric vector}
    \item{\code{M.Rhytidiadelphus.triquetrus}}{a numeric vector}
    \item{\code{M.Thuidium.tamariscinum.}}{a numeric vector}
    \item{\code{M.Bazzania.trilobata}}{a numeric vector}
    \item{\code{M.Hypnum.cupressiforme}}{a numeric vector}
    \item{\code{M.Sphagnum.spec.}}{a numeric vector}
    \item{\code{M.Polytrichum.spec}}{a numeric vector}
    \item{\code{M.Mnium.affine}}{a numeric vector}
    \item{\code{M.Plagiochila.asplenioides}}{a numeric vector}
    \item{\code{M.Rhizomnium.punctatum}}{a numeric vector}
    \item{\code{M.Rhodobryum.roseum.}}{a numeric vector}
    \item{\code{M.Lophocolea.bidentata}}{a numeric vector}
    \item{\code{M.Pellia.spec.}}{a numeric vector}
    \item{\code{M.Moose.spec}}{a numeric vector}
    \item{\code{K.Lycopodium.annotinum}}{a numeric vector}
    \item{\code{K.Lycopodium.clavatum.}}{a numeric vector}
    \item{\code{K.Selaginella.selaginoides}}{a numeric vector}
    \item{\code{K.Equisetum.hiemale.}}{a numeric vector}
    \item{\code{K.Botrychium.lunaria}}{a numeric vector}
    \item{\code{K.Asplenium.fontanum}}{a numeric vector}
    \item{\code{K.Asplenium.ruta.muraria}}{a numeric vector}
    \item{\code{K.Dryopteris.pseudomas}}{a numeric vector}
    \item{\code{K.Dryopteris.villarii.}}{a numeric vector}
    \item{\code{K.Dryopteris.expansa}}{a numeric vector}
    \item{\code{B.Pinus.mugo.ssparborea.}}{a numeric vector}
    \item{\code{S.Pinus.mugo.ssparborea.}}{a numeric vector}
    \item{\code{K.Pinus.mugo.ssparborea.}}{a numeric vector}
    \item{\code{B.Pinus.mugo.sspprostrata.}}{a numeric vector}
    \item{\code{S.Pinus.mugo.sspprostrata.}}{a numeric vector}
    \item{\code{K.Pinus.mugo.sspprostrata.}}{a numeric vector}
    \item{\code{B.Pinus.cembra}}{a numeric vector}
    \item{\code{S.Pinus.cembra}}{a numeric vector}
    \item{\code{K.Pinus.cembra}}{a numeric vector}
    \item{\code{K.Juniperus.communis.sspalpina}}{a numeric vector}
    \item{\code{S.Salix.triandra}}{a numeric vector}
    \item{\code{K.Salix.triandra}}{a numeric vector}
    \item{\code{K.Salix.retusa}}{a numeric vector}
    \item{\code{S.Salix.nigricans.}}{a numeric vector}
    \item{\code{K.Salix.nigricans.}}{a numeric vector}
    \item{\code{B.Salix.appendiculata.}}{a numeric vector}
    \item{\code{K.Salix.appendiculata.}}{a numeric vector}
    \item{\code{B.Salix.elaeagnos.}}{a numeric vector}
    \item{\code{S.Salix.elaeagnos.}}{a numeric vector}
    \item{\code{K.Salix.elaeagnos.}}{a numeric vector}
    \item{\code{K.Alnus.viridis.}}{a numeric vector}
    \item{\code{S.Alnus.incana}}{a numeric vector}
    \item{\code{B.Castanea.sativa.}}{a numeric vector}
    \item{\code{S.Castanea.sativa.}}{a numeric vector}
    \item{\code{K.Castanea.sativa.}}{a numeric vector}
    \item{\code{K.Polygonum.viviparum.}}{a numeric vector}
    \item{\code{K.Rumex.scutatus}}{a numeric vector}
    \item{\code{K.Rumex.crispus.}}{a numeric vector}
    \item{\code{K.Minuartia.rupestris.}}{a numeric vector}
    \item{\code{K.Cerastium.arvense.}}{a numeric vector}
    \item{\code{K.Cerastium.alpinum.}}{a numeric vector}
    \item{\code{K.Silene.vulgaris.sspglareosa.}}{a numeric vector}
    \item{\code{K.Silene.acaulis}}{a numeric vector}
    \item{\code{K.Silene.rupestris}}{a numeric vector}
    \item{\code{K.Silene.pusilla}}{a numeric vector}
    \item{\code{K.Saponaria.ocymoides.}}{a numeric vector}
    \item{\code{K.Dianthus.superbus.}}{a numeric vector}
    \item{\code{K.Anemone.narcissiflora.}}{a numeric vector}
    \item{\code{K.Pulsatilla.alpina.}}{a numeric vector}
    \item{\code{K.Pulsatilla.vernalis.}}{a numeric vector}
    \item{\code{K.Clematis.alpina.}}{a numeric vector}
    \item{\code{K.Ranunculus.alpestris}}{a numeric vector}
    \item{\code{K.Ranunculus.platanifolius}}{a numeric vector}
    \item{\code{K.Thalictrum.minus}}{a numeric vector}
    \item{\code{K.Chelidonium.majus.}}{a numeric vector}
    \item{\code{K.Hesperis.matronalis.}}{a numeric vector}
    \item{\code{K.Cardamine.impatiens.}}{a numeric vector}
    \item{\code{K.Arabis.serpyllifolia}}{a numeric vector}
    \item{\code{K.Arabis.pumila.}}{a numeric vector}
    \item{\code{K.Lunaria.rediviva}}{a numeric vector}
    \item{\code{K.Draba.tomentosa.}}{a numeric vector}
    \item{\code{K.Draba.stylaris}}{a numeric vector}
    \item{\code{K.Kernera.saxatilis.}}{a numeric vector}
    \item{\code{K.Biscutella.laevigata}}{a numeric vector}
    \item{\code{K.Lepidium.ruderale.}}{a numeric vector}
    \item{\code{K.Erucastrum.nasturtiifolium}}{a numeric vector}
    \item{\code{K.Sempervivum.montanum}}{a numeric vector}
    \item{\code{K.Sempervivum.tectorum}}{a numeric vector}
    \item{\code{K.Sedum.maximum.}}{a numeric vector}
    \item{\code{K.Sedum.album.}}{a numeric vector}
    \item{\code{K.Saxifraga.cuneifolia}}{a numeric vector}
    \item{\code{K.Saxifraga.aspera}}{a numeric vector}
    \item{\code{K.Saxifraga.aizoides}}{a numeric vector}
    \item{\code{K.Saxifraga.paniculata}}{a numeric vector}
    \item{\code{K.Parnassia.palustris.}}{a numeric vector}
    \item{\code{S.Rosa.pendulina}}{a numeric vector}
    \item{\code{K.Sanguisorba.officinalis.}}{a numeric vector}
    \item{\code{K.Dryas.octopetala}}{a numeric vector}
    \item{\code{K.Potentilla.anserina.}}{a numeric vector}
    \item{\code{K.Potentilla.grandiflora}}{a numeric vector}
    \item{\code{K.Alchemilla.alpina.s.l.}}{a numeric vector}
    \item{\code{S.Sorbus.chamaemespilus.}}{a numeric vector}
    \item{\code{K.Sorbus.chamaemespilus.}}{a numeric vector}
    \item{\code{S.Amelanchier.ovalis}}{a numeric vector}
    \item{\code{K.Amelanchier.ovalis}}{a numeric vector}
    \item{\code{S.Cotoneaster.integerrimus}}{a numeric vector}
    \item{\code{K.Cotoneaster.tomentosus}}{a numeric vector}
    \item{\code{S.Genista.germanica.}}{a numeric vector}
    \item{\code{K.Genista.germanica.}}{a numeric vector}
    \item{\code{S.Robinia.pseudoacacia}}{a numeric vector}
    \item{\code{K.Robinia.pseudoacacia}}{a numeric vector}
    \item{\code{K.Astragalus.onobrychis.}}{a numeric vector}
    \item{\code{K.Oxytropis.jacquinii.}}{a numeric vector}
    \item{\code{K.Vicia.incana}}{a numeric vector}
    \item{\code{K.Lathyrus.niger}}{a numeric vector}
    \item{\code{K.Lathyrus.linifolius.}}{a numeric vector}
    \item{\code{K.Trifolium.alpinum.}}{a numeric vector}
    \item{\code{K.Trifolium.montanum}}{a numeric vector}
    \item{\code{K.Trifolium.badium}}{a numeric vector}
    \item{\code{K.Trifolium.aureum}}{a numeric vector}
    \item{\code{K.Trifolium.arvense.}}{a numeric vector}
    \item{\code{K.Trifolium.incarnatum}}{a numeric vector}
    \item{\code{K.Lotus.uliginosus}}{a numeric vector}
    \item{\code{K.Onobrychis.montana}}{a numeric vector}
    \item{\code{K.Mercurialis.ovata.}}{a numeric vector}
    \item{\code{K.Euphorbia.maculata}}{a numeric vector}
    \item{\code{K.Polygala.alpina.}}{a numeric vector}
    \item{\code{S.Euonymus.latifolius.}}{a numeric vector}
    \item{\code{S.Buxus.sempervirens}}{a numeric vector}
    \item{\code{K.Buxus.sempervirens}}{a numeric vector}
    \item{\code{S.Rhamnus.catharticus.}}{a numeric vector}
    \item{\code{K.Rhamnus.pumilus.}}{a numeric vector}
    \item{\code{K.Daphne.striata}}{a numeric vector}
    \item{\code{K.Viola.alba}}{a numeric vector}
    \item{\code{K.Viola.collina.}}{a numeric vector}
    \item{\code{K.Viola.canina}}{a numeric vector}
    \item{\code{K.Helianthemum.grandiflorum.sspgrandiflo}}{a numeric vector}
    \item{\code{K.Lythrum.salicaria.}}{a numeric vector}
    \item{\code{K.Circaea.alpina}}{a numeric vector}
    \item{\code{K.Epilobium.collinum}}{a numeric vector}
    \item{\code{K.Epilobium.obscurum}}{a numeric vector}
    \item{\code{K.Astrantia.minor.}}{a numeric vector}
    \item{\code{K.Seseli.libanotis}}{a numeric vector}
    \item{\code{K.Ligusticum.mutellina}}{a numeric vector}
    \item{\code{K.Peucedanum.austriacum.}}{a numeric vector}
    \item{\code{K.Peucedanum.cervaria.}}{a numeric vector}
    \item{\code{K.Laserpitium.siler.}}{a numeric vector}
    \item{\code{K.Laserpitium.halleri.}}{a numeric vector}
    \item{\code{K.Pyrola.rotundifolia.}}{a numeric vector}
    \item{\code{K.Moneses.uniflora}}{a numeric vector}
    \item{\code{K.Erica.herbacea}}{a numeric vector}
    \item{\code{S.Rhododendron.ferrugineum}}{a numeric vector}
    \item{\code{K.Rhododendron.ferrugineum}}{a numeric vector}
    \item{\code{K.Rhododendron.hirsutum.}}{a numeric vector}
    \item{\code{K.Vaccinium.uliginosum.s.l}}{a numeric vector}
    \item{\code{K.Empetrum.nigrum.}}{a numeric vector}
    \item{\code{K.Primula.farinosa}}{a numeric vector}
    \item{\code{K.Primula.auricula}}{a numeric vector}
    \item{\code{K.Primula.hirsuta.}}{a numeric vector}
    \item{\code{K.Cyclamen.purpurascens.}}{a numeric vector}
    \item{\code{K.Gentiana.clusii.}}{a numeric vector}
    \item{\code{K.Gentiana.acaulis}}{a numeric vector}
    \item{\code{K.Gentiana.verna}}{a numeric vector}
    \item{\code{K.Gentiana.utriculosa.}}{a numeric vector}
    \item{\code{K.Gentiana.ciliata}}{a numeric vector}
    \item{\code{K.Gentianella.campestris}}{a numeric vector}
    \item{\code{K.Nymphoides.peltata}}{a numeric vector}
    \item{\code{K.Cuscuta.europaea}}{a numeric vector}
    \item{\code{K.Myosotis.alpestris}}{a numeric vector}
    \item{\code{K.Myosotis.strigulosa.}}{a numeric vector}
    \item{\code{K.Cynoglossum.officinale}}{a numeric vector}
    \item{\code{K.Ajuga.pyramidalis.}}{a numeric vector}
    \item{\code{K.Teucrium.montanum.}}{a numeric vector}
    \item{\code{K.Leonurus.cardiaca.}}{a numeric vector}
    \item{\code{K.Betonica.officinalis}}{a numeric vector}
    \item{\code{K.Prunella.grandiflora}}{a numeric vector}
    \item{\code{K.Acinos.alpinus}}{a numeric vector}
    \item{\code{K.Calamintha.grandiflora}}{a numeric vector}
    \item{\code{K.Thymus.froelichianus}}{a numeric vector}
    \item{\code{K.Lycopus.europaeus.}}{a numeric vector}
    \item{\code{K.Solanum.dulcamara.}}{a numeric vector}
    \item{\code{K.Verbascum.crassifolium}}{a numeric vector}
    \item{\code{K.Verbascum.chaixii.}}{a numeric vector}
    \item{\code{K.Veronica.alpina.}}{a numeric vector}
    \item{\code{K.Veronica.anagallis.aquatica.}}{a numeric vector}
    \item{\code{K.Melampyrum.pratense.}}{a numeric vector}
    \item{\code{K.Euphrasia.rostkoviana.sspmontana}}{a numeric vector}
    \item{\code{K.Euphrasia.hirtella}}{a numeric vector}
    \item{\code{K.Euphrasia.stricta.}}{a numeric vector}
    \item{\code{K.Euphrasia.minima}}{a numeric vector}
    \item{\code{K.Euphrasia.salisburgensis}}{a numeric vector}
    \item{\code{K.Bartsia.alpina}}{a numeric vector}
    \item{\code{K.Pedicularis.recutita}}{a numeric vector}
    \item{\code{K.Pedicularis.verticillata}}{a numeric vector}
    \item{\code{K.Rhinanthus.antiquus.}}{a numeric vector}
    \item{\code{K.Rhinanthus.glacialis}}{a numeric vector}
    \item{\code{K.Lathraea.squamaria}}{a numeric vector}
    \item{\code{K.Globularia.cordifolia.}}{a numeric vector}
    \item{\code{K.Globularia.nudicaulis.}}{a numeric vector}
    \item{\code{K.Pinguicula.alpina.}}{a numeric vector}
    \item{\code{K.Pinguicula.vulgaris.}}{a numeric vector}
    \item{\code{K.Galium.boreale}}{a numeric vector}
    \item{\code{K.Galium.verum}}{a numeric vector}
    \item{\code{K.Galium.lucidum}}{a numeric vector}
    \item{\code{K.Plantago.major.sspintermedia}}{a numeric vector}
    \item{\code{K.Plantago.alpina.}}{a numeric vector}
    \item{\code{K.Plantago.atrata.}}{a numeric vector}
    \item{\code{K.Linnaea.borealis}}{a numeric vector}
    \item{\code{S.Lonicera.caerulea.}}{a numeric vector}
    \item{\code{K.Lonicera.caerulea.}}{a numeric vector}
    \item{\code{S.Lonicera.periclymenum.}}{a numeric vector}
    \item{\code{K.Lonicera.periclymenum.}}{a numeric vector}
    \item{\code{K.Scabiosa.lucida.}}{a numeric vector}
    \item{\code{K.Campanula.barbata.}}{a numeric vector}
    \item{\code{K.Campanula.glomerata.}}{a numeric vector}
    \item{\code{K.Campanula.rapunculoides.}}{a numeric vector}
    \item{\code{K.Campanula.rhomboidalis}}{a numeric vector}
    \item{\code{K.Phyteuma.betonicifolium.}}{a numeric vector}
    \item{\code{K.Phyteuma.hemisphaericum.}}{a numeric vector}
    \item{\code{K.Aster.alpinus.}}{a numeric vector}
    \item{\code{K.Erigeron.annuus.sspannuus.}}{a numeric vector}
    \item{\code{K.Erigeron.alpinus}}{a numeric vector}
    \item{\code{K.Gnaphalium.sylvaticum.}}{a numeric vector}
    \item{\code{K.Antennaria.dioica.}}{a numeric vector}
    \item{\code{K.Buphthalmum.salicifolium}}{a numeric vector}
    \item{\code{K.Petasites.hybridus}}{a numeric vector}
    \item{\code{K.Petasites.paradoxus.}}{a numeric vector}
    \item{\code{K.Arnica.montana}}{a numeric vector}
    \item{\code{K.Senecio.abrotanifolius}}{a numeric vector}
    \item{\code{K.Senecio.viscosus}}{a numeric vector}
    \item{\code{K.Arctium.lappa.}}{a numeric vector}
    \item{\code{K.Cirsium.spinosissimum.}}{a numeric vector}
    \item{\code{K.Cirsium.acaule}}{a numeric vector}
    \item{\code{K.Centaurea.nervosa.}}{a numeric vector}
    \item{\code{K.Aposeris.foetida}}{a numeric vector}
    \item{\code{K.Taraxacum.alpinum.}}{a numeric vector}
    \item{\code{K.Crepis.aurea}}{a numeric vector}
    \item{\code{K.Crepis.pyrenaica}}{a numeric vector}
    \item{\code{K.Hieracium.lachenalii}}{a numeric vector}
    \item{\code{K.Hieracium.glaucum.}}{a numeric vector}
    \item{\code{K.Hieracium.laevigatum}}{a numeric vector}
    \item{\code{K.Potamogeton.gramineus.}}{a numeric vector}
    \item{\code{K.Tofieldia.calyculata}}{a numeric vector}
    \item{\code{K.Lloydia.serotina}}{a numeric vector}
    \item{\code{K.Allium.senescens.sspmontanum}}{a numeric vector}
    \item{\code{K.Streptopus.amplexifolius}}{a numeric vector}
    \item{\code{K.Juncus.filiformis.}}{a numeric vector}
    \item{\code{K.Juncus.articulatus}}{a numeric vector}
    \item{\code{K.Luzula.campestris.}}{a numeric vector}
    \item{\code{K.Festuca.arundinacea.}}{a numeric vector}
    \item{\code{K.Festuca.quadriflora.}}{a numeric vector}
    \item{\code{K.Festuca.violacea.s.l}}{a numeric vector}
    \item{\code{K.Festuca.halleri.}}{a numeric vector}
    \item{\code{K.Festuca.pallens.}}{a numeric vector}
    \item{\code{K.Festuca.valesiaca.}}{a numeric vector}
    \item{\code{K.Poa.supina}}{a numeric vector}
    \item{\code{K.Poa.palustris.}}{a numeric vector}
    \item{\code{K.Bromus.arvensis.}}{a numeric vector}
    \item{\code{K.Agropyron.repens}}{a numeric vector}
    \item{\code{K.Avenula.versicolor}}{a numeric vector}
    \item{\code{K.Hierochloe.odorata}}{a numeric vector}
    \item{\code{K.Agrostis.alpina.}}{a numeric vector}
    \item{\code{K.Phleum.alpinum}}{a numeric vector}
    \item{\code{K.Phleum.phleoides}}{a numeric vector}
    \item{\code{K.Phleum.hirsutum.}}{a numeric vector}
    \item{\code{K.Achnatherum.calamagrostis.}}{a numeric vector}
    \item{\code{K.Molinia.caerulea}}{a numeric vector}
    \item{\code{K.Eriophorum.latifolium.}}{a numeric vector}
    \item{\code{K.Eriophorum.vaginatum}}{a numeric vector}
    \item{\code{K.Eleocharis.palustris}}{a numeric vector}
    \item{\code{K.Carex.disticha}}{a numeric vector}
    \item{\code{K.Carex.davalliana}}{a numeric vector}
    \item{\code{K.Carex.strigosa}}{a numeric vector}
    \item{\code{K.Carex.panicea.}}{a numeric vector}
    \item{\code{K.Carex.flava.}}{a numeric vector}
    \item{\code{K.Carex.humilis.}}{a numeric vector}
    \item{\code{K.Carex.caryophyllea}}{a numeric vector}
    \item{\code{K.Carex.sempervirens}}{a numeric vector}
    \item{\code{K.Carex.firma.}}{a numeric vector}
    \item{\code{K.Cypripedium.calceolus.}}{a numeric vector}
    \item{\code{K.Cephalanthera.longifolia}}{a numeric vector}
    \item{\code{K.Cephalanthera.rubra.}}{a numeric vector}
    \item{\code{K.Listera.cordata.}}{a numeric vector}
    \item{\code{K.Goodyera.repens.}}{a numeric vector}
    \item{\code{K.Orchis.morio}}{a numeric vector}
    \item{\code{K.Orchis.mascula}}{a numeric vector}
    \item{\code{K.Aesculus.hippocastanum}}{a numeric vector}
    \item{\code{K.Phyteuma.nigrum.}}{a numeric vector}
    \item{\code{S.Spiraea.chamaedryfolia}}{a numeric vector}
    \item{\code{K.Spiraea.chamaedryfolia}}{a numeric vector}
    \item{\code{M.Polytrichum.commune.}}{a numeric vector}
    \item{\code{M.Plagiothecium.undulatum.}}{a numeric vector}
    \item{\code{M.Scleropodium.purum}}{a numeric vector}
    \item{\code{M.Cratoneuron.commutatum}}{a numeric vector}
    \item{\code{K.Selaginella.helvetica.}}{a numeric vector}
    \item{\code{K.Cryptogramma.crispa.}}{a numeric vector}
    \item{\code{K.Asplenium.septentrionale}}{a numeric vector}
    \item{\code{K.Asplenium.breynii.}}{a numeric vector}
    \item{\code{K.Woodsia.alpina}}{a numeric vector}
    \item{\code{K.Polystichum.braunii.}}{a numeric vector}
    \item{\code{S.Juniperus.sabina}}{a numeric vector}
    \item{\code{K.Juniperus.sabina}}{a numeric vector}
    \item{\code{K.Salix.foetida.}}{a numeric vector}
    \item{\code{B.Ostrya.carpinifolia.}}{a numeric vector}
    \item{\code{S.Ostrya.carpinifolia.}}{a numeric vector}
    \item{\code{K.Ostrya.carpinifolia.}}{a numeric vector}
    \item{\code{B.Quercus.pubescens.}}{a numeric vector}
    \item{\code{S.Quercus.pubescens.}}{a numeric vector}
    \item{\code{K.Quercus.pubescens.}}{a numeric vector}
    \item{\code{S.Humulus.lupulus.}}{a numeric vector}
    \item{\code{K.Humulus.lupulus.}}{a numeric vector}
    \item{\code{K.Parietaria.officinalis}}{a numeric vector}
    \item{\code{B.Viscum.album}}{a numeric vector}
    \item{\code{S.Viscum.album}}{a numeric vector}
    \item{\code{K.Fallopia.dumetorum}}{a numeric vector}
    \item{\code{K.Rumex.alpinus.}}{a numeric vector}
    \item{\code{K.Chenopodium.bonus.henricus}}{a numeric vector}
    \item{\code{K.Chenopodium.album.}}{a numeric vector}
    \item{\code{K.Minuartia.laricifolia.}}{a numeric vector}
    \item{\code{K.Cerastium.arvense.sspstrictum.}}{a numeric vector}
    \item{\code{K.Cerastium.glomeratum}}{a numeric vector}
    \item{\code{K.Cerastium.semidecandrum.}}{a numeric vector}
    \item{\code{K.Sagina.procumbens.}}{a numeric vector}
    \item{\code{K.Silene.otites.}}{a numeric vector}
    \item{\code{K.Silene.exscapa}}{a numeric vector}
    \item{\code{K.Silene.alba.}}{a numeric vector}
    \item{\code{K.Gypsophila.repens.}}{a numeric vector}
    \item{\code{K.Dianthus.carthusianorum.}}{a numeric vector}
    \item{\code{K.Aconitum.paniculatum}}{a numeric vector}
    \item{\code{K.Pulsatilla.apiifolia}}{a numeric vector}
    \item{\code{K.Pulsatilla.vulgaris.}}{a numeric vector}
    \item{\code{K.Ranunculus.bulbosus.}}{a numeric vector}
    \item{\code{K.Thalictrum.alpinum}}{a numeric vector}
    \item{\code{K.Rorippa.pyrenaica.}}{a numeric vector}
    \item{\code{K.Cardamine.amara.}}{a numeric vector}
    \item{\code{K.Cardamine.resedifolia.}}{a numeric vector}
    \item{\code{K.Arabis.glabra.}}{a numeric vector}
    \item{\code{K.Arabis.pauciflora.}}{a numeric vector}
    \item{\code{K.Arabis.ciliata}}{a numeric vector}
    \item{\code{K.Draba.aizoides}}{a numeric vector}
    \item{\code{K.Draba.dubia.}}{a numeric vector}
    \item{\code{K.Sempervivum.arachnoideum}}{a numeric vector}
    \item{\code{K.Sedum.telephium.}}{a numeric vector}
    \item{\code{K.Sedum.montanum}}{a numeric vector}
    \item{\code{K.Sedum.rupestre}}{a numeric vector}
    \item{\code{K.Sedum.sexangulare.}}{a numeric vector}
    \item{\code{K.Sedum.alpestre}}{a numeric vector}
    \item{\code{K.Sedum.dasyphyllum.}}{a numeric vector}
    \item{\code{K.Sedum.annuum}}{a numeric vector}
    \item{\code{K.Saxifraga.stellaris.}}{a numeric vector}
    \item{\code{K.Saxifraga.bryoides}}{a numeric vector}
    \item{\code{K.Saxifraga.moschata}}{a numeric vector}
    \item{\code{K.Saxifraga.oppositifolia.}}{a numeric vector}
    \item{\code{K.Saxifraga.retusa}}{a numeric vector}
    \item{\code{K.Saxifraga.cotyledon.}}{a numeric vector}
    \item{\code{S.Ribes.petraeum}}{a numeric vector}
    \item{\code{K.Ribes.petraeum}}{a numeric vector}
    \item{\code{K.Filipendula.vulgaris}}{a numeric vector}
    \item{\code{K.Potentilla.rupestris}}{a numeric vector}
    \item{\code{K.Potentilla.argentea.}}{a numeric vector}
    \item{\code{K.Potentilla.neumanniana}}{a numeric vector}
    \item{\code{K.Potentilla.pusilla}}{a numeric vector}
    \item{\code{K.Sibbaldia.procumbens}}{a numeric vector}
    \item{\code{K.Alchemilla.hybrida.s.l}}{a numeric vector}
    \item{\code{S.Mespilus.germanica}}{a numeric vector}
    \item{\code{K.Mespilus.germanica}}{a numeric vector}
    \item{\code{K.Prunus.cerasus}}{a numeric vector}
    \item{\code{B.Laburnum.anagyroides}}{a numeric vector}
    \item{\code{S.Laburnum.anagyroides}}{a numeric vector}
    \item{\code{K.Laburnum.anagyroides}}{a numeric vector}
    \item{\code{B.Laburnum.alpinum}}{a numeric vector}
    \item{\code{S.Laburnum.alpinum}}{a numeric vector}
    \item{\code{K.Laburnum.alpinum}}{a numeric vector}
    \item{\code{S.Cytisus.scoparius.}}{a numeric vector}
    \item{\code{K.Cytisus.scoparius.}}{a numeric vector}
    \item{\code{K.Genista.tinctoria.}}{a numeric vector}
    \item{\code{K.Astragalus.penduliflorus}}{a numeric vector}
    \item{\code{K.Astragalus.glycyphyllos.}}{a numeric vector}
    \item{\code{K.Astragalus.monspessulanus.}}{a numeric vector}
    \item{\code{K.Vicia.sativa}}{a numeric vector}
    \item{\code{K.Lathyrus.sylvestris.}}{a numeric vector}
    \item{\code{K.Lathyrus.latifolius.}}{a numeric vector}
    \item{\code{K.Lathyrus.heterophyllus}}{a numeric vector}
    \item{\code{K.Ononis.rotundifolia.}}{a numeric vector}
    \item{\code{K.Medicago.falcata}}{a numeric vector}
    \item{\code{K.Trifolium.pratense.sspnivale}}{a numeric vector}
    \item{\code{K.Trifolium.alpestre}}{a numeric vector}
    \item{\code{K.Lotus.delortii}}{a numeric vector}
    \item{\code{K.Onobrychis.viciifolia.}}{a numeric vector}
    \item{\code{K.Geranium.sanguineum.}}{a numeric vector}
    \item{\code{K.Geranium.rivulare.}}{a numeric vector}
    \item{\code{K.Geranium.rotundifolium}}{a numeric vector}
    \item{\code{K.Geranium.columbinum.}}{a numeric vector}
    \item{\code{K.Euphorbia.seguierana}}{a numeric vector}
    \item{\code{S.Ailanthus.altissima.}}{a numeric vector}
    \item{\code{K.Ailanthus.altissima.}}{a numeric vector}
    \item{\code{K.Polygala.calcarea.}}{a numeric vector}
    \item{\code{K.Vitis.vinifera.s.l}}{a numeric vector}
    \item{\code{K.Viola.thomasiana}}{a numeric vector}
    \item{\code{K.Viola.rupestris.}}{a numeric vector}
    \item{\code{K.Viola.tricolor}}{a numeric vector}
    \item{\code{K.Fumana.procumbens.}}{a numeric vector}
    \item{\code{K.Epilobium.palustre}}{a numeric vector}
    \item{\code{K.Epilobium.alsinifolium}}{a numeric vector}
    \item{\code{S.Cornus.mas}}{a numeric vector}
    \item{\code{K.Chaerophyllum.aureum}}{a numeric vector}
    \item{\code{K.Bupleurum.stellatum.}}{a numeric vector}
    \item{\code{K.Selinum.carvifolia}}{a numeric vector}
    \item{\code{K.Ligusticum.mutellinoides}}{a numeric vector}
    \item{\code{K.Peucedanum.oreoselinum}}{a numeric vector}
    \item{\code{K.Pastinaca.sativa}}{a numeric vector}
    \item{\code{K.Monotropa.hypophegea}}{a numeric vector}
    \item{\code{K.Empetrum.hermaphroditum.}}{a numeric vector}
    \item{\code{K.Androsace.carnea}}{a numeric vector}
    \item{\code{S.Fraxinus.ornus}}{a numeric vector}
    \item{\code{K.Gentiana.cruciata.}}{a numeric vector}
    \item{\code{K.Gentiana.brachyphylla.}}{a numeric vector}
    \item{\code{K.Gentiana.schleicheri}}{a numeric vector}
    \item{\code{K.Gentiana.tenella}}{a numeric vector}
    \item{\code{K.Gentiana.ramosa.}}{a numeric vector}
    \item{\code{K.Gentiana.germanica}}{a numeric vector}
    \item{\code{K.Convolvulus.arvensis}}{a numeric vector}
    \item{\code{K.Echium.vulgare}}{a numeric vector}
    \item{\code{K.Symphytum.tuberosum.}}{a numeric vector}
    \item{\code{K.Lappula.deflexa.}}{a numeric vector}
    \item{\code{K.Lamiastrum.flavidum.}}{a numeric vector}
    \item{\code{K.Verbascum.densiflorum.}}{a numeric vector}
    \item{\code{K.Verbascum.thapsus.}}{a numeric vector}
    \item{\code{K.Veronica.fruticans}}{a numeric vector}
    \item{\code{K.Veronica.praecox}}{a numeric vector}
    \item{\code{K.Veronica.spicata}}{a numeric vector}
    \item{\code{K.Tozzia.alpina.}}{a numeric vector}
    \item{\code{K.Euphrasia.versicolor}}{a numeric vector}
    \item{\code{K.Euphrasia.nemorosa}}{a numeric vector}
    \item{\code{K.Euphrasia.alpina}}{a numeric vector}
    \item{\code{K.Pedicularis.foliosa.}}{a numeric vector}
    \item{\code{K.Pedicularis.tuberosa}}{a numeric vector}
    \item{\code{K.Pedicularis.rostrato.spicata.ssphelvet}}{a numeric vector}
    \item{\code{K.Galium.aristatum}}{a numeric vector}
    \item{\code{K.Galium.laevigatum.}}{a numeric vector}
    \item{\code{K.Galium.rubrum.}}{a numeric vector}
    \item{\code{K.Galium.spurium}}{a numeric vector}
    \item{\code{K.Galium.tricornutum}}{a numeric vector}
    \item{\code{K.Cruciata.glabra.}}{a numeric vector}
    \item{\code{K.Scabiosa.gramuntia}}{a numeric vector}
    \item{\code{K.Campanula.persicifolia}}{a numeric vector}
    \item{\code{K.Phyteuma.scheuchzeri}}{a numeric vector}
    \item{\code{K.Jasione.montana.}}{a numeric vector}
    \item{\code{K.Inula.salicina}}{a numeric vector}
    \item{\code{K.Carpesium.cernuum.}}{a numeric vector}
    \item{\code{K.Leucanthemopsis.alpina}}{a numeric vector}
    \item{\code{K.Artemisia.absinthium}}{a numeric vector}
    \item{\code{K.Artemisia.campestris}}{a numeric vector}
    \item{\code{K.Senecio.incanus.}}{a numeric vector}
    \item{\code{K.Senecio.doronicum.}}{a numeric vector}
    \item{\code{K.Saussurea.discolor}}{a numeric vector}
    \item{\code{K.Cirsium.eriophorum}}{a numeric vector}
    \item{\code{K.Cirsium.heterophyllum.}}{a numeric vector}
    \item{\code{K.Cichorium.intybus.}}{a numeric vector}
    \item{\code{K.Hypochoeris.maculata}}{a numeric vector}
    \item{\code{K.Hypochoeris.uniflora}}{a numeric vector}
    \item{\code{K.Tragopogon.pratensis}}{a numeric vector}
    \item{\code{K.Lactuca.perennis}}{a numeric vector}
    \item{\code{K.Crepis.conyzifolia}}{a numeric vector}
    \item{\code{K.Crepis.tectorum.}}{a numeric vector}
    \item{\code{K.Crepis.capillaris.}}{a numeric vector}
    \item{\code{K.Hieracium.saussureoides.}}{a numeric vector}
    \item{\code{K.Hieracium.piloselloides.}}{a numeric vector}
    \item{\code{K.Hieracium.pictum}}{a numeric vector}
    \item{\code{K.Hieracium.amplexicaule}}{a numeric vector}
    \item{\code{K.Hieracium.sabaudum}}{a numeric vector}
    \item{\code{K.Hieracium.umbellatum}}{a numeric vector}
    \item{\code{K.Anthericum.ramosum}}{a numeric vector}
    \item{\code{K.Lilium.bulbiferum.}}{a numeric vector}
    \item{\code{K.Allium.strictum.}}{a numeric vector}
    \item{\code{K.Allium.sphaerocephalon}}{a numeric vector}
    \item{\code{K.Ruscus.aculeatus}}{a numeric vector}
    \item{\code{K.Juncus.trifidus.}}{a numeric vector}
    \item{\code{K.Juncus.compressus.}}{a numeric vector}
    \item{\code{K.Luzula.lutea}}{a numeric vector}
    \item{\code{K.Festuca.varia.}}{a numeric vector}
    \item{\code{K.Festuca.tenuifolia}}{a numeric vector}
    \item{\code{K.Lolium.multiflorum}}{a numeric vector}
    \item{\code{K.Oreochloa.disticha}}{a numeric vector}
    \item{\code{K.Melica.ciliata}}{a numeric vector}
    \item{\code{K.Agropyron.intermedium.}}{a numeric vector}
    \item{\code{K.Secale.cereale}}{a numeric vector}
    \item{\code{K.Koeleria.vallesiana.}}{a numeric vector}
    \item{\code{K.Trisetum.distichophyllum}}{a numeric vector}
    \item{\code{K.Agrostis.schraderiana.}}{a numeric vector}
    \item{\code{K.Stipa.joannis.}}{a numeric vector}
    \item{\code{K.Digitaria.sanguinalis.}}{a numeric vector}
    \item{\code{K.Carex.dioica}}{a numeric vector}
    \item{\code{K.Carex.rostrata}}{a numeric vector}
    \item{\code{K.Carex.halleriana}}{a numeric vector}
    \item{\code{K.Carex.umbrosa.}}{a numeric vector}
    \item{\code{K.Carex.liparocarpos}}{a numeric vector}
    \item{\code{K.Epipogium.aphyllum}}{a numeric vector}
    \item{\code{K.Nigritella.nigra}}{a numeric vector}
    \item{\code{K.Aceras.anthropophorum.}}{a numeric vector}
    \item{\code{K.Ophris.insectifera}}{a numeric vector}
    \item{\code{K.Trachycarpus.fortunei.}}{a numeric vector}
    \item{\code{K.Allium.spec.}}{a numeric vector}
    \item{\code{B.Populus.canadensis}}{a numeric vector}
    \item{\code{K.Rumex.aquaticus.}}{a numeric vector}
    \item{\code{K.Impatiens.glandulifera}}{a numeric vector}
    \item{\code{K.Viola.riviniana.}}{a numeric vector}
    \item{\code{K.Lysimachia.nummularia.}}{a numeric vector}
    \item{\code{K.Mentha.aquatica.}}{a numeric vector}
    \item{\code{K.Orobanche.alba}}{a numeric vector}
    \item{\code{K.Orobanche.caryophyllacea}}{a numeric vector}
    \item{\code{K.Achillea.nobilis}}{a numeric vector}
    \item{\code{K.Calycocorsus.stipitatus.}}{a numeric vector}
    \item{\code{K.Phragmites.australis}}{a numeric vector}
    \item{\code{K.Carex.distans.}}{a numeric vector}
    \item{\code{K.Carex.pulicaris.}}{a numeric vector}
    \item{\code{K.Orchis.pallens}}{a numeric vector}
    \item{\code{S.Cornus.stolonifera}}{a numeric vector}
    \item{\code{K.Myosotis.decumbens}}{a numeric vector}
    \item{\code{F.Bartflechte..Usnea.spec...}}{a numeric vector}
    \item{\code{F.Blasenflechte..Hypogymnia.physodes..}}{a numeric vector}
    \item{\code{F.Krustenflechten.}}{a numeric vector}
    \item{\code{F.Pseudevernia.furfuracea.}}{a numeric vector}
    \item{\code{F.Schuesselflechten..Parmelia.}}{a numeric vector}
    \item{\code{K.Lycopodiella.inundata.}}{a numeric vector}
    \item{\code{K.Sisymbrium.strictissimum}}{a numeric vector}
    \item{\code{K.Astragalus.australis}}{a numeric vector}
    \item{\code{K.Trifolium.thalii}}{a numeric vector}
    \item{\code{K.Coronilla.varia.}}{a numeric vector}
    \item{\code{K.Geranium.molle}}{a numeric vector}
    \item{\code{K.Rhamnus.saxatilis.}}{a numeric vector}
    \item{\code{K.Gentiana.nivalis}}{a numeric vector}
    \item{\code{K.Swertia.perennis}}{a numeric vector}
    \item{\code{K.Leucanthemum.adustum}}{a numeric vector}
    \item{\code{K.Doronicum.grandiflorum}}{a numeric vector}
    \item{\code{K.Senecio.erucifolius.}}{a numeric vector}
    \item{\code{K.Arctium.minus.}}{a numeric vector}
    \item{\code{K.Leontodon.incanus.}}{a numeric vector}
    \item{\code{K.Hieracium.villosum}}{a numeric vector}
    \item{\code{K.Juncus.alpino.articulatus.}}{a numeric vector}
    \item{\code{K.Poa.laxa}}{a numeric vector}
    \item{\code{K.Poa.glauca}}{a numeric vector}
    \item{\code{K.Triticum..kultivierte.Sippen}}{a numeric vector}
    \item{\code{K.Dactylorhiza.incarnata}}{a numeric vector}
    \item{\code{K.Ranunculus.auficom..cassub}}{a numeric vector}
    \item{\code{K.Viola.spec}}{a numeric vector}
    \item{\code{F.Lungenflechte..Lobaria.pulmonaria.}}{a numeric vector}
    \item{\code{F.Moosbart..Bryoria.spec..}}{a numeric vector}
    \item{\code{K.Asplenium.onopteris.}}{a numeric vector}
    \item{\code{S.Salix.helvetica.}}{a numeric vector}
    \item{\code{K.Arenaria.serpyllifolia}}{a numeric vector}
    \item{\code{K.Cerastium.carinthiacum.sspaustroalpinu}}{a numeric vector}
    \item{\code{K.Sagina.saginoides.}}{a numeric vector}
    \item{\code{K.Erysimum.helveticum.}}{a numeric vector}
    \item{\code{K.Erysimum.virgatum.}}{a numeric vector}
    \item{\code{K.Arabis.nova.}}{a numeric vector}
    \item{\code{K.Agrimonia.eupatoria.}}{a numeric vector}
    \item{\code{K.Cytisus.hirsutus}}{a numeric vector}
    \item{\code{K.Lathyrus.laevigatus.sspoccidentalis.}}{a numeric vector}
    \item{\code{K.Cornus.mas}}{a numeric vector}
    \item{\code{B.Fraxinus.ornus}}{a numeric vector}
    \item{\code{K.Fraxinus.ornus}}{a numeric vector}
    \item{\code{K.Lappula.squarrosa.}}{a numeric vector}
    \item{\code{K.Veronica.verna}}{a numeric vector}
    \item{\code{K.Pedicularis.rostrato.capitata.}}{a numeric vector}
    \item{\code{K.Lonicera.japonica.}}{a numeric vector}
    \item{\code{K.Knautia.godetii.}}{a numeric vector}
    \item{\code{K.Hieracium.intybaceum}}{a numeric vector}
    \item{\code{K.Asparagus.tenuifolius.}}{a numeric vector}
    \item{\code{K.Festuca.amethystina.}}{a numeric vector}
    \item{\code{S.Trachycarpus.fortunei.}}{a numeric vector}
    \item{\code{K.Primula.pedemontana.}}{a numeric vector}
    \item{\code{K.Potentilla.spec.}}{a numeric vector}
    \item{\code{K.Salsola.kali}}{a numeric vector}
    \item{\code{K.Chrysosplenium.oppositifolium.}}{a numeric vector}
    \item{\code{K.Hypericum.tetrapterum.}}{a numeric vector}
    \item{\code{K.Primula.veris.}}{a numeric vector}
    \item{\code{K.Symphytum.bulbosum}}{a numeric vector}
    \item{\code{K.Galium.glaucum}}{a numeric vector}
    \item{\code{K.Campanula.cervicaria}}{a numeric vector}
    \item{\code{K.Senecio.halleri.}}{a numeric vector}
    \item{\code{K.Orchis.militaris}}{a numeric vector}
    \item{\code{K.Aphanes.microcarpa}}{a numeric vector}
    \item{\code{K.Vicia.ervilia.}}{a numeric vector}
    \item{\code{K.Coronilla.vaginalis.}}{a numeric vector}
    \item{\code{K.Pyrola.chlorantha.}}{a numeric vector}
    \item{\code{K.Primula.integrifolia}}{a numeric vector}
    \item{\code{K.Cuscuta.australis.}}{a numeric vector}
    \item{\code{K.Rhaponticum.scariosum.}}{a numeric vector}
    \item{\code{K.Carex.brunnescens.}}{a numeric vector}
    \item{\code{K.Lavandula.angustifolia}}{a numeric vector}
    \item{\code{K.Asplenium.billotii}}{a numeric vector}
    \item{\code{K.Salix.helvetica.}}{a numeric vector}
    \item{\code{K.Scleranthus.annuus}}{a numeric vector}
    \item{\code{K.Pulsatilla.montana}}{a numeric vector}
    \item{\code{K.Alyssoides.utriculatum}}{a numeric vector}
    \item{\code{S.Rosa.obtusifolia}}{a numeric vector}
    \item{\code{K.Alchemilla.fissa.s.l}}{a numeric vector}
    \item{\code{K.Polygala.pedemontana}}{a numeric vector}
    \item{\code{K.Primula.vulgaris}}{a numeric vector}
    \item{\code{K.Buglossoides.purpurocaerulea}}{a numeric vector}
    \item{\code{S.Solanum.dulcamara.}}{a numeric vector}
    \item{\code{K.Linaria.angustissima}}{a numeric vector}
    \item{\code{K.Globularia.punctata.}}{a numeric vector}
    \item{\code{K.Lactuca.serriola}}{a numeric vector}
    \item{\code{K.Hieracium.alpinum.}}{a numeric vector}
    \item{\code{K.Luzula.forsteri.}}{a numeric vector}
    \item{\code{K.Poa.violacea}}{a numeric vector}
    \item{\code{K.Bromus.condensatus}}{a numeric vector}
    \item{\code{K.Carex.buxbaumii.}}{a numeric vector}
    \item{\code{K.Campanula.bertolae}}{a numeric vector}
    \item{\code{K.Carlina.stricta.}}{a numeric vector}
    \item{\code{K.Dianthus.spec.}}{a numeric vector}
  }
}
\details{
See object wssit for corresponding site information
}
\source{
Wohlgemuth, T., Moser B., Braendli, U.-B., Kull, P. & Schuetz, M. 2008. 
Diversity of forest plant species at the community and landscape scales in Switzerland. 
Plant Biosystems 142: 604-613.
}
\references{
Wildi, O. 2013. Data Analysis in Vegetation Ecology. 2nd ed. Wiley-Blackwell,
Chichester.
}
\examples{
summary(ws500)
}
\keyword{datasets}
