\name{EKv}
\alias{EKv}
\docType{data}
\title{
Swiss forest vegetation data base 1972, vegetation information
}
\description{
Swiss forest vegetation data base 1972, vegetation  information. Site data is in data frame EKs.
}
\usage{data(EKv)}
\format{
  A data frame with 2533 observations on the following 1259 variables.
  \describe{
    \item{\code{Convallaria_majalis_}}{a numeric vector}
    \item{\code{S.Cornus_sanguinea}}{a numeric vector}
    \item{\code{S.Coronilla_emerus}}{a numeric vector}
    \item{\code{Coronilla_emerus}}{a numeric vector}
    \item{\code{S.Corylus_avellana}}{a numeric vector}
    \item{\code{Dactylis_glomerata}}{a numeric vector}
    \item{\code{Digitalis_grandiflora_}}{a numeric vector}
    \item{\code{Epipactis_latifolia_}}{a numeric vector}
    \item{\code{B.Fagus_silvatica_}}{a numeric vector}
    \item{\code{Fragaria_vesca}}{a numeric vector}
    \item{\code{Geranium_robertianum}}{a numeric vector}
    \item{\code{Hepatica_triloba}}{a numeric vector}
    \item{\code{Hypericum_montanum}}{a numeric vector}
    \item{\code{Knautia_arvensis}}{a numeric vector}
    \item{\code{Lathyrus_pratensis}}{a numeric vector}
    \item{\code{B.Acer_platanoides}}{a numeric vector}
    \item{\code{S.Acer_pseudoplatanus_}}{a numeric vector}
    \item{\code{S.Lonicera_xylosteum}}{a numeric vector}
    \item{\code{Mercurialis_perennis}}{a numeric vector}
    \item{\code{Moehringia_muscosa}}{a numeric vector}
    \item{\code{Molinia_litoralis_}}{a numeric vector}
    \item{\code{Mycelis_muralis_}}{a numeric vector}
    \item{\code{Origanum_vulgare}}{a numeric vector}
    \item{\code{S.Picea_excelsa_}}{a numeric vector}
    \item{\code{S.Pirus_malus_}}{a numeric vector}
    \item{\code{Platanthera_bifolia_}}{a numeric vector}
    \item{\code{Poa_nemoralis_}}{a numeric vector}
    \item{\code{Polygala_chamaebuxus}}{a numeric vector}
    \item{\code{Polypodium_vulgare}}{a numeric vector}
    \item{\code{B.Prunus_avium}}{a numeric vector}
    \item{\code{Agrostis_tenuis_}}{a numeric vector}
    \item{\code{S.Rhamnus_cathartica}}{a numeric vector}
    \item{\code{Satureja_vulgaris_}}{a numeric vector}
    \item{\code{Seseli_libanotis}}{a numeric vector}
    \item{\code{Sesleria_coerulea_}}{a numeric vector}
    \item{\code{Silene_nutans_}}{a numeric vector}
    \item{\code{Silene_vulgaris_}}{a numeric vector}
    \item{\code{Solidago_virga.aurea}}{a numeric vector}
    \item{\code{B.Sorbus_aria_}}{a numeric vector}
    \item{\code{S.Sorbus_aria_}}{a numeric vector}
    \item{\code{S.Sorbus_aucuparia}}{a numeric vector}
    \item{\code{B.Tilia_cordata_}}{a numeric vector}
    \item{\code{S.Ulmus_scabra}}{a numeric vector}
    \item{\code{Vaccinium_myrtillus_}}{a numeric vector}
    \item{\code{Valeriana_officinalis_}}{a numeric vector}
    \item{\code{Valeriana_tripteris_}}{a numeric vector}
    \item{\code{Veronica_latifolia}}{a numeric vector}
    \item{\code{Vincetoxicum_officinale_}}{a numeric vector}
    \item{\code{M.Camptothecium_lutescens_}}{a numeric vector}
    \item{\code{M.Pterogonium_ornipodioides_}}{a numeric vector}
    \item{\code{M.Rhytidium_rugosum_}}{a numeric vector}
    \item{\code{Arabis_hirsuta}}{a numeric vector}
    \item{\code{Asplenium_trichomanes_}}{a numeric vector}
    \item{\code{Brachypodium_pinnatum_}}{a numeric vector}
    \item{\code{Brachypodium_silvaticum_}}{a numeric vector}
    \item{\code{S.Quercus_spec._}}{a numeric vector}
    \item{\code{Rubus_fruticosus}}{a numeric vector}
    \item{\code{M.Tortella_spec.}}{a numeric vector}
    \item{\code{M.Thuidium_spec.}}{a numeric vector}
    \item{\code{M.Polytrichum_spec._}}{a numeric vector}
    \item{\code{M.Dicranum_spec.}}{a numeric vector}
    \item{\code{Calamagrostis_varia_}}{a numeric vector}
    \item{\code{M.Ctenidium_spec._}}{a numeric vector}
    \item{\code{Campanula_glomerata_}}{a numeric vector}
    \item{\code{M.Hypnum_spec.}}{a numeric vector}
    \item{\code{Campanula_rotundifolia}}{a numeric vector}
    \item{\code{S.Rosa_canina._arvensis._spec.}}{a numeric vector}
    \item{\code{Campanula_trachelium}}{a numeric vector}
    \item{\code{Carex_digitata}}{a numeric vector}
    \item{\code{Galium_mollugo_.ag._}}{a numeric vector}
    \item{\code{Hieracium_murorum_.ag.}}{a numeric vector}
    \item{\code{M.Homalothecium_spec._}}{a numeric vector}
    \item{\code{M.Cirriphyllum_spec.}}{a numeric vector}
    \item{\code{S.Fagus_silvatica_}}{a numeric vector}
    \item{\code{Fagus_silvatica_}}{a numeric vector}
    \item{\code{Festuca_heterophylla}}{a numeric vector}
    \item{\code{Festuca_rubra_}}{a numeric vector}
    \item{\code{Fraxinus_excelsior}}{a numeric vector}
    \item{\code{Galium_odoratum_}}{a numeric vector}
    \item{\code{Hedera_helix}}{a numeric vector}
    \item{\code{Acer_pseudoplatanus_}}{a numeric vector}
    \item{\code{Luzula_nemorosa_}}{a numeric vector}
    \item{\code{Luzula_pilosa_}}{a numeric vector}
    \item{\code{Picea_excelsa_}}{a numeric vector}
    \item{\code{Pinus_silvestris}}{a numeric vector}
    \item{\code{Polygonatum_multiflorum_}}{a numeric vector}
    \item{\code{Ajuga_reptans_}}{a numeric vector}
    \item{\code{Viola_silvestris}}{a numeric vector}
    \item{\code{M.Atrichum_undulatum}}{a numeric vector}
    \item{\code{M.Pleurozium_schreberi}}{a numeric vector}
    \item{\code{Anemone_nemorosa}}{a numeric vector}
    \item{\code{M.Rhytidiadelphus_triqueter_}}{a numeric vector}
    \item{\code{B.Quercus_spec._}}{a numeric vector}
    \item{\code{Quercus_spec._}}{a numeric vector}
    \item{\code{M.Fissidens_spec._}}{a numeric vector}
    \item{\code{M.Hylocomium_spec.}}{a numeric vector}
    \item{\code{M.Eurhynchium_spec._}}{a numeric vector}
    \item{\code{Carex_montana_}}{a numeric vector}
    \item{\code{Carex_silvatica_}}{a numeric vector}
    \item{\code{Carpinus_betulus}}{a numeric vector}
    \item{\code{M.Diplophyllum_spec.}}{a numeric vector}
    \item{\code{Dryopteris_filix.mas}}{a numeric vector}
    \item{\code{S.Abies_alba}}{a numeric vector}
    \item{\code{Pinus_strobus_}}{a numeric vector}
    \item{\code{Pteridium_aquilinum_}}{a numeric vector}
    \item{\code{M.Blepharostoma_trichophyllum_}}{a numeric vector}
    \item{\code{Dryopteris_dilatata._spinulosa}}{a numeric vector}
    \item{\code{Athyrium_filix.femina_}}{a numeric vector}
    \item{\code{M.Dicranella_spec.}}{a numeric vector}
    \item{\code{M.Isopterygium_spec.}}{a numeric vector}
    \item{\code{S.Ilex_aquifolium_}}{a numeric vector}
    \item{\code{Abies_alba}}{a numeric vector}
    \item{\code{Prenanthes_purpurea_}}{a numeric vector}
    \item{\code{M.Diphyscium_foliosum_}}{a numeric vector}
    \item{\code{Lathyrus_montanus_}}{a numeric vector}
    \item{\code{Luzula_silvatica}}{a numeric vector}
    \item{\code{B.Pinus_silvestris}}{a numeric vector}
    \item{\code{M.Brachythecium_spec._}}{a numeric vector}
    \item{\code{M.Calypogeia_spec.}}{a numeric vector}
    \item{\code{M.Isothecium_spec.}}{a numeric vector}
    \item{\code{B.Picea_excelsa_}}{a numeric vector}
    \item{\code{M.Mnium_spec._}}{a numeric vector}
    \item{\code{Betula_spec.}}{a numeric vector}
    \item{\code{M.Plagiochila_spec._}}{a numeric vector}
    \item{\code{M.Cladonia_spec.}}{a numeric vector}
    \item{\code{Acer_campestre}}{a numeric vector}
    \item{\code{Veronica_officinalis}}{a numeric vector}
    \item{\code{Vicia_sepium}}{a numeric vector}
    \item{\code{Melampyrum_pratense_}}{a numeric vector}
    \item{\code{Phyteuma_spicatum_}}{a numeric vector}
    \item{\code{Prunus_avium}}{a numeric vector}
    \item{\code{M.Plagiothecium_spec._}}{a numeric vector}
    \item{\code{B.Pseudotsuga_taxifolia_}}{a numeric vector}
    \item{\code{B.Abies_alba}}{a numeric vector}
    \item{\code{Deschampsia_flexuosa}}{a numeric vector}
    \item{\code{B.Betula_spec.}}{a numeric vector}
    \item{\code{Teucrium_scorodonia_}}{a numeric vector}
    \item{\code{M.Peltigera_rufescens_}}{a numeric vector}
    \item{\code{Epilobium_montanum}}{a numeric vector}
    \item{\code{M.Lepidozia_reptans_}}{a numeric vector}
    \item{\code{M.Leucobryum_glaucum}}{a numeric vector}
    \item{\code{Hieracium_umbellatum_.ag._}}{a numeric vector}
    \item{\code{M.Pseudoscleropodium_purum}}{a numeric vector}
    \item{\code{M.Ceratodon_purpureus_}}{a numeric vector}
    \item{\code{M.Tetraphis_pellucida_}}{a numeric vector}
    \item{\code{Calluna_vulgaris}}{a numeric vector}
    \item{\code{S.Frangula_alnus}}{a numeric vector}
    \item{\code{Molinia_coerulea}}{a numeric vector}
    \item{\code{B.Pinus_strobus_}}{a numeric vector}
    \item{\code{Carex_pilulifera}}{a numeric vector}
    \item{\code{S.Juniperus_communis}}{a numeric vector}
    \item{\code{Epipactis_atropurpurea}}{a numeric vector}
    \item{\code{Galium_silvaticum_}}{a numeric vector}
    \item{\code{Knautia_silvatica_}}{a numeric vector}
    \item{\code{B.Acer_pseudoplatanus_}}{a numeric vector}
    \item{\code{Melica_nutans_}}{a numeric vector}
    \item{\code{Neottia_nidus.avis}}{a numeric vector}
    \item{\code{Oxalis_acetosella_}}{a numeric vector}
    \item{\code{S.Viburnum_lantana}}{a numeric vector}
    \item{\code{Bromus_benekenii}}{a numeric vector}
    \item{\code{S.Rubus_fruticosus}}{a numeric vector}
    \item{\code{Carex_flacca}}{a numeric vector}
    \item{\code{Cephalanthera_damasonium}}{a numeric vector}
    \item{\code{Cephalanthera_rubra_}}{a numeric vector}
    \item{\code{S.Fraxinus_excelsior}}{a numeric vector}
    \item{\code{S.Hedera_helix}}{a numeric vector}
    \item{\code{S.Acer_campestre}}{a numeric vector}
    \item{\code{S.Prunus_avium}}{a numeric vector}
    \item{\code{Angelica_silvestris_}}{a numeric vector}
    \item{\code{S.Crataegus_spec._}}{a numeric vector}
    \item{\code{Majanthemum_bifolium}}{a numeric vector}
    \item{\code{S.Rubus_idaeus}}{a numeric vector}
    \item{\code{S.Sambucus_racemosa_}}{a numeric vector}
    \item{\code{Elymus_europaeus}}{a numeric vector}
    \item{\code{Festuca_altissima_}}{a numeric vector}
    \item{\code{Heracleum_sphondylium_}}{a numeric vector}
    \item{\code{Lathyrus_vernus_}}{a numeric vector}
    \item{\code{S.Lonicera_alpigena_}}{a numeric vector}
    \item{\code{S.Lonicera_nigra}}{a numeric vector}
    \item{\code{Actaea_spicata}}{a numeric vector}
    \item{\code{Adenostyles_alliariae_}}{a numeric vector}
    \item{\code{Adenostyles_glabra}}{a numeric vector}
    \item{\code{Paris_quadrifolia_}}{a numeric vector}
    \item{\code{Polygonatum_verticillatum_}}{a numeric vector}
    \item{\code{Primula_elatior_}}{a numeric vector}
    \item{\code{Pyrola_secunda}}{a numeric vector}
    \item{\code{S.Rosa_pendulina}}{a numeric vector}
    \item{\code{Saxifraga_rotundifolia}}{a numeric vector}
    \item{\code{Senecio_fuchsii_}}{a numeric vector}
    \item{\code{B.Sorbus_aucuparia}}{a numeric vector}
    \item{\code{Sorbus_aucuparia}}{a numeric vector}
    \item{\code{M.Rhytidiadelphus_loreus}}{a numeric vector}
    \item{\code{Arabis_alpina_}}{a numeric vector}
    \item{\code{Aruncus_silvester_}}{a numeric vector}
    \item{\code{Lamium_galeobdolon_.ag._}}{a numeric vector}
    \item{\code{S.Ribes_spec._}}{a numeric vector}
    \item{\code{Cardamine_pentaphyllos}}{a numeric vector}
    \item{\code{Melampyrum_silvaticum_}}{a numeric vector}
    \item{\code{Milium_effusum}}{a numeric vector}
    \item{\code{Polystichum_lobatum_}}{a numeric vector}
    \item{\code{Polystichum_lonchitis_}}{a numeric vector}
    \item{\code{Ranunculus_aconitifolius_.ag._}}{a numeric vector}
    \item{\code{Asplenium_viride}}{a numeric vector}
    \item{\code{Lastrea_dryopteris}}{a numeric vector}
    \item{\code{Cardamine_heptaphylla_}}{a numeric vector}
    \item{\code{Corylus_avellana}}{a numeric vector}
    \item{\code{Euphorbia_dulcis}}{a numeric vector}
    \item{\code{S.Ligustrum_vulgare_}}{a numeric vector}
    \item{\code{Lonicera_nigra}}{a numeric vector}
    \item{\code{Lonicera_xylosteum}}{a numeric vector}
    \item{\code{S.Prunus_spinosa}}{a numeric vector}
    \item{\code{Rosa_pendulina}}{a numeric vector}
    \item{\code{Rubus_idaeus}}{a numeric vector}
    \item{\code{Veronica_chamaedrys_}}{a numeric vector}
    \item{\code{Rosa_canina._arvensis._spec.}}{a numeric vector}
    \item{\code{Galeopsis_tetrahit}}{a numeric vector}
    \item{\code{Galium_rotundifolium}}{a numeric vector}
    \item{\code{Scrophularia_nodosa_}}{a numeric vector}
    \item{\code{S.Tilia_cordata_}}{a numeric vector}
    \item{\code{M.Gyroweisia_tenuis_}}{a numeric vector}
    \item{\code{M.Scapania_spec.}}{a numeric vector}
    \item{\code{M.Bartramia_spec._}}{a numeric vector}
    \item{\code{M.Lophocolea_spec.}}{a numeric vector}
    \item{\code{Carex_pilosa}}{a numeric vector}
    \item{\code{Sambucus_nigra}}{a numeric vector}
    \item{\code{B.Carpinus_betulus}}{a numeric vector}
    \item{\code{S.Carpinus_betulus}}{a numeric vector}
    \item{\code{S.Viburnum_opulus_}}{a numeric vector}
    \item{\code{Monotropa_spec._}}{a numeric vector}
    \item{\code{B.Larix_decidua_}}{a numeric vector}
    \item{\code{Lastrea_phegopteris_}}{a numeric vector}
    \item{\code{Moehringia_trinervia}}{a numeric vector}
    \item{\code{Lycopodium_selago_}}{a numeric vector}
    \item{\code{M.Neckera_spec._}}{a numeric vector}
    \item{\code{Luzula_nivea}}{a numeric vector}
    \item{\code{Viburnum_lantana}}{a numeric vector}
    \item{\code{Deschampsia_caespitosa}}{a numeric vector}
    \item{\code{Lastrea_oreopteris}}{a numeric vector}
    \item{\code{Lysimachia_nemorum}}{a numeric vector}
    \item{\code{Petasites_albus_}}{a numeric vector}
    \item{\code{Blechnum_spicant}}{a numeric vector}
    \item{\code{Sanicula_europaea_}}{a numeric vector}
    \item{\code{Ranunculus_nemorosus_.ag._}}{a numeric vector}
    \item{\code{Homogyne_alpina_}}{a numeric vector}
    \item{\code{Hieracium_sabaudum}}{a numeric vector}
    \item{\code{Luzula_forsteri_}}{a numeric vector}
    \item{\code{S.Sorbus_torminalis_}}{a numeric vector}
    \item{\code{S.Castanea_sativa_}}{a numeric vector}
    \item{\code{Euphorbia_amygdaloides}}{a numeric vector}
    \item{\code{Melica_uniflora_}}{a numeric vector}
    \item{\code{Melittis_melissophyllum_}}{a numeric vector}
    \item{\code{Potentilla_sterilis_}}{a numeric vector}
    \item{\code{Primula_vulgaris}}{a numeric vector}
    \item{\code{Anthoxanthum_spec.}}{a numeric vector}
    \item{\code{Pyrola_rotundifolia_}}{a numeric vector}
    \item{\code{Orchis_maculata_}}{a numeric vector}
    \item{\code{Vaccinium_vitis.idaea_}}{a numeric vector}
    \item{\code{Cephalanthera_longifolia}}{a numeric vector}
    \item{\code{S.Lonicera_periclymenum_}}{a numeric vector}
    \item{\code{Impatiens_noli.tangere}}{a numeric vector}
    \item{\code{M.Barbula_spec._}}{a numeric vector}
    \item{\code{Luzula_campestris_.ag.}}{a numeric vector}
    \item{\code{M.Chiloscyphus_spec.}}{a numeric vector}
    \item{\code{M.Metzgeria_spec._}}{a numeric vector}
    \item{\code{Festuca_gigantea}}{a numeric vector}
    \item{\code{Poa_trivialis_}}{a numeric vector}
    \item{\code{Sambucus_racemosa_}}{a numeric vector}
    \item{\code{Viburnum_opulus_}}{a numeric vector}
    \item{\code{Carex_brizoides_}}{a numeric vector}
    \item{\code{Cornus_sanguinea}}{a numeric vector}
    \item{\code{S.Daphne_mezereum_}}{a numeric vector}
    \item{\code{Daphne_mezereum_}}{a numeric vector}
    \item{\code{Evonymus_europaea_}}{a numeric vector}
    \item{\code{B.Fraxinus_excelsior}}{a numeric vector}
    \item{\code{B.Acer_campestre}}{a numeric vector}
    \item{\code{Quercus_rubra_}}{a numeric vector}
    \item{\code{B.Sorbus_torminalis_}}{a numeric vector}
    \item{\code{Sorbus_torminalis_}}{a numeric vector}
    \item{\code{Tamus_communis}}{a numeric vector}
    \item{\code{Pulmonaria_spec.}}{a numeric vector}
    \item{\code{Carex_umbrosa_}}{a numeric vector}
    \item{\code{S.Juglans_regia_}}{a numeric vector}
    \item{\code{Juglans_regia_}}{a numeric vector}
    \item{\code{S.Acer_platanoides}}{a numeric vector}
    \item{\code{B.Robinia_pseudo.acacia_}}{a numeric vector}
    \item{\code{Crataegus_spec._}}{a numeric vector}
    \item{\code{Ligustrum_vulgare_}}{a numeric vector}
    \item{\code{Ilex_aquifolium_}}{a numeric vector}
    \item{\code{Rubus_saxatilis_}}{a numeric vector}
    \item{\code{Valeriana_montana_}}{a numeric vector}
    \item{\code{M.Ptilium_crista.castrensis_}}{a numeric vector}
    \item{\code{Centaurea_montana_}}{a numeric vector}
    \item{\code{Acer_platanoides}}{a numeric vector}
    \item{\code{B.Prunus_spinosa}}{a numeric vector}
    \item{\code{Carex_pallescens}}{a numeric vector}
    \item{\code{Salix_caprea}}{a numeric vector}
    \item{\code{B.Picea_sitchensis}}{a numeric vector}
    \item{\code{Tilia_cordata_}}{a numeric vector}
    \item{\code{S.Taxus_baccata_}}{a numeric vector}
    \item{\code{Taxus_baccata_}}{a numeric vector}
    \item{\code{B.Quercus_rubra_}}{a numeric vector}
    \item{\code{Hypericum_pulchrum}}{a numeric vector}
    \item{\code{Frangula_alnus}}{a numeric vector}
    \item{\code{Prunus_spinosa}}{a numeric vector}
    \item{\code{Pseudotsuga_taxifolia_}}{a numeric vector}
    \item{\code{Geum_urbanum}}{a numeric vector}
    \item{\code{B.Hedera_helix}}{a numeric vector}
    \item{\code{Sorbus_aria_}}{a numeric vector}
    \item{\code{B.Taxus_baccata_}}{a numeric vector}
    \item{\code{Calamagrostis_arundinacea_}}{a numeric vector}
    \item{\code{Luzula_flavescens_}}{a numeric vector}
    \item{\code{Rubus_caesius_}}{a numeric vector}
    \item{\code{Taraxacum_spec._}}{a numeric vector}
    \item{\code{Gentiana_asclepiadea}}{a numeric vector}
    \item{\code{Aegopodium_podagraria_}}{a numeric vector}
    \item{\code{M.Dolichotheca_seligeri_}}{a numeric vector}
    \item{\code{M.Porella_spec._}}{a numeric vector}
    \item{\code{Euphorbia_cyparissias_}}{a numeric vector}
    \item{\code{Hypericum_perforatum}}{a numeric vector}
    \item{\code{Impatiens_parviflora}}{a numeric vector}
    \item{\code{M.Anomodon_spec.}}{a numeric vector}
    \item{\code{M.Bryum_spec._}}{a numeric vector}
    \item{\code{Galeopsis_pubescens_}}{a numeric vector}
    \item{\code{Phyteuma_betonicifolium_}}{a numeric vector}
    \item{\code{Salvia_glutinosa}}{a numeric vector}
    \item{\code{M.Heterocladium_spec._}}{a numeric vector}
    \item{\code{Cytisus_scoparius_}}{a numeric vector}
    \item{\code{Silene_rupestris}}{a numeric vector}
    \item{\code{S.Cytisus_scoparius_}}{a numeric vector}
    \item{\code{Cruciata_glabra_}}{a numeric vector}
    \item{\code{Festuca_ovina_}}{a numeric vector}
    \item{\code{Nardus_stricta}}{a numeric vector}
    \item{\code{S.Populus_tremula_}}{a numeric vector}
    \item{\code{Saxifraga_cuneifolia}}{a numeric vector}
    \item{\code{Sieglingia_decumbens}}{a numeric vector}
    \item{\code{S.Alnus_glutinosa_}}{a numeric vector}
    \item{\code{S.Betula_spec.}}{a numeric vector}
    \item{\code{Carex_fritschii_}}{a numeric vector}
    \item{\code{B.Castanea_sativa_}}{a numeric vector}
    \item{\code{Crocus_albiflorus_}}{a numeric vector}
    \item{\code{S.Evonymus_europaea_}}{a numeric vector}
    \item{\code{S.Mespilus_germanica}}{a numeric vector}
    \item{\code{B.Populus_tremula_}}{a numeric vector}
    \item{\code{Populus_tremula_}}{a numeric vector}
    \item{\code{S.Robinia_pseudo.acacia_}}{a numeric vector}
    \item{\code{S.Sambucus_nigra}}{a numeric vector}
    \item{\code{B.Alnus_glutinosa_}}{a numeric vector}
    \item{\code{Betonica_officinalis}}{a numeric vector}
    \item{\code{Peucedanum_oreoselinum}}{a numeric vector}
    \item{\code{Sedum_maximum_}}{a numeric vector}
    \item{\code{Vinca_minor_}}{a numeric vector}
    \item{\code{Anthericum_liliago}}{a numeric vector}
    \item{\code{Carex_humilis_}}{a numeric vector}
    \item{\code{Castanea_sativa_}}{a numeric vector}
    \item{\code{S.Rhododendron_ferrugineum}}{a numeric vector}
    \item{\code{S.Pinus_strobus_}}{a numeric vector}
    \item{\code{Robinia_pseudo.acacia_}}{a numeric vector}
    \item{\code{Asplenium_adiantum.nigrum_}}{a numeric vector}
    \item{\code{Verbascum_spec._}}{a numeric vector}
    \item{\code{Galium_verum}}{a numeric vector}
    \item{\code{Genista_tinctoria_}}{a numeric vector}
    \item{\code{S.Larix_decidua_}}{a numeric vector}
    \item{\code{Pimpinella_major}}{a numeric vector}
    \item{\code{Poa_pratensis_}}{a numeric vector}
    \item{\code{Potentilla_erecta_}}{a numeric vector}
    \item{\code{Rumex_acetosa_}}{a numeric vector}
    \item{\code{Serratula_tinctoria_}}{a numeric vector}
    \item{\code{Holcus_mollis_}}{a numeric vector}
    \item{\code{Hoplismenus_undulatifolius}}{a numeric vector}
    \item{\code{S.Tamus_communis}}{a numeric vector}
    \item{\code{Viola_hirta_}}{a numeric vector}
    \item{\code{Chrysanthemum_leucanthemum_.ag._}}{a numeric vector}
    \item{\code{Festuca_varia_}}{a numeric vector}
    \item{\code{Silene_dioeca_}}{a numeric vector}
    \item{\code{B.Corylus_avellana}}{a numeric vector}
    \item{\code{S.Cytisus_nigricans_}}{a numeric vector}
    \item{\code{Cytisus_nigricans_}}{a numeric vector}
    \item{\code{Paradisia_liliastrum}}{a numeric vector}
    \item{\code{Pedicularis_tuberosa}}{a numeric vector}
    \item{\code{Phyteuma_scheuchzeri}}{a numeric vector}
    \item{\code{Erica_carnea}}{a numeric vector}
    \item{\code{Rhododendron_ferrugineum}}{a numeric vector}
    \item{\code{S.Amelanchier_ovalis}}{a numeric vector}
    \item{\code{Genista_pilosa}}{a numeric vector}
    \item{\code{Genista_germanica_}}{a numeric vector}
    \item{\code{Arnica_montana}}{a numeric vector}
    \item{\code{Chaerophyllum_aureum}}{a numeric vector}
    \item{\code{Satureja_calamintha_}}{a numeric vector}
    \item{\code{Primula_hirsuta_}}{a numeric vector}
    \item{\code{Ranunculus_montanus_}}{a numeric vector}
    \item{\code{Astrantia_minor_}}{a numeric vector}
    \item{\code{Carex_sempervirens}}{a numeric vector}
    \item{\code{Molopospermum_peloponnesiacum_}}{a numeric vector}
    \item{\code{S.Laburnum_anagyroides}}{a numeric vector}
    \item{\code{Betonica_alopecuros_}}{a numeric vector}
    \item{\code{Hieracium_pilosella_.ag.}}{a numeric vector}
    \item{\code{Juniperus_communis}}{a numeric vector}
    \item{\code{Achillea_millefolium}}{a numeric vector}
    \item{\code{Luzula_spadicea_}}{a numeric vector}
    \item{\code{Viola_biflora_}}{a numeric vector}
    \item{\code{Cardamine_resedifolia_}}{a numeric vector}
    \item{\code{Larix_decidua_}}{a numeric vector}
    \item{\code{Corallorrhiza_trifida_}}{a numeric vector}
    \item{\code{Fragaria_moschata_}}{a numeric vector}
    \item{\code{S.Alnus_viridis_}}{a numeric vector}
    \item{\code{Carex_ornithopoda_}}{a numeric vector}
    \item{\code{Geranium_silvaticum_}}{a numeric vector}
    \item{\code{Poa_chaixii_}}{a numeric vector}
    \item{\code{S.Laburnum_alpinum}}{a numeric vector}
    \item{\code{Laburnum_alpinum}}{a numeric vector}
    \item{\code{Galium_aristatum}}{a numeric vector}
    \item{\code{Listera_cordata_}}{a numeric vector}
    \item{\code{Polygonatum_officinale}}{a numeric vector}
    \item{\code{S.Salix_appendiculata_}}{a numeric vector}
    \item{\code{Galium_pumilum}}{a numeric vector}
    \item{\code{Myosotis_silvatica}}{a numeric vector}
    \item{\code{M.Tortula_spec._}}{a numeric vector}
    \item{\code{Goodyera_repens_}}{a numeric vector}
    \item{\code{Urtica_dioeca_}}{a numeric vector}
    \item{\code{Hieracium_lachenalii}}{a numeric vector}
    \item{\code{Campanula_rhomboidalis}}{a numeric vector}
    \item{\code{Geranium_bohemicum}}{a numeric vector}
    \item{\code{Stachys_alpina}}{a numeric vector}
    \item{\code{Veratrum_album}}{a numeric vector}
    \item{\code{Stellaria_nemorum_}}{a numeric vector}
    \item{\code{Achillea_macrophylla}}{a numeric vector}
    \item{\code{Agrostis_stolonifera}}{a numeric vector}
    \item{\code{B.Alnus_viridis_}}{a numeric vector}
    \item{\code{Alnus_viridis_}}{a numeric vector}
    \item{\code{Salix_appendiculata_}}{a numeric vector}
    \item{\code{Cryptogramma_crispa_}}{a numeric vector}
    \item{\code{Calamagrostis_villosa_}}{a numeric vector}
    \item{\code{Leontodon_helveticus}}{a numeric vector}
    \item{\code{Polystichum_braunii_}}{a numeric vector}
    \item{\code{Circaea_alpina}}{a numeric vector}
    \item{\code{S.Salix_caprea}}{a numeric vector}
    \item{\code{Phyteuma_ovatum_}}{a numeric vector}
    \item{\code{B.Laburnum_anagyroides}}{a numeric vector}
    \item{\code{Peucedanum_ostruthium_}}{a numeric vector}
    \item{\code{Helleborus_viridis}}{a numeric vector}
    \item{\code{Aquilegia_vulgaris_.ag._}}{a numeric vector}
    \item{\code{Streptopus_amplexifolius}}{a numeric vector}
    \item{\code{B.Salix_caprea}}{a numeric vector}
    \item{\code{Lonicera_alpigena_}}{a numeric vector}
    \item{\code{Tilia_platyphyllos}}{a numeric vector}
    \item{\code{B.Ulmus_scabra}}{a numeric vector}
    \item{\code{Ulmus_scabra}}{a numeric vector}
    \item{\code{Chaerophyllum_villarsii_.ag.}}{a numeric vector}
    \item{\code{Laserpitium_gaudinii}}{a numeric vector}
    \item{\code{Pyrola_minor}}{a numeric vector}
    \item{\code{Thalictrum_aquilegiifolium}}{a numeric vector}
    \item{\code{Viola_mirabilis_}}{a numeric vector}
    \item{\code{Arabis_turrita}}{a numeric vector}
    \item{\code{S.Prunus_padus}}{a numeric vector}
    \item{\code{Asarum_europaeum}}{a numeric vector}
    \item{\code{Circaea_lutetiana_}}{a numeric vector}
    \item{\code{Eupatorium_cannabinum_}}{a numeric vector}
    \item{\code{Prunella_vulgaris_}}{a numeric vector}
    \item{\code{Ranunculus_repens_}}{a numeric vector}
    \item{\code{Stachys_silvatica_}}{a numeric vector}
    \item{\code{Epipactis_purpurata_}}{a numeric vector}
    \item{\code{Ranunculus_ficaria}}{a numeric vector}
    \item{\code{Cardamine_flexuosa}}{a numeric vector}
    \item{\code{M.Pellia_spec.}}{a numeric vector}
    \item{\code{Equisetum_silvaticum}}{a numeric vector}
    \item{\code{Ranunculus_lanuginosus}}{a numeric vector}
    \item{\code{Arum_maculatum}}{a numeric vector}
    \item{\code{Veronica_montana}}{a numeric vector}
    \item{\code{Tussilago_farfara_}}{a numeric vector}
    \item{\code{B.Pinus_nigra_}}{a numeric vector}
    \item{\code{Rhamnus_cathartica}}{a numeric vector}
    \item{\code{S.Pinus_silvestris}}{a numeric vector}
    \item{\code{M.Homalia_trichomanoides}}{a numeric vector}
    \item{\code{Cystopteris_fragilis}}{a numeric vector}
    \item{\code{Laserpitium_latifolium}}{a numeric vector}
    \item{\code{M.Didymodon_rubellus}}{a numeric vector}
    \item{\code{M.Rhynchostegium_spec.}}{a numeric vector}
    \item{\code{Circaea_x_intermedia}}{a numeric vector}
    \item{\code{M.Pohlia_spec.}}{a numeric vector}
    \item{\code{M.Campylium_spec._}}{a numeric vector}
    \item{\code{Phyllitis_scolopendrium_}}{a numeric vector}
    \item{\code{Carex_pendula_}}{a numeric vector}
    \item{\code{Aconitum_vulparia_.ag.}}{a numeric vector}
    \item{\code{Carex_alba}}{a numeric vector}
    \item{\code{Stellaria_media_}}{a numeric vector}
    \item{\code{Calamagrostis_epigeios}}{a numeric vector}
    \item{\code{Carex_contigua}}{a numeric vector}
    \item{\code{Juncus_effusus}}{a numeric vector}
    \item{\code{Carex_remota}}{a numeric vector}
    \item{\code{Silene_flos.cuculi}}{a numeric vector}
    \item{\code{Equisetum_arvense_}}{a numeric vector}
    \item{\code{S.Clematis_vitalba}}{a numeric vector}
    \item{\code{Vicia_dumetorum_}}{a numeric vector}
    \item{\code{Dactylis_aschersoniana}}{a numeric vector}
    \item{\code{Lonicera_periclymenum_}}{a numeric vector}
    \item{\code{Chrysosplenium_alternifolium}}{a numeric vector}
    \item{\code{Corydalis_cava}}{a numeric vector}
    \item{\code{Leucojum_vernum_}}{a numeric vector}
    \item{\code{Trifolium_pratense}}{a numeric vector}
    \item{\code{Allium_ursinum}}{a numeric vector}
    \item{\code{M.Plagiopus_oederi}}{a numeric vector}
    \item{\code{M.Pogonatum_spec._}}{a numeric vector}
    \item{\code{Anemone_ranunculoides_}}{a numeric vector}
    \item{\code{Cardamine_bulbifera_}}{a numeric vector}
    \item{\code{Cardamine_kitaibelii}}{a numeric vector}
    \item{\code{Cardamine_pratensis_}}{a numeric vector}
    \item{\code{Rumex_sanguineus}}{a numeric vector}
    \item{\code{Cirsium_oleraceum_}}{a numeric vector}
    \item{\code{S.Berberis_vulgaris_}}{a numeric vector}
    \item{\code{Cirsium_arvense_}}{a numeric vector}
    \item{\code{Clematis_vitalba}}{a numeric vector}
    \item{\code{Gentiana_ciliata}}{a numeric vector}
    \item{\code{Lamium_maculatum}}{a numeric vector}
    \item{\code{Lotus_uliginosus}}{a numeric vector}
    \item{\code{Lapsana_communis}}{a numeric vector}
    \item{\code{Ranunculus_auricomus_.ag._}}{a numeric vector}
    \item{\code{Glechoma_hederaceum_}}{a numeric vector}
    \item{\code{Adoxa_moschatellina_}}{a numeric vector}
    \item{\code{S.Rosa_spinosissima_}}{a numeric vector}
    \item{\code{Daphne_laureola_}}{a numeric vector}
    \item{\code{Aesculus_hippocastanum}}{a numeric vector}
    \item{\code{Prunus_padus}}{a numeric vector}
    \item{\code{Stellaria_alsine}}{a numeric vector}
    \item{\code{Galium_palustre_.ag.}}{a numeric vector}
    \item{\code{Alnus_incana}}{a numeric vector}
    \item{\code{Carex_leporina}}{a numeric vector}
    \item{\code{Dryopteris_borreri}}{a numeric vector}
    \item{\code{Equisetum_maximum_}}{a numeric vector}
    \item{\code{S.Evonymus_latifolia}}{a numeric vector}
    \item{\code{Lilium_martagon_}}{a numeric vector}
    \item{\code{Valeriana_dioeca}}{a numeric vector}
    \item{\code{M.Hookeria_lucens_}}{a numeric vector}
    \item{\code{S.Pseudotsuga_taxifolia_}}{a numeric vector}
    \item{\code{B.Tilia_platyphyllos}}{a numeric vector}
    \item{\code{S.Tilia_platyphyllos}}{a numeric vector}
    \item{\code{Atropa_belladonna_}}{a numeric vector}
    \item{\code{Filipendula_ulmaria_}}{a numeric vector}
    \item{\code{S.Daphne_laureola_}}{a numeric vector}
    \item{\code{Scilla_bifolia}}{a numeric vector}
    \item{\code{B.Cornus_mas}}{a numeric vector}
    \item{\code{S.Cornus_mas}}{a numeric vector}
    \item{\code{Cyclamen_europaeum}}{a numeric vector}
    \item{\code{B.Ilex_aquifolium_}}{a numeric vector}
    \item{\code{Lathyrus_venetus}}{a numeric vector}
    \item{\code{Potentilla_micrantha}}{a numeric vector}
    \item{\code{Primula_veris_}}{a numeric vector}
    \item{\code{Ulmus_carpinifolia}}{a numeric vector}
    \item{\code{Equisetum_hiemale_}}{a numeric vector}
    \item{\code{Listera_ovata_}}{a numeric vector}
    \item{\code{Helleborus_foetidus_}}{a numeric vector}
    \item{\code{Orchis_mascula}}{a numeric vector}
    \item{\code{Trifolium_medium}}{a numeric vector}
    \item{\code{Ribes_spec._}}{a numeric vector}
    \item{\code{Berberis_vulgaris_}}{a numeric vector}
    \item{\code{Chrysanthemum_corymbosum}}{a numeric vector}
    \item{\code{Lithospermum_purpureo.coeruleum_}}{a numeric vector}
    \item{\code{Agrostis_gigantea_}}{a numeric vector}
    \item{\code{Sambucus_ebulus_}}{a numeric vector}
    \item{\code{Alliaria_officinalis}}{a numeric vector}
    \item{\code{Campanula_persicifolia}}{a numeric vector}
    \item{\code{B.Ulmus_carpinifolia}}{a numeric vector}
    \item{\code{Bromus_erectus}}{a numeric vector}
    \item{\code{B.Pirus_malus_}}{a numeric vector}
    \item{\code{B.Alnus_incana}}{a numeric vector}
    \item{\code{B.Acer_opalus_}}{a numeric vector}
    \item{\code{S.Acer_opalus_}}{a numeric vector}
    \item{\code{Cotoneaster_spec._}}{a numeric vector}
    \item{\code{Hypericum_hirsutum}}{a numeric vector}
    \item{\code{Galium_aparine_.ag._}}{a numeric vector}
    \item{\code{Holcus_lanatus}}{a numeric vector}
    \item{\code{Torilis_japonica}}{a numeric vector}
    \item{\code{Peucedanum_cervaria_}}{a numeric vector}
    \item{\code{B.Cornus_sanguinea}}{a numeric vector}
    \item{\code{S.Rubus_caesius_}}{a numeric vector}
    \item{\code{S.Pirus_piraster}}{a numeric vector}
    \item{\code{Ornithogalum_pyrenaicum_}}{a numeric vector}
    \item{\code{S.Ulmus_carpinifolia}}{a numeric vector}
    \item{\code{Geum_rivale_}}{a numeric vector}
    \item{\code{Gagea_lutea_}}{a numeric vector}
    \item{\code{Asperula_taurina}}{a numeric vector}
    \item{\code{B.Tilia_spec._}}{a numeric vector}
    \item{\code{Cardamine_impatiens_}}{a numeric vector}
    \item{\code{S.Alnus_incana}}{a numeric vector}
    \item{\code{Prunus_domestica}}{a numeric vector}
    \item{\code{B.Prunus_padus}}{a numeric vector}
    \item{\code{Orobanche_vulgaris}}{a numeric vector}
    \item{\code{Rumex_arifolius_}}{a numeric vector}
    \item{\code{Galanthus_nivalis_}}{a numeric vector}
    \item{\code{B.Viscum_album}}{a numeric vector}
    \item{\code{B.Juglans_regia_}}{a numeric vector}
    \item{\code{Campanula_rapunculoides_}}{a numeric vector}
    \item{\code{Pirus_malus_}}{a numeric vector}
    \item{\code{Epipactis_microphylla_}}{a numeric vector}
    \item{\code{S.Aesculus_hippocastanum}}{a numeric vector}
    \item{\code{Prunella_grandiflora}}{a numeric vector}
    \item{\code{Scabiosa_columbaria_}}{a numeric vector}
    \item{\code{Anthericum_ramosum}}{a numeric vector}
    \item{\code{S.Cotoneaster_spec._}}{a numeric vector}
    \item{\code{Arctium_spec._}}{a numeric vector}
    \item{\code{Lathyrus_niger}}{a numeric vector}
    \item{\code{M.Trichostomum_spec.}}{a numeric vector}
    \item{\code{Vicia_silvatica_}}{a numeric vector}
    \item{\code{M.Radula_complanata_}}{a numeric vector}
    \item{\code{Crepis_praemorsa}}{a numeric vector}
    \item{\code{Geranium_sanguineum_}}{a numeric vector}
    \item{\code{Pirus_piraster}}{a numeric vector}
    \item{\code{Heracleum_alpinum_}}{a numeric vector}
    \item{\code{M.Rhodobryum_roseum_}}{a numeric vector}
    \item{\code{Bellidiastrum_michelii}}{a numeric vector}
    \item{\code{S.Rhamnus_alpina}}{a numeric vector}
    \item{\code{Carduus_defloratus}}{a numeric vector}
    \item{\code{Digitalis_lutea_}}{a numeric vector}
    \item{\code{Gymnadenia_conopea}}{a numeric vector}
    \item{\code{Teucrium_chamaedrys_}}{a numeric vector}
    \item{\code{Anthyllis_vulneraria}}{a numeric vector}
    \item{\code{Bupleurum_falcatum}}{a numeric vector}
    \item{\code{Gentiana_lutea}}{a numeric vector}
    \item{\code{Lastrea_robertiana}}{a numeric vector}
    \item{\code{Amelanchier_ovalis}}{a numeric vector}
    \item{\code{Acer_opalus_}}{a numeric vector}
    \item{\code{Agropyron_caninum_}}{a numeric vector}
    \item{\code{Sorbus_mougeotii}}{a numeric vector}
    \item{\code{B.Laburnum_alpinum}}{a numeric vector}
    \item{\code{Melica_ciliata}}{a numeric vector}
    \item{\code{S.Buxus_sempervirens}}{a numeric vector}
    \item{\code{Centaurea_triumfettii_}}{a numeric vector}
    \item{\code{S.Sorbus_mougeotii}}{a numeric vector}
    \item{\code{Cypripedium_calceolus_}}{a numeric vector}
    \item{\code{Evonymus_latifolia}}{a numeric vector}
    \item{\code{M.Amblystegiella_confervoides_}}{a numeric vector}
    \item{\code{Vicia_cracca_.ag._}}{a numeric vector}
    \item{\code{Erigeron_atticus}}{a numeric vector}
    \item{\code{Thymus_serpyllum_.ag._}}{a numeric vector}
    \item{\code{Galium_lucidum}}{a numeric vector}
    \item{\code{Globularia_cordifolia_}}{a numeric vector}
    \item{\code{Hippocrepis_comosa}}{a numeric vector}
    \item{\code{Inula_conyza}}{a numeric vector}
    \item{\code{Achnatherum_calamagrostis_}}{a numeric vector}
    \item{\code{Saxifraga_aizoon}}{a numeric vector}
    \item{\code{Viola_odorata_}}{a numeric vector}
    \item{\code{Campanula_cochleariifolia_}}{a numeric vector}
    \item{\code{Carlina_vulgaris}}{a numeric vector}
    \item{\code{Thesium_alpinum_}}{a numeric vector}
    \item{\code{Thlaspi_montanum}}{a numeric vector}
    \item{\code{Lotus_corniculatus}}{a numeric vector}
    \item{\code{Asplenium_ruta.muraria}}{a numeric vector}
    \item{\code{M.Grimmia_spec._}}{a numeric vector}
    \item{\code{Athyrium_alpestre_}}{a numeric vector}
    \item{\code{Viola_alba}}{a numeric vector}
    \item{\code{Cirsium_palustre}}{a numeric vector}
    \item{\code{M.Blasia_pusilla}}{a numeric vector}
    \item{\code{Aconitum_compactum_.ag._}}{a numeric vector}
    \item{\code{Cirsium_vulgare_}}{a numeric vector}
    \item{\code{Aconitum_paniculatum}}{a numeric vector}
    \item{\code{Trollius_europaeus}}{a numeric vector}
    \item{\code{Carex_ferruginea}}{a numeric vector}
    \item{\code{Crepis_blattarioides}}{a numeric vector}
    \item{\code{M.Rhytidiadelphus_squarrosus}}{a numeric vector}
    \item{\code{M.Trichocolea_tomentella}}{a numeric vector}
    \item{\code{M.Fegatella_spec._}}{a numeric vector}
    \item{\code{Crepis_paludosa_}}{a numeric vector}
    \item{\code{Carduus_nutans}}{a numeric vector}
    \item{\code{M.Riccardia_multifida_}}{a numeric vector}
    \item{\code{Hypericum_maculatum_}}{a numeric vector}
    \item{\code{Cicerbita_alpina}}{a numeric vector}
    \item{\code{S.Rosa_obtusifolia}}{a numeric vector}
    \item{\code{Astrantia_major_}}{a numeric vector}
    \item{\code{Cirsium_erisithales_}}{a numeric vector}
    \item{\code{Heracleum_montanum}}{a numeric vector}
    \item{\code{Chaerophyllum_nitidum_}}{a numeric vector}
    \item{\code{Caltha_palustris}}{a numeric vector}
    \item{\code{M.Cephalozia_spec.}}{a numeric vector}
    \item{\code{Campanula_scheuchzeri_}}{a numeric vector}
    \item{\code{Gentiana_amarella_}}{a numeric vector}
    \item{\code{Achillea_atrata_}}{a numeric vector}
    \item{\code{Veronica_alpina_}}{a numeric vector}
    \item{\code{Epilobium_parviflorum_}}{a numeric vector}
    \item{\code{M.Tritomaria_exsecta}}{a numeric vector}
    \item{\code{Salix_nigricans_}}{a numeric vector}
    \item{\code{Astragalus_glycyphyllos_}}{a numeric vector}
    \item{\code{Lycopodium_annotinum}}{a numeric vector}
    \item{\code{M.Dicranodontium_spec.}}{a numeric vector}
    \item{\code{Satureja_grandiflora}}{a numeric vector}
    \item{\code{Chaerophyllum_silvestre_}}{a numeric vector}
    \item{\code{Cardamine_trifolia}}{a numeric vector}
    \item{\code{Lunaria_rediviva}}{a numeric vector}
    \item{\code{Ranunculus_acer_.ag.}}{a numeric vector}
    \item{\code{Poa_hybrida_}}{a numeric vector}
    \item{\code{Epilobium_angustifolium_}}{a numeric vector}
    \item{\code{Soldanella_alpina_}}{a numeric vector}
    \item{\code{Aposeris_foetida}}{a numeric vector}
    \item{\code{Gentiana_prostrata}}{a numeric vector}
    \item{\code{Polygonum_bistorta}}{a numeric vector}
    \item{\code{Veronica_beccabunga_.ag.}}{a numeric vector}
    \item{\code{Senecio_alpinus_}}{a numeric vector}
    \item{\code{Festuca_pratensis_}}{a numeric vector}
    \item{\code{Gentiana_punctata_}}{a numeric vector}
    \item{\code{Laserpitium_prutenicum}}{a numeric vector}
    \item{\code{Petasites_paradoxus_}}{a numeric vector}
    \item{\code{Epilobium_alpestre}}{a numeric vector}
    \item{\code{Myosotis_alpestris}}{a numeric vector}
    \item{\code{Delphinium_elatum_}}{a numeric vector}
    \item{\code{Saxifraga_stellaris_}}{a numeric vector}
    \item{\code{Carduus_personata_}}{a numeric vector}
    \item{\code{S.Pinus_montana_}}{a numeric vector}
    \item{\code{S.Lonicera_coerulea_}}{a numeric vector}
    \item{\code{Chaerophyllum_temulum_}}{a numeric vector}
    \item{\code{Geranium_lividum}}{a numeric vector}
    \item{\code{Myrrhis_odorata_}}{a numeric vector}
    \item{\code{Campanula_latifolia_}}{a numeric vector}
    \item{\code{Sedum_album_}}{a numeric vector}
    \item{\code{Rhamnus_alpina}}{a numeric vector}
    \item{\code{Bupleurum_longifolium_}}{a numeric vector}
    \item{\code{Cynoglossum_germanicum}}{a numeric vector}
    \item{\code{Colchicum_autumnale_}}{a numeric vector}
    \item{\code{Alchemilla_xanthochlora_}}{a numeric vector}
    \item{\code{Crepis_biennis}}{a numeric vector}
    \item{\code{Parnassia_palustris_}}{a numeric vector}
    \item{\code{Petasites_hybridus}}{a numeric vector}
    \item{\code{Hypericum_humifusum_}}{a numeric vector}
    \item{\code{Plantago_media}}{a numeric vector}
    \item{\code{B.Fraxinus_ornus}}{a numeric vector}
    \item{\code{S.Fraxinus_ornus}}{a numeric vector}
    \item{\code{B.Ostrya_carpinifolia_}}{a numeric vector}
    \item{\code{Peucedanum_venetum}}{a numeric vector}
    \item{\code{S.Prunus_mahaleb}}{a numeric vector}
    \item{\code{S.Ruscus_aculeatus}}{a numeric vector}
    \item{\code{Carex_halleriana}}{a numeric vector}
    \item{\code{Lilium_croceum}}{a numeric vector}
    \item{\code{Viola_scotophylla_}}{a numeric vector}
    \item{\code{Laserpitium_siler_}}{a numeric vector}
    \item{\code{Teucrium_montanum_}}{a numeric vector}
    \item{\code{Allium_oleraceum}}{a numeric vector}
    \item{\code{S.Crataegus_monticola_}}{a numeric vector}
    \item{\code{Ophrys_apifera}}{a numeric vector}
    \item{\code{Prunus_mahaleb}}{a numeric vector}
    \item{\code{Asplenium_fontanum}}{a numeric vector}
    \item{\code{B.Clematis_vitalba}}{a numeric vector}
    \item{\code{S.Crataegus_palmstruchii}}{a numeric vector}
    \item{\code{Orchis_pallens}}{a numeric vector}
    \item{\code{Staphylea_pinnata_}}{a numeric vector}
    \item{\code{Viola_collina_}}{a numeric vector}
    \item{\code{S.Rubus_saxatilis_}}{a numeric vector}
    \item{\code{M.Climacium_dendroides}}{a numeric vector}
    \item{\code{Briza_media_}}{a numeric vector}
    \item{\code{Galium_rubrum_}}{a numeric vector}
    \item{\code{Achillea_stricta}}{a numeric vector}
    \item{\code{Laserpitium_halleri_}}{a numeric vector}
    \item{\code{Laburnum_anagyroides}}{a numeric vector}
    \item{\code{Polystichum_setiferum_}}{a numeric vector}
    \item{\code{Fraxinus_ornus}}{a numeric vector}
    \item{\code{Helleborus_niger}}{a numeric vector}
    \item{\code{Polygonum_convolvulus_}}{a numeric vector}
    \item{\code{Ruscus_aculeatus}}{a numeric vector}
    \item{\code{Cornus_mas}}{a numeric vector}
    \item{\code{Geranium_nodosum}}{a numeric vector}
    \item{\code{Iris_graminea_}}{a numeric vector}
    \item{\code{Asparagus_tenuifolius_}}{a numeric vector}
    \item{\code{Carex_acutiformis_}}{a numeric vector}
    \item{\code{Physalis_alkekengi}}{a numeric vector}
    \item{\code{Alnus_glutinosa_}}{a numeric vector}
    \item{\code{Chrysosplenium_oppositifolium_}}{a numeric vector}
    \item{\code{Glyceria_fluitans_}}{a numeric vector}
    \item{\code{Lysimachia_vulgaris_}}{a numeric vector}
    \item{\code{Myosotis_palustris}}{a numeric vector}
    \item{\code{Scirpus_silvaticus}}{a numeric vector}
    \item{\code{Typhoides_arundinacea_}}{a numeric vector}
    \item{\code{Mentha_aquatica_}}{a numeric vector}
    \item{\code{Phragmites_communis_}}{a numeric vector}
    \item{\code{M.Conocephalum_spec.}}{a numeric vector}
    \item{\code{Carex_strigosa}}{a numeric vector}
    \item{\code{Rumex_obtusifolius}}{a numeric vector}
    \item{\code{Lysimachia_nummularia_}}{a numeric vector}
    \item{\code{Lythrum_salicaria_}}{a numeric vector}
    \item{\code{Polygonum_hydropiper_.ag._}}{a numeric vector}
    \item{\code{Scrophularia_alata}}{a numeric vector}
    \item{\code{Solanum_dulcamara_}}{a numeric vector}
    \item{\code{Dipsacus_pilosus}}{a numeric vector}
    \item{\code{Cardamine_amara_}}{a numeric vector}
    \item{\code{S.Humulus_lupulus_}}{a numeric vector}
    \item{\code{B.Populus_alba_._nigra}}{a numeric vector}
    \item{\code{M.Cratoneurum_spec._}}{a numeric vector}
    \item{\code{M.Calliergonella_cuspidata}}{a numeric vector}
    \item{\code{Carex_flava_.ag.}}{a numeric vector}
    \item{\code{M.Sphagnum_spec.}}{a numeric vector}
    \item{\code{Carex_elata_}}{a numeric vector}
    \item{\code{Humulus_lupulus_}}{a numeric vector}
    \item{\code{S.Salix_elaeagnos_}}{a numeric vector}
    \item{\code{Solidago_canadensis_}}{a numeric vector}
    \item{\code{B.Ulmus_spec._}}{a numeric vector}
    \item{\code{S.Salix_nigricans_}}{a numeric vector}
    \item{\code{S.Ulmus_spec._}}{a numeric vector}
    \item{\code{B.Salix_alba}}{a numeric vector}
    \item{\code{S.Populus_alba_._nigra}}{a numeric vector}
    \item{\code{Iris_pseudacorus}}{a numeric vector}
    \item{\code{B.Salix_fragilis}}{a numeric vector}
    \item{\code{S.Salix_purpurea}}{a numeric vector}
    \item{\code{S.Salix_triandra}}{a numeric vector}
    \item{\code{Solidago_serotina_}}{a numeric vector}
    \item{\code{Symphytum_officinale}}{a numeric vector}
    \item{\code{Galium_uliginosum_}}{a numeric vector}
    \item{\code{Bromus_racemosus}}{a numeric vector}
    \item{\code{Convolvulus_sepium}}{a numeric vector}
    \item{\code{Hypericum_acutum}}{a numeric vector}
    \item{\code{Stellaria_aquatica}}{a numeric vector}
    \item{\code{Veronica_hederifolia}}{a numeric vector}
    \item{\code{Equisetum_palustre}}{a numeric vector}
    \item{\code{Saponaria_officinalis_}}{a numeric vector}
    \item{\code{Carduus_crispus_}}{a numeric vector}
    \item{\code{Lycopus_europaeus_}}{a numeric vector}
    \item{\code{Arrhenatherum_elatius_}}{a numeric vector}
    \item{\code{B.Salix_daphnoides}}{a numeric vector}
    \item{\code{B.Salix_elaeagnos_}}{a numeric vector}
    \item{\code{B.Frangula_alnus}}{a numeric vector}
    \item{\code{Geranium_palustre_}}{a numeric vector}
    \item{\code{B.Salix_cinerea_}}{a numeric vector}
    \item{\code{S.Salix_cinerea_}}{a numeric vector}
    \item{\code{B.Salix_purpurea}}{a numeric vector}
    \item{\code{Salix_cinerea_}}{a numeric vector}
    \item{\code{S.Salix_daphnoides}}{a numeric vector}
    \item{\code{Lithospermum_officinale_}}{a numeric vector}
    \item{\code{S.Hippophae_rhamnoides}}{a numeric vector}
    \item{\code{Veronica_serpyllifolia}}{a numeric vector}
    \item{\code{B.Salix_pentandra_}}{a numeric vector}
    \item{\code{M.Amblystegium_spec.}}{a numeric vector}
    \item{\code{M.Limnobium_palustre}}{a numeric vector}
    \item{\code{Saponaria_ocymoides_}}{a numeric vector}
    \item{\code{Narcissus_poeticus}}{a numeric vector}
    \item{\code{Trifolium_alpestre}}{a numeric vector}
    \item{\code{S.Ostrya_carpinifolia_}}{a numeric vector}
    \item{\code{Knautia_drymeia_}}{a numeric vector}
    \item{\code{Polygonum_dumetorum_}}{a numeric vector}
    \item{\code{Thalictrum_minus}}{a numeric vector}
    \item{\code{Daucus_carota_}}{a numeric vector}
    \item{\code{Dianthus_seguieri_}}{a numeric vector}
    \item{\code{Euphrasia_nemorosa}}{a numeric vector}
    \item{\code{Leontodon_hispidus}}{a numeric vector}
    \item{\code{Linum_catharticum_}}{a numeric vector}
    \item{\code{Ostrya_carpinifolia_}}{a numeric vector}
    \item{\code{Picris_hieracioides_}}{a numeric vector}
    \item{\code{Pimpinella_saxifraga}}{a numeric vector}
    \item{\code{Plantago_lanceolata_}}{a numeric vector}
    \item{\code{Sanguisorba_minor_}}{a numeric vector}
    \item{\code{Trifolium_repens}}{a numeric vector}
    \item{\code{Trifolium_rubens}}{a numeric vector}
    \item{\code{Carlina_acaulis_}}{a numeric vector}
    \item{\code{Cerastium_caespitosum_}}{a numeric vector}
    \item{\code{Euphrasia_stricta_}}{a numeric vector}
    \item{\code{Myosotis_caespitosa_}}{a numeric vector}
    \item{\code{Plantago_major}}{a numeric vector}
    \item{\code{Sedum_sexangulare_}}{a numeric vector}
    \item{\code{Selinum_carvifolia}}{a numeric vector}
    \item{\code{Vicia_angustifolia}}{a numeric vector}
    \item{\code{Succisa_pratensis_}}{a numeric vector}
    \item{\code{Symphytum_tuberosum_}}{a numeric vector}
    \item{\code{Mespilus_germanica}}{a numeric vector}
    \item{\code{Centaurea_dubia_}}{a numeric vector}
    \item{\code{Phyteuma_scorzonerifolium_}}{a numeric vector}
    \item{\code{S.Genista_germanica_}}{a numeric vector}
    \item{\code{Silene_alba_}}{a numeric vector}
    \item{\code{Gnaphalium_silvaticum_}}{a numeric vector}
    \item{\code{Rumex_acetosella}}{a numeric vector}
    \item{\code{Viola_tricolor}}{a numeric vector}
    \item{\code{Agrimonia_eupatoria_}}{a numeric vector}
    \item{\code{Euphrasia_cisalpina_}}{a numeric vector}
    \item{\code{Asperula_purpurea_}}{a numeric vector}
    \item{\code{Centaurium_umbellatum_}}{a numeric vector}
    \item{\code{Rumex_scutatus}}{a numeric vector}
    \item{\code{Torilis_anthriscus}}{a numeric vector}
    \item{\code{S.Lonicera_japonica_}}{a numeric vector}
    \item{\code{Thesium_bavarum_}}{a numeric vector}
    \item{\code{B.Lonicera_periclymenum_}}{a numeric vector}
    \item{\code{B.Crataegus_spec._}}{a numeric vector}
    \item{\code{Dianthus_superbus_}}{a numeric vector}
    \item{\code{B.Pirus_piraster}}{a numeric vector}
    \item{\code{Populus_alba_._nigra}}{a numeric vector}
    \item{\code{Myosotis_arvensis_}}{a numeric vector}
    \item{\code{Potentilla_alba_}}{a numeric vector}
    \item{\code{Carex_verna_}}{a numeric vector}
    \item{\code{Erigeron_canadensis_}}{a numeric vector}
    \item{\code{Genista_sagittalis}}{a numeric vector}
    \item{\code{Cardamine_enneaphyllos}}{a numeric vector}
    \item{\code{Doronicum_pardalianches_}}{a numeric vector}
    \item{\code{Veratrum_nigrum_}}{a numeric vector}
    \item{\code{B.Celtis_australis}}{a numeric vector}
    \item{\code{S.Celtis_australis}}{a numeric vector}
    \item{\code{Celtis_australis}}{a numeric vector}
    \item{\code{Lonicera_japonica_}}{a numeric vector}
    \item{\code{Clematis_recta}}{a numeric vector}
    \item{\code{Laurus_nobilis}}{a numeric vector}
    \item{\code{S.Cotinus_coggygria_}}{a numeric vector}
    \item{\code{Muscari_comosum_}}{a numeric vector}
    \item{\code{Aster_amellus_}}{a numeric vector}
    \item{\code{Cnidium_silaifolium_}}{a numeric vector}
    \item{\code{Melissa_officinalis_}}{a numeric vector}
    \item{\code{Carex_austroalpina}}{a numeric vector}
    \item{\code{Limodorum_abortivum_}}{a numeric vector}
    \item{\code{Peucedanum_austriacum_}}{a numeric vector}
    \item{\code{Buphthalmum_salicifolium}}{a numeric vector}
    \item{\code{Solanum_nigrum}}{a numeric vector}
    \item{\code{Chamaecytisus_hirsutus}}{a numeric vector}
    \item{\code{Triticum_spec.}}{a numeric vector}
    \item{\code{Cotinus_coggygria_}}{a numeric vector}
    \item{\code{Sonchus_spec._}}{a numeric vector}
    \item{\code{Rhamnus_saxatilis_}}{a numeric vector}
    \item{\code{Inula_hirta_}}{a numeric vector}
    \item{\code{S.Lonicera_caprifolium}}{a numeric vector}
    \item{\code{Lonicera_caprifolium}}{a numeric vector}
    \item{\code{Crepis_froelichiana_}}{a numeric vector}
    \item{\code{Leontodon_incanus_}}{a numeric vector}
    \item{\code{Ononis_pusilla}}{a numeric vector}
    \item{\code{Stachys_recta_}}{a numeric vector}
    \item{\code{Koeleria_pyramidata_.ag.}}{a numeric vector}
    \item{\code{S.Rhamnus_saxatilis_}}{a numeric vector}
    \item{\code{Ballota_nigra_}}{a numeric vector}
    \item{\code{Bupleurum_ranunculoides_}}{a numeric vector}
    \item{\code{Chrysopogon_gryllus_}}{a numeric vector}
    \item{\code{Dianthus_silvester}}{a numeric vector}
    \item{\code{Thesium_linophyllon_}}{a numeric vector}
    \item{\code{Centaurea_scabiosa}}{a numeric vector}
    \item{\code{Melampyrum_cristatum}}{a numeric vector}
    \item{\code{Orobanche_hederae_}}{a numeric vector}
    \item{\code{Dictamnus_albus_}}{a numeric vector}
    \item{\code{Inula_salicina}}{a numeric vector}
    \item{\code{Rosa_gallica}}{a numeric vector}
    \item{\code{Rosa_jundzillii_}}{a numeric vector}
    \item{\code{Asperula_tinctoria}}{a numeric vector}
    \item{\code{Asperula_cyanchica}}{a numeric vector}
    \item{\code{Phleum_boehmeri_}}{a numeric vector}
    \item{\code{Vicia_tenuifolia}}{a numeric vector}
    \item{\code{Campanula_rapunculus}}{a numeric vector}
    \item{\code{Helianthemum_nummularium_.ag._}}{a numeric vector}
    \item{\code{Aceras_anthropophorum_}}{a numeric vector}
    \item{\code{Ranunculus_bulbosus_}}{a numeric vector}
    \item{\code{S.Rosa_micrantha}}{a numeric vector}
    \item{\code{Epilobium_collinum}}{a numeric vector}
    \item{\code{Epilobium_lanceolatum_}}{a numeric vector}
    \item{\code{Hieracium_cymosum_}}{a numeric vector}
    \item{\code{S.Rosa_agrestis_}}{a numeric vector}
    \item{\code{Sedum_montanum}}{a numeric vector}
    \item{\code{Silene_coronaria}}{a numeric vector}
    \item{\code{Trifolium_arvense_}}{a numeric vector}
    \item{\code{Vicia_hirsuta_}}{a numeric vector}
    \item{\code{Vicia_onobrychioides}}{a numeric vector}
    \item{\code{Anacamptis_pyramidalis}}{a numeric vector}
    \item{\code{Potentilla_puberula_}}{a numeric vector}
    \item{\code{S.Daphne_alpina_}}{a numeric vector}
    \item{\code{Orchis_militaris}}{a numeric vector}
    \item{\code{Polygala_amarella_}}{a numeric vector}
    \item{\code{S.Rosa_jundzillii_}}{a numeric vector}
    \item{\code{B.Sorbus_mougeotii}}{a numeric vector}
    \item{\code{S.Vitis_silvestris}}{a numeric vector}
    \item{\code{Sedum_acre}}{a numeric vector}
    \item{\code{Fragaria_viridis}}{a numeric vector}
    \item{\code{Ajuga_genevensis}}{a numeric vector}
    \item{\code{Sedum_rupestre}}{a numeric vector}
    \item{\code{Coronilla_coronata}}{a numeric vector}
    \item{\code{Galeopsis_ladanum_}}{a numeric vector}
    \item{\code{Orlaya_grandiflora}}{a numeric vector}
    \item{\code{Sempervivum_alpinum_}}{a numeric vector}
    \item{\code{Aster_linosyris_}}{a numeric vector}
    \item{\code{Lactuca_perennis}}{a numeric vector}
    \item{\code{S.Sorbus_latifolia}}{a numeric vector}
    \item{\code{Prunus_cerasus}}{a numeric vector}
    \item{\code{Platanthera_chlorantha}}{a numeric vector}
    \item{\code{Filipendula_hexapetala}}{a numeric vector}
    \item{\code{Galium_boreale}}{a numeric vector}
    \item{\code{Veronica_teucrium_}}{a numeric vector}
    \item{\code{Coronilla_varia_}}{a numeric vector}
    \item{\code{Arabis_pauciflora_}}{a numeric vector}
    \item{\code{Colutea_arborescens_}}{a numeric vector}
    \item{\code{Cirsium_acaule}}{a numeric vector}
    \item{\code{Helictotrichon_pratense_}}{a numeric vector}
    \item{\code{Trifolium_montanum}}{a numeric vector}
    \item{\code{M.Abietinella_abietina}}{a numeric vector}
    \item{\code{Gymnadenia_odoratissima_}}{a numeric vector}
    \item{\code{Peucedanum_verticillare_}}{a numeric vector}
    \item{\code{Vicia_incana}}{a numeric vector}
    \item{\code{Polygala_vulgaris_}}{a numeric vector}
    \item{\code{Viola_canina}}{a numeric vector}
    \item{\code{S.Genista_tinctoria_}}{a numeric vector}
    \item{\code{Hypochoeris_radicata}}{a numeric vector}
    \item{\code{Jasione_montana_}}{a numeric vector}
    \item{\code{Odontites_lutea_}}{a numeric vector}
    \item{\code{Trifolium_campestre_}}{a numeric vector}
    \item{\code{Hieracium_piloselloides_}}{a numeric vector}
    \item{\code{Antennaria_dioeca_}}{a numeric vector}
    \item{\code{Polygala_comosa_}}{a numeric vector}
    \item{\code{Euphrasia_micrantha_}}{a numeric vector}
    \item{\code{Lathraea_squamaria}}{a numeric vector}
    \item{\code{Dianthus_carthusianorum_}}{a numeric vector}
    \item{\code{Orobanche_rapum.genistae}}{a numeric vector}
    \item{\code{Artemisia_vulgaris}}{a numeric vector}
    \item{\code{Carex_tomentosa_}}{a numeric vector}
    \item{\code{Chelidonium_majus_}}{a numeric vector}
    \item{\code{Chenopodium_album_}}{a numeric vector}
    \item{\code{Poa_annua_}}{a numeric vector}
    \item{\code{Sinapis_arvensis}}{a numeric vector}
    \item{\code{Bellis_perennis_}}{a numeric vector}
    \item{\code{Rorippa_spec._}}{a numeric vector}
    \item{\code{Cardamine_hirsuta_}}{a numeric vector}
    \item{\code{Mentha_spicata}}{a numeric vector}
    \item{\code{B.Salix_triandra}}{a numeric vector}
    \item{\code{Festuca_arundinacea_}}{a numeric vector}
    \item{\code{Matricaria_chamomilla_}}{a numeric vector}
    \item{\code{S.Salix_viminalis_}}{a numeric vector}
    \item{\code{Lamium_purpureum}}{a numeric vector}
    \item{\code{Raphanus_raphanistrum_}}{a numeric vector}
    \item{\code{Senecio_vulgaris}}{a numeric vector}
    \item{\code{Sisymbrium_officinale_}}{a numeric vector}
    \item{\code{Barbarea_vulgaris_}}{a numeric vector}
    \item{\code{Capsella_bursa.pastoris_}}{a numeric vector}
    \item{\code{Brassica_oleracea_}}{a numeric vector}
    \item{\code{Scutellaria_galericulata}}{a numeric vector}
    \item{\code{Cladium_mariscus}}{a numeric vector}
    \item{\code{Comarum_palustre}}{a numeric vector}
    \item{\code{Lastrea_thelypteris_}}{a numeric vector}
    \item{\code{Menyanthes_trifoliata_}}{a numeric vector}
    \item{\code{Carex_paradoxa}}{a numeric vector}
    \item{\code{Carex_elongata}}{a numeric vector}
    \item{\code{S.Salix_aurita}}{a numeric vector}
    \item{\code{Peucedanum_palustre_}}{a numeric vector}
    \item{\code{Equisetum_limosum_}}{a numeric vector}
    \item{\code{Stachys_palustris_}}{a numeric vector}
    \item{\code{Carex_riparia_}}{a numeric vector}
    \item{\code{Agrostis_canina_}}{a numeric vector}
    \item{\code{Carex_gracilis}}{a numeric vector}
    \item{\code{S.Solanum_dulcamara_}}{a numeric vector}
    \item{\code{Carex_fusca_}}{a numeric vector}
    \item{\code{Carex_davalliana}}{a numeric vector}
    \item{\code{Carex_panicea_}}{a numeric vector}
    \item{\code{B.Rhamnus_cathartica}}{a numeric vector}
    \item{\code{Oxycoccus_quadripetalus_}}{a numeric vector}
    \item{\code{Vaccinium_uliginosum}}{a numeric vector}
    \item{\code{M.Aulacomnium_spec._}}{a numeric vector}
    \item{\code{M.Orthodicranum_spec._}}{a numeric vector}
    \item{\code{Trichophorum_caespitosum}}{a numeric vector}
    \item{\code{Carex_canescens_}}{a numeric vector}
    \item{\code{M.Bazzania_spec.}}{a numeric vector}
    \item{\code{M.Hylocomiastrum_umbratum_}}{a numeric vector}
    \item{\code{Hieracium_prenanthoides_}}{a numeric vector}
    \item{\code{Allium_victorialis}}{a numeric vector}
    \item{\code{Cirsium_helenioides_}}{a numeric vector}
    \item{\code{Phyteuma_orbiculare_}}{a numeric vector}
    \item{\code{Pyrola_uniflora_}}{a numeric vector}
    \item{\code{Lycopodium_clavatum_}}{a numeric vector}
    \item{\code{Ligusticum_mutellina}}{a numeric vector}
    \item{\code{B.Pinus_cembra}}{a numeric vector}
    \item{\code{Pinus_cembra}}{a numeric vector}
    \item{\code{Poa_alpina}}{a numeric vector}
    \item{\code{Potentilla_aurea}}{a numeric vector}
    \item{\code{Alchemilla_alpina_}}{a numeric vector}
    \item{\code{Campanula_barbata_}}{a numeric vector}
    \item{\code{B.Salix_appendiculata_}}{a numeric vector}
    \item{\code{Sorbus_chamaemespilus_}}{a numeric vector}
    \item{\code{S.Sorbus_chamaemespilus_}}{a numeric vector}
    \item{\code{M.Barbilophozia_spec._}}{a numeric vector}
    \item{\code{B.Pinus_montana_}}{a numeric vector}
    \item{\code{M.Ptilidium_spec._}}{a numeric vector}
    \item{\code{Tozzia_alpina_}}{a numeric vector}
    \item{\code{Poa_remota}}{a numeric vector}
    \item{\code{Carex_pauciflora}}{a numeric vector}
    \item{\code{M.Encalypta_spec._}}{a numeric vector}
    \item{\code{Epilobium_hirsutum}}{a numeric vector}
    \item{\code{Selaginella_helvetica_}}{a numeric vector}
    \item{\code{M.Cetraria_spec.}}{a numeric vector}
    \item{\code{Pastinaca_sativa}}{a numeric vector}
    \item{\code{Pyrola_chlorantha_}}{a numeric vector}
    \item{\code{Leontodon_hyoseroides_}}{a numeric vector}
    \item{\code{Tofieldia_calyculata}}{a numeric vector}
    \item{\code{Thesium_pyrenaicum}}{a numeric vector}
    \item{\code{Galeopsis_speciosa}}{a numeric vector}
    \item{\code{M.Timmia_spec.}}{a numeric vector}
    \item{\code{M.Drepanocladus_spec._}}{a numeric vector}
    \item{\code{Clematis_alpina_}}{a numeric vector}
    \item{\code{Viola_rupestris_}}{a numeric vector}
    \item{\code{Galium_triflorum}}{a numeric vector}
    \item{\code{Linnaea_borealis}}{a numeric vector}
    \item{\code{Juniperus_nana}}{a numeric vector}
    \item{\code{Agrostis_schraderiana_}}{a numeric vector}
    \item{\code{Gentiana_purpurea_}}{a numeric vector}
    \item{\code{Pulsatilla_sulphurea}}{a numeric vector}
    \item{\code{Lonicera_coerulea_}}{a numeric vector}
    \item{\code{Selaginella_selaginoides}}{a numeric vector}
    \item{\code{Carex_stellulata}}{a numeric vector}
    \item{\code{Deschampsia_litoralis_}}{a numeric vector}
    \item{\code{Eriophorum_vaginatum}}{a numeric vector}
    \item{\code{M.Lophozia_spec.}}{a numeric vector}
    \item{\code{Sempervivum_montanum}}{a numeric vector}
    \item{\code{S.Juniperus_nana}}{a numeric vector}
    \item{\code{S.Pinus_cembra}}{a numeric vector}
    \item{\code{M.Anastrepta_orcadensis_}}{a numeric vector}
    \item{\code{M.Jungermannia_spec.}}{a numeric vector}
    \item{\code{Carex_brunnescens_}}{a numeric vector}
    \item{\code{Epipogium_aphyllum}}{a numeric vector}
    \item{\code{M.Pannaria_nebulosa_}}{a numeric vector}
    \item{\code{Phleum_alpinum}}{a numeric vector}
    \item{\code{Dryas_octopetala}}{a numeric vector}
    \item{\code{Gentiana_verna}}{a numeric vector}
    \item{\code{Geranium_rivulare_}}{a numeric vector}
    \item{\code{Hieracium_staticifolium_}}{a numeric vector}
    \item{\code{Trisetum_distichophyllum}}{a numeric vector}
    \item{\code{Astragalus_alpinus}}{a numeric vector}
    \item{\code{Arctostaphylos_alpina_}}{a numeric vector}
    \item{\code{Geum_montanum_}}{a numeric vector}
    \item{\code{Empetrum_nigrum_}}{a numeric vector}
    \item{\code{S.Rosa_stylosa}}{a numeric vector}
    \item{\code{Centaurea_nervosa_}}{a numeric vector}
    \item{\code{Pleurospermum_austriacum}}{a numeric vector}
    \item{\code{Crepis_alpestris}}{a numeric vector}
    \item{\code{Arctostaphylos_uva.ursi_}}{a numeric vector}
    \item{\code{M.Baeomyces_roseus}}{a numeric vector}
    \item{\code{Empetrum_hermaphroditum_}}{a numeric vector}
    \item{\code{M.Paraleucobryum_longifolium}}{a numeric vector}
    \item{\code{Trientalis_europaea_}}{a numeric vector}
    \item{\code{Helictotrichon_versicolor_}}{a numeric vector}
    \item{\code{Hieracium_villosum_.ag._}}{a numeric vector}
    \item{\code{Polygonum_viviparum_}}{a numeric vector}
    \item{\code{Luzula_lutea}}{a numeric vector}
    \item{\code{Euphrasia_minima}}{a numeric vector}
    \item{\code{Euphrasia_montana_}}{a numeric vector}
    \item{\code{Gentiana_campestris_}}{a numeric vector}
    \item{\code{Gentiana_kochiana_}}{a numeric vector}
    \item{\code{Leucorchis_albida_}}{a numeric vector}
    \item{\code{Loiseleuria_procumbens}}{a numeric vector}
    \item{\code{Galium_anisophyllum_}}{a numeric vector}
    \item{\code{Phyteuma_hemisphaericum_}}{a numeric vector}
    \item{\code{Pinguicula_leptoceras_}}{a numeric vector}
    \item{\code{Lotus_alpinus_}}{a numeric vector}
    \item{\code{Poa_glauca}}{a numeric vector}
    \item{\code{Hieracium_auricula}}{a numeric vector}
    \item{\code{Sempervivum_arachnoideum}}{a numeric vector}
    \item{\code{Sedum_ochroleucum_}}{a numeric vector}
    \item{\code{Asplenium_septentrionale}}{a numeric vector}
    \item{\code{Allium_montanum_}}{a numeric vector}
    \item{\code{Ononis_spec.}}{a numeric vector}
    \item{\code{Ophrys_insectifera}}{a numeric vector}
    \item{\code{B.Sorbus_latifolia}}{a numeric vector}
    \item{\code{Rhinanthus_angustifolius}}{a numeric vector}
    \item{\code{Orchis_traunsteineri}}{a numeric vector}
    \item{\code{Pinguicula_vulgaris_}}{a numeric vector}
    \item{\code{Carex_hostiana}}{a numeric vector}
    \item{\code{Festuca_amethystina_}}{a numeric vector}
    \item{\code{Blackstonia_perfoliata}}{a numeric vector}
    \item{\code{Lathyrus_silvester}}{a numeric vector}
    \item{\code{Tragopogon_spec.}}{a numeric vector}
    \item{\code{Centaurea_jacea_}}{a numeric vector}
    \item{\code{Polygala_amara}}{a numeric vector}
    \item{\code{Gentiana_germanica}}{a numeric vector}
    \item{\code{S.Colutea_arborescens_}}{a numeric vector}
    \item{\code{Cirsium_tuberosum_}}{a numeric vector}
    \item{\code{Juncus_inflexus_}}{a numeric vector}
    \item{\code{Senecio_erucifolius_}}{a numeric vector}
    \item{\code{S.Rosa_eglanteria_}}{a numeric vector}
    \item{\code{B.Juniperus_communis}}{a numeric vector}
    \item{\code{Pinus_montana_}}{a numeric vector}
    \item{\code{Athamanta_cretensis_}}{a numeric vector}
    \item{\code{Scabiosa_lucida_}}{a numeric vector}
    \item{\code{Euphrasia_rostkoviana_}}{a numeric vector}
    \item{\code{Orchis_ustulata_}}{a numeric vector}
    \item{\code{Euphorbia_verrucosa_}}{a numeric vector}
    \item{\code{Globularia_elongata_}}{a numeric vector}
    \item{\code{Potentilla_heptaphylla}}{a numeric vector}
    \item{\code{Pulsatilla_vulgaris_}}{a numeric vector}
    \item{\code{Salvia_pratensis}}{a numeric vector}
    \item{\code{Scorzonera_austriaca}}{a numeric vector}
    \item{\code{Allium_sphaerocephalum}}{a numeric vector}
    \item{\code{Daphne_cneorum}}{a numeric vector}
    \item{\code{Festuca_vallesiaca}}{a numeric vector}
    \item{\code{Veronica_spicata}}{a numeric vector}
    \item{\code{Potentilla_verna}}{a numeric vector}
    \item{\code{M.Solorina_saccata}}{a numeric vector}
    \item{\code{Euphrasia_salisburgensis}}{a numeric vector}
    \item{\code{Hieracium_amplexicaule}}{a numeric vector}
    \item{\code{Campanula_bononiensis_}}{a numeric vector}
    \item{\code{Coronilla_minima}}{a numeric vector}
    \item{\code{Euphorbia_seguieriana_}}{a numeric vector}
    \item{\code{Odontites_viscosa_}}{a numeric vector}
    \item{\code{Stipa_capillata_}}{a numeric vector}
    \item{\code{M.Ditrichum_spec._}}{a numeric vector}
    \item{\code{Linum_tenuifolium_}}{a numeric vector}
    \item{\code{Pulsatilla_montana}}{a numeric vector}
    \item{\code{Astragalus_onobrychis_}}{a numeric vector}
    \item{\code{Fumana_ericoides}}{a numeric vector}
    \item{\code{Astragalus_monspessulanus_}}{a numeric vector}
    \item{\code{Rosa_eglanteria_}}{a numeric vector}
    \item{\code{Lathyrus_heterophyllus}}{a numeric vector}
    \item{\code{Coronilla_vaginalis_}}{a numeric vector}
    \item{\code{S.Polygala_chamaebuxus}}{a numeric vector}
    \item{\code{Globularia_nudicaulis_}}{a numeric vector}
    \item{\code{Thesium_rostratum_}}{a numeric vector}
    \item{\code{Rosa_micrantha}}{a numeric vector}
    \item{\code{Centaurea_rhaetica}}{a numeric vector}
    \item{\code{Satureja_nepetoides_}}{a numeric vector}
    \item{\code{Ononis_rotundifolia_}}{a numeric vector}
    \item{\code{Centaurea_uniflora}}{a numeric vector}
    \item{\code{Ononis_natrix_}}{a numeric vector}
    \item{\code{Hippophae_rhamnoides}}{a numeric vector}
    \item{\code{Medicago_lupulina_}}{a numeric vector}
    \item{\code{Oxytropis_pilosa}}{a numeric vector}
    \item{\code{Salix_elaeagnos_}}{a numeric vector}
    \item{\code{Salix_purpurea}}{a numeric vector}
    \item{\code{Artemisia_campestris}}{a numeric vector}
    \item{\code{Gypsophila_repens_}}{a numeric vector}
    \item{\code{Hieracium_peletierianum_}}{a numeric vector}
    \item{\code{S.Juniperus_sabina}}{a numeric vector}
    \item{\code{Minuartia_laricifolia_}}{a numeric vector}
    \item{\code{Astragalus_exscapus_}}{a numeric vector}
    \item{\code{Crepis_nemausensis}}{a numeric vector}
    \item{\code{Satureja_acinos_}}{a numeric vector}
    \item{\code{Sisymbrium_irio_}}{a numeric vector}
    \item{\code{Daphne_striata}}{a numeric vector}
    \item{\code{Hieracium_porrifolium_.ag.}}{a numeric vector}
    \item{\code{Helianthemum_alpestre_}}{a numeric vector}
    \item{\code{Biscutella_levigata_}}{a numeric vector}
    \item{\code{M.Distichium_capillaceum}}{a numeric vector}
    \item{\code{Aster_alpinus_}}{a numeric vector}
    \item{\code{Epilobium_fleischeri}}{a numeric vector}
    \item{\code{Onobrychis_montana}}{a numeric vector}
    \item{\code{Oxytropis_jacquinii_}}{a numeric vector}
    \item{\code{Alchemilla_conjuncta_.ag._}}{a numeric vector}
    \item{\code{Gentiana_clusii_}}{a numeric vector}
    \item{\code{Kernera_saxatilis_}}{a numeric vector}
    \item{\code{Minuartia_verna_}}{a numeric vector}
    \item{\code{Cerastium_alpinum_}}{a numeric vector}
    \item{\code{Rhododendron_hirsutum_}}{a numeric vector}
    \item{\code{S.Rhododendron_hirsutum_}}{a numeric vector}
    \item{\code{Pinguicula_alpina_}}{a numeric vector}
    \item{\code{Carex_ericetorum}}{a numeric vector}
    \item{\code{Saxifraga_aizoides}}{a numeric vector}
    \item{\code{Bartsia_alpina}}{a numeric vector}
    \item{\code{Carex_firma_}}{a numeric vector}
    \item{\code{Salix_retusa}}{a numeric vector}
    \item{\code{Salix_breviserrata}}{a numeric vector}
    \item{\code{Salix_hastata_}}{a numeric vector}
    \item{\code{Salix_reticulata}}{a numeric vector}
    \item{\code{Silene_acaulis}}{a numeric vector}
    \item{\code{Festuca_pumila}}{a numeric vector}
    \item{\code{Pedicularis_verticillata}}{a numeric vector}
    \item{\code{Veronica_aphylla}}{a numeric vector}
    \item{\code{M.Meesea_uliginosa}}{a numeric vector}
    \item{\code{Carex_capillaris}}{a numeric vector}
    \item{\code{Carex_rupestris_}}{a numeric vector}
    \item{\code{Poa_cenisia_}}{a numeric vector}
    \item{\code{Salix_waldsteiniana_}}{a numeric vector}
    \item{\code{Saxifraga_caesia}}{a numeric vector}
    \item{\code{M.Collema_spec._}}{a numeric vector}
    \item{\code{Equisetum_variegatum}}{a numeric vector}
    \item{\code{Polygala_alpestris}}{a numeric vector}
    \item{\code{Pulsatilla_vernalis_}}{a numeric vector}
    \item{\code{M.Mylia_anomala_}}{a numeric vector}
    \item{\code{M.Icmadophila_ericetorum}}{a numeric vector}
    \item{\code{Pulsatilla_alpina_}}{a numeric vector}
    \item{\code{Salix_aurita}}{a numeric vector}
    \item{\code{Andromeda_polifolia_}}{a numeric vector}
    \item{\code{Carex_rostrata}}{a numeric vector}
    \item{\code{M.Funaria_hygrometrica}}{a numeric vector}
    \item{\code{Drosera_rotundifolia}}{a numeric vector}
    \item{\code{Melampyrum_nemorosum}}{a numeric vector}
  }
}
\details{
Old taxonomy, not updated.
}
\source{
Ellenberg, H. & Kloetzli, F. 1972.
Waldgesellschaften und Waldstandorte der Schweiz. Mitt. Eidgenoess.
Forsch. anst. Wald Schnee Landsch. 48(4): 587--930.

Keller, W., Wohlgemuth, T., Kuhn, N., Schuetz, M. &
Wildi, O. 1998. Waldgesellschaften der Schweiz auf floristischer
Grundlage. Mitteilungen der Eidgenoessischen Forschungsanstalt fuer
Wald, Schnee und Landschaft (WSL) 73, Vol. 2.
}
\references{
Wildi, O. 2013. Data Analysis in Vegetation Ecology. 2nd ed. Wiley-Blackwell,
Chichester.
}
\examples{
summary(EKv)
}
\keyword{datasets}
