\name{mveg}
\alias{mveg}
\docType{data}
\title{
Ellenberg's 1956 meadow data
}
\description{
Ellenberg's 1956 meadow data. No site factors available.
}
\usage{data(mveg)}
\format{
A data frame with 25 observations on the 94 species, the variables (cover percentages). Species names are abbreviations.
}
\details{
No site factors available for this data frame.
}
\source{
Mueller-Dombois, D. & Ellenberg, H. 1974. Aims and
Methods of Vegetation Ecology. John Wiley & Sons, New York, Chichester,
Brisbane, Toronto.
}
\references{
Wildi, O. 2013. Data Analysis in Vegetation Ecology. 2nd ed. Wiley-Blackwell,
Chichester.
}
\examples{
summary(mveg)
}
\keyword{datasets}
