\name{sn59sit}
\alias{sn59sit}
\docType{data}
\title{
Time series from the Swiss National Park, 59 plots. Site data.
}
\description{
Time series from the Swiss National Park, 59 plots. Site data: Plot names and yr of sampling.
}
\usage{data(sn59sit)}
\format{
  A data frame with 751 observations on the following 2 variables.
  \describe{
    \item{\code{Plot.no}}{a factor with levels \code{AC1} \code{Ac9} \code{CiN1} \code{CiN2} \code{Cxs1} \code{De1} \code{De2} \code{Elyn} \code{Fe1} \code{Fe1a} \code{FN1} \code{FN2} \code{FN3} \code{Gri1} \code{Gri2} \code{Lav1} \code{Lav2} \code{Lav3} \code{Lav4} \code{Lav5} \code{Mi15} \code{Mi16} \code{Mi17} \code{Mi80} \code{MU21} \code{N1} \code{N10} \code{N11} \code{N14} \code{N2} \code{N3} \code{N4} \code{N5} \code{N6} \code{N7} \code{N8} \code{N9} \code{PF1} \code{PF13} \code{PF2} \code{Pin1} \code{PiN1} \code{Pin2} \code{PiN2} \code{Pin3} \code{Pin4} \code{Pin5} \code{S32} \code{S37} \code{S40} \code{S41} \code{S5} \code{Sesl} \code{Tr1} \code{Tr2} \code{Tr3} \code{Tr4} \code{Tr5} \code{Tr6}}
    \item{\code{Year}}{a numeric vector}
  }
}
\source{
Wildi, O. & Schuetz, M. 2000. Reconstruction of a
long-term recovery process from pasture to forest. Community Ecology 1: 25-32.
}
\references{
Wildi, O. 2013. Data Analysis in Vegetation Ecology. 2nd ed. Wiley-Blackwell,
Chichester.
}
\examples{
summary(sn59sit)
}
\keyword{datasets}
