\name{sveg}
\alias{sveg}
\docType{data}
\title{
Schlaenggli vegetation data 
}
\description{
Schlaenggli vegetation data. 63 plots arranged in a square grid. Site factors in data frame \code{\link{ssit}}.
}
\usage{data(sveg)}
\format{
A data frame with 63 observations on 119 species as variables. Species abundance measured on a 0 to 6 scale.
}
\source{
Wildi, O. 1977. Beschreibung exzentrischer Hochmoore
mit Hilfe quantitativer Methoden. Veroeff. Geobot. Inst. ETH, Stiftung Ruebel 60: 128S.
}
\references{
Wildi, O. 2017. Data Analysis in Vegetation Ecology. 3rd ed. CABI, Oxfordshire, Boston.
}
\examples{
summary(sveg)
}
\keyword{datasets}
