\name{tveg}
\alias{tveg}
\docType{data}
\title{
Time series from the Swiss National Park, Plot Tr6. Vegetation data.
}
\description{
Time series from the Swiss National Park, Plot Tr6. Vegetation data.
}
\usage{data(tveg)}
\format{
  A data frame with 16 observations on the following 6 variables.
  \describe{
    \item{\code{Aconitum}}{a numeric vector}
    \item{\code{Deschampsia}}{a numeric vector}
    \item{\code{Trisetum}}{a numeric vector}
    \item{\code{Festuca}}{a numeric vector}
    \item{\code{Carex}}{a numeric vector}
    \item{\code{Pinus}}{a numeric vector}
  }
}
\details{
A subset of data frame sn59veg
}
\source{
Wildi, O. & Schuetz, M. 2000. Reconstruction of a
long-term recovery process from pasture to forest. Community Ecology 1: 25-32.
}
\references{
Wildi, O. 2013. Data Analysis in Vegetation Ecology. 2nd ed. Wiley-Blackwell,
Chichester.
}
\examples{
summary(tveg)
}
\keyword{datasets}
