\name{wetveg}
\alias{wetveg}
\docType{data}
\title{
Random sample of Swiss wetland vegetation, vegetation information.
}
\description{
Random sample of Swiss wetland vegetation, vegetation information.
}
\usage{data(wetveg)}
\format{
  A data frame with 1500 observations on the following 1164 variables.
  \describe{
    \item{\code{Abies_alba_Mill.}}{a numeric vector}
    \item{\code{Acer_campestre}}{a numeric vector}
    \item{\code{Acer_platanoides}}{a numeric vector}
    \item{\code{Acer_pseudoplatanus}}{a numeric vector}
    \item{\code{Achillea_atrata}}{a numeric vector}
    \item{\code{Achillea_millefolium}}{a numeric vector}
    \item{\code{Acinos_alpinus}}{a numeric vector}
    \item{\code{Aconitum_vulparia}}{a numeric vector}
    \item{\code{Aconitum_napellus}}{a numeric vector}
    \item{\code{Aconitum_variegatum_subsp._paniculatum}}{a numeric vector}
    \item{\code{Actaea_spicata}}{a numeric vector}
    \item{\code{Adenostyles_alliariae}}{a numeric vector}
    \item{\code{Adenostyles_glabra}}{a numeric vector}
    \item{\code{Aegopodium_podagraria}}{a numeric vector}
    \item{\code{Agropyron_caninum}}{a numeric vector}
    \item{\code{Agropyron_repens}}{a numeric vector}
    \item{\code{Agrostis_alpina}}{a numeric vector}
    \item{\code{Agrostis_canina}}{a numeric vector}
    \item{\code{Agrostis_capillaris}}{a numeric vector}
    \item{\code{Agrostis_rupestris}}{a numeric vector}
    \item{\code{Agrostis_schraderiana}}{a numeric vector}
    \item{\code{Agrostis_stolonifera}}{a numeric vector}
    \item{\code{Ajuga_pyramidalis}}{a numeric vector}
    \item{\code{Ajuga_reptans}}{a numeric vector}
    \item{\code{Alchemilla_alpina}}{a numeric vector}
    \item{\code{Alchemilla_conjuncta}}{a numeric vector}
    \item{\code{Alchemilla_coriacea}}{a numeric vector}
    \item{\code{Alchemilla_decumbens}}{a numeric vector}
    \item{\code{Alchemilla_fissa}}{a numeric vector}
    \item{\code{Alchemilla_hybrida}}{a numeric vector}
    \item{\code{Alchemilla_xanthochlora}}{a numeric vector}
    \item{\code{Alisma_plantago.aquatica}}{a numeric vector}
    \item{\code{Alliaria_petiolata}}{a numeric vector}
    \item{\code{Allium_carinatum}}{a numeric vector}
    \item{\code{Allium_schoenoprasum}}{a numeric vector}
    \item{\code{Allium_sphaerocephalon}}{a numeric vector}
    \item{\code{Allium_ursinum}}{a numeric vector}
    \item{\code{Alnus_glutinosa}}{a numeric vector}
    \item{\code{Alnus_incana}}{a numeric vector}
    \item{\code{Alnus_viridis}}{a numeric vector}
    \item{\code{Alopecurus_geniculatus}}{a numeric vector}
    \item{\code{Alopecurus_pratensis}}{a numeric vector}
    \item{\code{Amblyodon_dealbatus}}{a numeric vector}
    \item{\code{Amblystegium_humile}}{a numeric vector}
    \item{\code{Amblystegium_riparium}}{a numeric vector}
    \item{\code{Amblystegium_serpens}}{a numeric vector}
    \item{\code{Amblystegium_varium}}{a numeric vector}
    \item{\code{Anastrophyllum_minutum}}{a numeric vector}
    \item{\code{Andromeda_polifolia}}{a numeric vector}
    \item{\code{Androsace_chamaejasme}}{a numeric vector}
    \item{\code{Anemone_narcissiflora}}{a numeric vector}
    \item{\code{Anemone_nemorosa}}{a numeric vector}
    \item{\code{Aneura_pinguis}}{a numeric vector}
    \item{\code{Angelica_sylvestris}}{a numeric vector}
    \item{\code{Antennaria_dioica}}{a numeric vector}
    \item{\code{Anthelia_julacea}}{a numeric vector}
    \item{\code{Anthoxanthum_odoratum}}{a numeric vector}
    \item{\code{Anthriscus_sylvestris}}{a numeric vector}
    \item{\code{Anthyllis_vulneraria}}{a numeric vector}
    \item{\code{Aposeris_foetida}}{a numeric vector}
    \item{\code{Aquilegia_atrata}}{a numeric vector}
    \item{\code{Arabis_alpina}}{a numeric vector}
    \item{\code{Arabis_hirsuta}}{a numeric vector}
    \item{\code{Arabis_subcoriacea}}{a numeric vector}
    \item{\code{Arctostaphylos_alpina}}{a numeric vector}
    \item{\code{Arctostaphylos_uva.ursi}}{a numeric vector}
    \item{\code{Arenaria_serpyllifolia}}{a numeric vector}
    \item{\code{Arnica_montana}}{a numeric vector}
    \item{\code{Arrhenatherum_elatius}}{a numeric vector}
    \item{\code{Artemisia_vulgaris}}{a numeric vector}
    \item{\code{Asarum_europaeum}}{a numeric vector}
    \item{\code{Asparagus_officinalis}}{a numeric vector}
    \item{\code{Asperula_cynanchica}}{a numeric vector}
    \item{\code{Asplenium_ruta.muraria}}{a numeric vector}
    \item{\code{Asplenium_viride}}{a numeric vector}
    \item{\code{Aster_bellidiastrum}}{a numeric vector}
    \item{\code{Astragalus_alpinus}}{a numeric vector}
    \item{\code{Astragalus_frigidus}}{a numeric vector}
    \item{\code{Astragalus_penduliflorus}}{a numeric vector}
    \item{\code{Astrantia_major}}{a numeric vector}
    \item{\code{Astrantia_minor}}{a numeric vector}
    \item{\code{Athyrium_distentifolium}}{a numeric vector}
    \item{\code{Athyrium_filix.femina}}{a numeric vector}
    \item{\code{Atrichum_undulatum}}{a numeric vector}
    \item{\code{Aulacomnium_androgynum}}{a numeric vector}
    \item{\code{Aulacomnium_palustre}}{a numeric vector}
    \item{\code{Avenella_flexuosa}}{a numeric vector}
    \item{\code{Barbarea_vulgaris}}{a numeric vector}
    \item{\code{Barbilophozia_attenuata}}{a numeric vector}
    \item{\code{Barbilophozia_barbata}}{a numeric vector}
    \item{\code{Barbilophozia_floerkei}}{a numeric vector}
    \item{\code{Barbilophozia_lycopodioides}}{a numeric vector}
    \item{\code{Barbilophozia_kunzeana}}{a numeric vector}
    \item{\code{Barbula_fallax}}{a numeric vector}
    \item{\code{Barbula_gigantea}}{a numeric vector}
    \item{\code{Barbula_hornschuchiana}}{a numeric vector}
    \item{\code{Barbula_reflexa}}{a numeric vector}
    \item{\code{Barbula_unguiculata}}{a numeric vector}
    \item{\code{Bartramia_ithyphylla}}{a numeric vector}
    \item{\code{Bartsia_alpina}}{a numeric vector}
    \item{\code{Bazzania_tricrenata}}{a numeric vector}
    \item{\code{Bazzania_trilobata}}{a numeric vector}
    \item{\code{Bellis_perennis}}{a numeric vector}
    \item{\code{Betula_nana}}{a numeric vector}
    \item{\code{Betula_pendula}}{a numeric vector}
    \item{\code{Betula_pubescens}}{a numeric vector}
    \item{\code{Biscutella_laevigata}}{a numeric vector}
    \item{\code{Blechnum_spicant}}{a numeric vector}
    \item{\code{Blepharostoma_trichophyllum}}{a numeric vector}
    \item{\code{Blindia_acuta}}{a numeric vector}
    \item{\code{Blysmus_compressus}}{a numeric vector}
    \item{\code{Botrychium_lunaria}}{a numeric vector}
    \item{\code{Brachypodium_pinnatum}}{a numeric vector}
    \item{\code{Brachypodium_sylvaticum}}{a numeric vector}
    \item{\code{Brachythecium_albicans}}{a numeric vector}
    \item{\code{Brachythecium_campestre}}{a numeric vector}
    \item{\code{Brachythecium_erythrorrhizon}}{a numeric vector}
    \item{\code{Brachythecium_glareosum}}{a numeric vector}
    \item{\code{Brachythecium_mildeanum}}{a numeric vector}
    \item{\code{Brachythecium_plumosum}}{a numeric vector}
    \item{\code{Brachythecium_populeum}}{a numeric vector}
    \item{\code{Brachythecium_reflexum}}{a numeric vector}
    \item{\code{Brachythecium_rivulare}}{a numeric vector}
    \item{\code{Brachythecium_rutabulum}}{a numeric vector}
    \item{\code{Brachythecium_salebrosum}}{a numeric vector}
    \item{\code{Brachythecium_starkei}}{a numeric vector}
    \item{\code{Brachythecium_turgidum}}{a numeric vector}
    \item{\code{Brachythecium_velutinum}}{a numeric vector}
    \item{\code{Brassica_rapa}}{a numeric vector}
    \item{\code{Briza_media}}{a numeric vector}
    \item{\code{Bromus_arvensis}}{a numeric vector}
    \item{\code{Bromus_erectus}}{a numeric vector}
    \item{\code{Bromus_hordeaceus}}{a numeric vector}
    \item{\code{Bromus_racemosus}}{a numeric vector}
    \item{\code{Bryonia_dioica}}{a numeric vector}
    \item{\code{Bryum_argenteum}}{a numeric vector}
    \item{\code{Bryum_capillare}}{a numeric vector}
    \item{\code{Bryum_muehlenbeckii}}{a numeric vector}
    \item{\code{Bryum_pallens}}{a numeric vector}
    \item{\code{Bryum_pallescens}}{a numeric vector}
    \item{\code{Bryum_pseudotriquetrum}}{a numeric vector}
    \item{\code{Bryum_turbinatum}}{a numeric vector}
    \item{\code{Bryum_weigelii}}{a numeric vector}
    \item{\code{Buphthalmum_salicifolium}}{a numeric vector}
    \item{\code{Bupleurum_stellatum}}{a numeric vector}
    \item{\code{Calamagrostis_canescens}}{a numeric vector}
    \item{\code{Calamagrostis_epigejos}}{a numeric vector}
    \item{\code{Calamagrostis_varia}}{a numeric vector}
    \item{\code{Calamagrostis_villosa}}{a numeric vector}
    \item{\code{Calliergon_cordifolium}}{a numeric vector}
    \item{\code{Calliergon_giganteum}}{a numeric vector}
    \item{\code{Calliergon_sarmentosum}}{a numeric vector}
    \item{\code{Calliergon_stramineum}}{a numeric vector}
    \item{\code{Calliergon_trifarium}}{a numeric vector}
    \item{\code{Calliergonella_cuspidata}}{a numeric vector}
    \item{\code{Callitriche_hamulata}}{a numeric vector}
    \item{\code{Callitriche_palustris}}{a numeric vector}
    \item{\code{Calluna_vulgaris}}{a numeric vector}
    \item{\code{Caltha_palustris}}{a numeric vector}
    \item{\code{Calycocorsus_stipitatus}}{a numeric vector}
    \item{\code{Calypogeia_fissa}}{a numeric vector}
    \item{\code{Calypogeia_integristipula}}{a numeric vector}
    \item{\code{Calypogeia_muelleriana}}{a numeric vector}
    \item{\code{Calypogeia_neesiana}}{a numeric vector}
    \item{\code{Calypogeia_sphagnicola}}{a numeric vector}
    \item{\code{Calystegia_sepium}}{a numeric vector}
    \item{\code{Campanula_barbata}}{a numeric vector}
    \item{\code{Campanula_cochleariifolia}}{a numeric vector}
    \item{\code{Campanula_glomerata}}{a numeric vector}
    \item{\code{Campanula_patula_subsp._patula}}{a numeric vector}
    \item{\code{Campanula_rhomboidalis}}{a numeric vector}
    \item{\code{Campanula_rotundifolia}}{a numeric vector}
    \item{\code{Campylium_calcareum}}{a numeric vector}
    \item{\code{Campylium_chrysophyllum}}{a numeric vector}
    \item{\code{Campylium_elodes}}{a numeric vector}
    \item{\code{Campylium_polygamum}}{a numeric vector}
    \item{\code{Campylium_stellatum}}{a numeric vector}
    \item{\code{Campylopus_flexuosus}}{a numeric vector}
    \item{\code{Campylopus_fragilis}}{a numeric vector}
    \item{\code{Campylopus_introflexus}}{a numeric vector}
    \item{\code{Campylopus_pyriformis}}{a numeric vector}
    \item{\code{Campylopus_schwarzii}}{a numeric vector}
    \item{\code{Campylopus_subulatus}}{a numeric vector}
    \item{\code{Capsella_bursa.pastoris}}{a numeric vector}
    \item{\code{Cardamine_amara}}{a numeric vector}
    \item{\code{Cardamine_flexuosa}}{a numeric vector}
    \item{\code{Cardamine_hirsuta}}{a numeric vector}
    \item{\code{Cardamine_pentaphyllos}}{a numeric vector}
    \item{\code{Cardamine_pratensis}}{a numeric vector}
    \item{\code{Cardamine_resedifolia}}{a numeric vector}
    \item{\code{Carduus_defloratus}}{a numeric vector}
    \item{\code{Carduus_nutans}}{a numeric vector}
    \item{\code{Carduus_personata}}{a numeric vector}
    \item{\code{Carex_acuta}}{a numeric vector}
    \item{\code{Carex_acutiformis}}{a numeric vector}
    \item{\code{Carex_appropinquata}}{a numeric vector}
    \item{\code{Carex_atrata}}{a numeric vector}
    \item{\code{Carex_atrofusca}}{a numeric vector}
    \item{\code{Carex_bicolor}}{a numeric vector}
    \item{\code{Carex_brunnescens}}{a numeric vector}
    \item{\code{Carex_buxbaumii}}{a numeric vector}
    \item{\code{Carex_canescens}}{a numeric vector}
    \item{\code{Carex_capillaris}}{a numeric vector}
    \item{\code{Carex_caryophyllea}}{a numeric vector}
    \item{\code{Carex_curvula_subsp._curvula}}{a numeric vector}
    \item{\code{Carex_curvula_subsp._rosae}}{a numeric vector}
    \item{\code{Carex_davalliana}}{a numeric vector}
    \item{\code{Carex_diandra}}{a numeric vector}
    \item{\code{Carex_digitata}}{a numeric vector}
    \item{\code{Carex_dioica}}{a numeric vector}
    \item{\code{Carex_distans}}{a numeric vector}
    \item{\code{Carex_disticha}}{a numeric vector}
    \item{\code{Carex_echinata}}{a numeric vector}
    \item{\code{Carex_elata}}{a numeric vector}
    \item{\code{Carex_elata.1}}{a numeric vector}
    \item{\code{Carex_elongata}}{a numeric vector}
    \item{\code{Carex_ericetorum}}{a numeric vector}
    \item{\code{Carex_ferruginea}}{a numeric vector}
    \item{\code{Carex_flacca}}{a numeric vector}
    \item{\code{Carex_flava}}{a numeric vector}
    \item{\code{Carex_frigida}}{a numeric vector}
    \item{\code{Carex_hirta}}{a numeric vector}
    \item{\code{Carex_hostiana}}{a numeric vector}
    \item{\code{Carex_lasiocarpa}}{a numeric vector}
    \item{\code{Carex_leporina}}{a numeric vector}
    \item{\code{Carex_limosa}}{a numeric vector}
    \item{\code{Carex_microglochin}}{a numeric vector}
    \item{\code{Carex_montana}}{a numeric vector}
    \item{\code{Carex_muricata}}{a numeric vector}
    \item{\code{Carex_nigra}}{a numeric vector}
    \item{\code{Carex_ornithopoda}}{a numeric vector}
    \item{\code{Carex_otrubae}}{a numeric vector}
    \item{\code{Carex_pallescens}}{a numeric vector}
    \item{\code{Carex_panicea}}{a numeric vector}
    \item{\code{Carex_paniculata}}{a numeric vector}
    \item{\code{Carex_pauciflora}}{a numeric vector}
    \item{\code{Carex_paupercula}}{a numeric vector}
    \item{\code{Carex_pendula}}{a numeric vector}
    \item{\code{Carex_pilulifera}}{a numeric vector}
    \item{\code{Carex_pulicaris}}{a numeric vector}
    \item{\code{Carex_remota}}{a numeric vector}
    \item{\code{Carex_riparia}}{a numeric vector}
    \item{\code{Carex_rostrata}}{a numeric vector}
    \item{\code{Carex_sempervirens}}{a numeric vector}
    \item{\code{Carex_sylvatica}}{a numeric vector}
    \item{\code{Carex_tomentosa}}{a numeric vector}
    \item{\code{Carex_umbrosa}}{a numeric vector}
    \item{\code{Carex_vaginata}}{a numeric vector}
    \item{\code{Carex_vesicaria}}{a numeric vector}
    \item{\code{Carlina_acaulis}}{a numeric vector}
    \item{\code{Carlina_vulgaris}}{a numeric vector}
    \item{\code{Carum_carvi}}{a numeric vector}
    \item{\code{Castanea_sativa}}{a numeric vector}
    \item{\code{Catoscopium_nigritum}}{a numeric vector}
    \item{\code{Centaurea_jacea}}{a numeric vector}
    \item{\code{Centaurea_montana}}{a numeric vector}
    \item{\code{Centaurea_nervosa}}{a numeric vector}
    \item{\code{Centaurea_scabiosa}}{a numeric vector}
    \item{\code{Cephalanthera_longifolia}}{a numeric vector}
    \item{\code{Cephalanthera_rubra}}{a numeric vector}
    \item{\code{Cephalozia_bicuspidata}}{a numeric vector}
    \item{\code{Cephalozia_connivens}}{a numeric vector}
    \item{\code{Cephalozia_loitlesbergeri}}{a numeric vector}
    \item{\code{Cerastium_alpinum}}{a numeric vector}
    \item{\code{Cerastium_arvense}}{a numeric vector}
    \item{\code{Cerastium_brachypetalum}}{a numeric vector}
    \item{\code{Cerastium_fontanum_subsp._fontanum}}{a numeric vector}
    \item{\code{Cerastium_fontanum_subsp._vulgare}}{a numeric vector}
    \item{\code{Ceratodon_purpureus}}{a numeric vector}
    \item{\code{Chaerophyllum_aureum}}{a numeric vector}
    \item{\code{Chaerophyllum_hirsutum}}{a numeric vector}
    \item{\code{Chamorchis_alpina}}{a numeric vector}
    \item{\code{Chenopodium_bonus.henricus}}{a numeric vector}
    \item{\code{Chiloscyphus_polyanthos}}{a numeric vector}
    \item{\code{Leucanthemum_vulgare_aggr.}}{a numeric vector}
    \item{\code{Chrysosplenium_alternifolium}}{a numeric vector}
    \item{\code{Cicerbita_alpina}}{a numeric vector}
    \item{\code{Cinclidium_stygium}}{a numeric vector}
    \item{\code{Circaea_alpina}}{a numeric vector}
    \item{\code{Circaea_lutetiana}}{a numeric vector}
    \item{\code{Cirriphyllum_cirrosum}}{a numeric vector}
    \item{\code{Cirriphyllum_piliferum}}{a numeric vector}
    \item{\code{Cirsium_acaule}}{a numeric vector}
    \item{\code{Cirsium_arvense}}{a numeric vector}
    \item{\code{Cirsium_eriophorum_subsp._eriophorum}}{a numeric vector}
    \item{\code{Cirsium_helenioides}}{a numeric vector}
    \item{\code{Cirsium_oleraceum}}{a numeric vector}
    \item{\code{Cirsium_palustre}}{a numeric vector}
    \item{\code{Cirsium_rivulare}}{a numeric vector}
    \item{\code{Cirsium_spinosissimum}}{a numeric vector}
    \item{\code{Cirsium_vulgare}}{a numeric vector}
    \item{\code{Cladium_mariscus}}{a numeric vector}
    \item{\code{Climacium_dendroides}}{a numeric vector}
    \item{\code{Clinopodium_vulgare}}{a numeric vector}
    \item{\code{Coeloglossum_viride}}{a numeric vector}
    \item{\code{Colchicum_autumnale}}{a numeric vector}
    \item{\code{Conocephalum_conicum}}{a numeric vector}
    \item{\code{Convolvulus_arvensis}}{a numeric vector}
    \item{\code{Conyza_canadensis}}{a numeric vector}
    \item{\code{Cornus_sanguinea}}{a numeric vector}
    \item{\code{Cornus_sericea}}{a numeric vector}
    \item{\code{Corylus_avellana}}{a numeric vector}
    \item{\code{Cotoneaster_divaricata}}{a numeric vector}
    \item{\code{Cotoneaster_integerrimus}}{a numeric vector}
    \item{\code{Crataegus_monogyna}}{a numeric vector}
    \item{\code{Cratoneuron_commutatum}}{a numeric vector}
    \item{\code{Cratoneuron_decipiens}}{a numeric vector}
    \item{\code{Cratoneuron_filicinum}}{a numeric vector}
    \item{\code{Crepis_aurea}}{a numeric vector}
    \item{\code{Crepis_biennis}}{a numeric vector}
    \item{\code{Crepis_capillaris}}{a numeric vector}
    \item{\code{Crepis_conyzifolia}}{a numeric vector}
    \item{\code{Crepis_mollis}}{a numeric vector}
    \item{\code{Crepis_paludosa}}{a numeric vector}
    \item{\code{Crepis_pyrenaica}}{a numeric vector}
    \item{\code{Crepis_vesicaria_subsp._taraxacifolia}}{a numeric vector}
    \item{\code{Crocus_albiflorus}}{a numeric vector}
    \item{\code{Cruciata_laevipes}}{a numeric vector}
    \item{\code{Ctenidium_molluscum}}{a numeric vector}
    \item{\code{Cynosurus_cristatus}}{a numeric vector}
    \item{\code{Cystopteris_fragilis}}{a numeric vector}
    \item{\code{Dactylis_glomerata}}{a numeric vector}
    \item{\code{Dactylorhiza_cruenta}}{a numeric vector}
    \item{\code{Dactylorhiza_fistulosa}}{a numeric vector}
    \item{\code{Dactylorhiza_maculata}}{a numeric vector}
    \item{\code{Dactylorhiza_incarnata}}{a numeric vector}
    \item{\code{Dactylorhiza_traunsteineri}}{a numeric vector}
    \item{\code{Danthonia_decumbens}}{a numeric vector}
    \item{\code{Daphne_mezereum}}{a numeric vector}
    \item{\code{Daphne_striata}}{a numeric vector}
    \item{\code{Daucus_carota}}{a numeric vector}
    \item{\code{Deschampsia_cespitosa}}{a numeric vector}
    \item{\code{Desmatodon_latifolius}}{a numeric vector}
    \item{\code{Dianthus_barbatus}}{a numeric vector}
    \item{\code{Dianthus_carthusianorum}}{a numeric vector}
    \item{\code{Dianthus_superbus}}{a numeric vector}
    \item{\code{Dichodontium_pellucidum}}{a numeric vector}
    \item{\code{Dicranella_cerviculata}}{a numeric vector}
    \item{\code{Dicranella_crispa}}{a numeric vector}
    \item{\code{Dicranella_grevilleana}}{a numeric vector}
    \item{\code{Dicranella_heteromalla}}{a numeric vector}
    \item{\code{Dicranella_palustris}}{a numeric vector}
    \item{\code{Dicranella_rufescens}}{a numeric vector}
    \item{\code{Dicranodontium_denudatum}}{a numeric vector}
    \item{\code{Dicranoweisia_crispula}}{a numeric vector}
    \item{\code{Dicranum_bergeri}}{a numeric vector}
    \item{\code{Dicranum_scoparium}}{a numeric vector}
    \item{\code{Dicranum_fuscescens}}{a numeric vector}
    \item{\code{Dicranum_montanum}}{a numeric vector}
    \item{\code{Dicranum_muehlenbeckii}}{a numeric vector}
    \item{\code{Dicranum_polysetum}}{a numeric vector}
    \item{\code{Dicranum_spurium}}{a numeric vector}
    \item{\code{Diphasiastrum_alpinum}}{a numeric vector}
    \item{\code{Diplophyllum_taxifolium}}{a numeric vector}
    \item{\code{Dipsacus_fullonum}}{a numeric vector}
    \item{\code{Distichium_capillaceum}}{a numeric vector}
    \item{\code{Ditrichum_flexicaule}}{a numeric vector}
    \item{\code{Ditrichum_heteromallum}}{a numeric vector}
    \item{\code{Draba_siliquosa}}{a numeric vector}
    \item{\code{Drepanocladus_aduncus}}{a numeric vector}
    \item{\code{Drepanocladus_revolvens}}{a numeric vector}
    \item{\code{Drepanocladus_exannulatus}}{a numeric vector}
    \item{\code{Drepanocladus_fluitans}}{a numeric vector}
    \item{\code{Drepanocladus_lycopodioides}}{a numeric vector}
    \item{\code{Drepanocladus_uncinatus}}{a numeric vector}
    \item{\code{Drepanocladus_vernicosus}}{a numeric vector}
    \item{\code{Drosera_intermedia}}{a numeric vector}
    \item{\code{Drosera_longifolia}}{a numeric vector}
    \item{\code{Drosera_rotundifolia}}{a numeric vector}
    \item{\code{Drosera_x}}{a numeric vector}
    \item{\code{Dryas_octopetala}}{a numeric vector}
    \item{\code{Dryopteris_affinis}}{a numeric vector}
    \item{\code{Dryopteris_carthusiana}}{a numeric vector}
    \item{\code{Dryopteris_filix.mas}}{a numeric vector}
    \item{\code{Dryopteris_remota}}{a numeric vector}
    \item{\code{Dryptodon_patens}}{a numeric vector}
    \item{\code{Echium_vulgare}}{a numeric vector}
    \item{\code{Eleocharis_palustris}}{a numeric vector}
    \item{\code{Eleocharis_quinqueflora}}{a numeric vector}
    \item{\code{Elyna_myosuroides}}{a numeric vector}
    \item{\code{Empetrum_nigrum}}{a numeric vector}
    \item{\code{Encalypta_streptocarpa}}{a numeric vector}
    \item{\code{Entodon_concinnus}}{a numeric vector}
    \item{\code{Epilobium_alpestre}}{a numeric vector}
    \item{\code{Epilobium_alsinifolium}}{a numeric vector}
    \item{\code{Epilobium_anagallidifolium}}{a numeric vector}
    \item{\code{Epilobium_angustifolium}}{a numeric vector}
    \item{\code{Epilobium_ciliatum}}{a numeric vector}
    \item{\code{Epilobium_collinum}}{a numeric vector}
    \item{\code{Epilobium_hirsutum}}{a numeric vector}
    \item{\code{Epilobium_montanum}}{a numeric vector}
    \item{\code{Epilobium_nutans}}{a numeric vector}
    \item{\code{Epilobium_obscurum}}{a numeric vector}
    \item{\code{Epilobium_palustre}}{a numeric vector}
    \item{\code{Epilobium_parviflorum}}{a numeric vector}
    \item{\code{Epilobium_roseum}}{a numeric vector}
    \item{\code{Epilobium_tetragonum}}{a numeric vector}
    \item{\code{Epipactis_helleborine}}{a numeric vector}
    \item{\code{Epipactis_palustris}}{a numeric vector}
    \item{\code{Epipactis_purpurata}}{a numeric vector}
    \item{\code{Equisetum_arvense}}{a numeric vector}
    \item{\code{Equisetum_fluviatile}}{a numeric vector}
    \item{\code{Equisetum_hyemale}}{a numeric vector}
    \item{\code{Equisetum_palustre}}{a numeric vector}
    \item{\code{Equisetum_sylvaticum}}{a numeric vector}
    \item{\code{Equisetum_telmateia}}{a numeric vector}
    \item{\code{Equisetum_variegatum}}{a numeric vector}
    \item{\code{Erica_carnea}}{a numeric vector}
    \item{\code{Erigeron_acer_subsp._angulosus}}{a numeric vector}
    \item{\code{Erigeron_alpinus}}{a numeric vector}
    \item{\code{Erigeron_annuus}}{a numeric vector}
    \item{\code{Eriophorum_angustifolium}}{a numeric vector}
    \item{\code{Eriophorum_latifolium}}{a numeric vector}
    \item{\code{Eriophorum_scheuchzeri}}{a numeric vector}
    \item{\code{Eriophorum_vaginatum}}{a numeric vector}
    \item{\code{Euonymus_europaeus}}{a numeric vector}
    \item{\code{Eupatorium_cannabinum}}{a numeric vector}
    \item{\code{Euphorbia_cyparissias}}{a numeric vector}
    \item{\code{Euphorbia_dulcis}}{a numeric vector}
    \item{\code{Euphorbia_helioscopia}}{a numeric vector}
    \item{\code{Euphrasia_alpina}}{a numeric vector}
    \item{\code{Euphrasia_hirtella}}{a numeric vector}
    \item{\code{Euphrasia_minima}}{a numeric vector}
    \item{\code{Euphrasia_picta}}{a numeric vector}
    \item{\code{Euphrasia_pulchella}}{a numeric vector}
    \item{\code{Euphrasia_rostkoviana_subsp._montana}}{a numeric vector}
    \item{\code{Euphrasia_rostkoviana_subsp._rostkoviana}}{a numeric vector}
    \item{\code{Eurhynchium_hians}}{a numeric vector}
    \item{\code{Eurhynchium_praelongum}}{a numeric vector}
    \item{\code{Eurhynchium_speciosum}}{a numeric vector}
    \item{\code{Eurhynchium_striatum}}{a numeric vector}
    \item{\code{Fagus_sylvatica}}{a numeric vector}
    \item{\code{Festuca_arundinacea_subsp._arundinacea}}{a numeric vector}
    \item{\code{Festuca_gigantea}}{a numeric vector}
    \item{\code{Festuca_norica}}{a numeric vector}
    \item{\code{Festuca_ovina}}{a numeric vector}
    \item{\code{Festuca_pratensis}}{a numeric vector}
    \item{\code{Festuca_pulchella}}{a numeric vector}
    \item{\code{Festuca_quadriflora}}{a numeric vector}
    \item{\code{Festuca_rubra}}{a numeric vector}
    \item{\code{Festuca_rupicaprina}}{a numeric vector}
    \item{\code{Festuca_varia}}{a numeric vector}
    \item{\code{Festuca_violacea}}{a numeric vector}
    \item{\code{Filipendula_ulmaria}}{a numeric vector}
    \item{\code{Fissidens_adianthoides}}{a numeric vector}
    \item{\code{Fissidens_osmundoides}}{a numeric vector}
    \item{\code{Fissidens_taxifolius}}{a numeric vector}
    \item{\code{Fontinalis_antipyretica}}{a numeric vector}
    \item{\code{Fragaria_vesca}}{a numeric vector}
    \item{\code{Frangula_alnus}}{a numeric vector}
    \item{\code{Fraxinus_excelsior}}{a numeric vector}
    \item{\code{Funaria_hygrometrica}}{a numeric vector}
    \item{\code{Galeopsis_tetrahit}}{a numeric vector}
    \item{\code{Galium_anisophyllon}}{a numeric vector}
    \item{\code{Galium_aparine}}{a numeric vector}
    \item{\code{Galium_boreale}}{a numeric vector}
    \item{\code{Galium_elongatum}}{a numeric vector}
    \item{\code{Galium_mollugo}}{a numeric vector}
    \item{\code{Galium_odoratum}}{a numeric vector}
    \item{\code{Galium_palustre}}{a numeric vector}
    \item{\code{Galium_pumilum}}{a numeric vector}
    \item{\code{Galium_rotundifolium}}{a numeric vector}
    \item{\code{Galium_saxatile}}{a numeric vector}
    \item{\code{Galium_spurium}}{a numeric vector}
    \item{\code{Galium_uliginosum}}{a numeric vector}
    \item{\code{Galium_verum}}{a numeric vector}
    \item{\code{Genista_sagittalis}}{a numeric vector}
    \item{\code{Genista_tinctoria}}{a numeric vector}
    \item{\code{Gentiana_acaulis}}{a numeric vector}
    \item{\code{Gentiana_asclepiadea}}{a numeric vector}
    \item{\code{Gentiana_bavarica}}{a numeric vector}
    \item{\code{Gentiana_campestris}}{a numeric vector}
    \item{\code{Gentiana_ciliata}}{a numeric vector}
    \item{\code{Gentiana_clusii}}{a numeric vector}
    \item{\code{Gentiana_germanica}}{a numeric vector}
    \item{\code{Gentiana_lutea}}{a numeric vector}
    \item{\code{Gentiana_nivalis}}{a numeric vector}
    \item{\code{Gentiana_pneumonanthe}}{a numeric vector}
    \item{\code{Gentiana_punctata}}{a numeric vector}
    \item{\code{Gentiana_purpurea}}{a numeric vector}
    \item{\code{Gentiana_tenella}}{a numeric vector}
    \item{\code{Gentiana_utriculosa}}{a numeric vector}
    \item{\code{Gentiana_verna}}{a numeric vector}
    \item{\code{Geranium_columbinum}}{a numeric vector}
    \item{\code{Geranium_palustre}}{a numeric vector}
    \item{\code{Geranium_pusillum}}{a numeric vector}
    \item{\code{Geranium_pyrenaicum}}{a numeric vector}
    \item{\code{Geranium_robertianum}}{a numeric vector}
    \item{\code{Geranium_sylvaticum}}{a numeric vector}
    \item{\code{Geum_montanum}}{a numeric vector}
    \item{\code{Geum_rivale}}{a numeric vector}
    \item{\code{Geum_urbanum}}{a numeric vector}
    \item{\code{Glechoma_hederacea}}{a numeric vector}
    \item{\code{Globularia_cordifolia}}{a numeric vector}
    \item{\code{Globularia_nudicaulis}}{a numeric vector}
    \item{\code{Glyceria_fluitans}}{a numeric vector}
    \item{\code{Gnaphalium_norvegicum}}{a numeric vector}
    \item{\code{Gnaphalium_supinum}}{a numeric vector}
    \item{\code{Gnaphalium_sylvaticum}}{a numeric vector}
    \item{\code{Gymnadenia_conopsea}}{a numeric vector}
    \item{\code{Gymnadenia_odoratissima}}{a numeric vector}
    \item{\code{Gymnocarpium_dryopteris}}{a numeric vector}
    \item{\code{Gymnocarpium_robertianum}}{a numeric vector}
    \item{\code{Gymnocolea_inflata}}{a numeric vector}
    \item{\code{Gypsophila_repens}}{a numeric vector}
    \item{\code{Hedera_helix}}{a numeric vector}
    \item{\code{Hedysarum_hedysaroides}}{a numeric vector}
    \item{\code{Helianthemum_nummularium}}{a numeric vector}
    \item{\code{Helictotrichon_pratense}}{a numeric vector}
    \item{\code{Helictotrichon_pubescens}}{a numeric vector}
    \item{\code{Helictotrichon_versicolor}}{a numeric vector}
    \item{\code{Heracleum_mantegazzianum}}{a numeric vector}
    \item{\code{Heracleum_sphondylium}}{a numeric vector}
    \item{\code{Herminium_monorchis}}{a numeric vector}
    \item{\code{Herzogiella_seligeri}}{a numeric vector}
    \item{\code{Heterocladium_dimorphum}}{a numeric vector}
    \item{\code{Hieracium_alpinum}}{a numeric vector}
    \item{\code{Hieracium_angustifolium}}{a numeric vector}
    \item{\code{Hieracium_aurantiacum}}{a numeric vector}
    \item{\code{Hieracium_bifidum}}{a numeric vector}
    \item{\code{Hieracium_hoppeanum}}{a numeric vector}
    \item{\code{Hieracium_jurassicum}}{a numeric vector}
    \item{\code{Hieracium_lachenalii}}{a numeric vector}
    \item{\code{Hieracium_lactucella}}{a numeric vector}
    \item{\code{Hieracium_murorum}}{a numeric vector}
    \item{\code{Hieracium_piliferum}}{a numeric vector}
    \item{\code{Hieracium_pilosella}}{a numeric vector}
    \item{\code{Hieracium_piloselloides}}{a numeric vector}
    \item{\code{Hieracium_prenanthoides}}{a numeric vector}
    \item{\code{Hieracium_sabaudum}}{a numeric vector}
    \item{\code{Hieracium_umbellatum}}{a numeric vector}
    \item{\code{Hieracium_villosum}}{a numeric vector}
    \item{\code{Hippocrepis_comosa}}{a numeric vector}
    \item{\code{Hippuris_vulgaris}}{a numeric vector}
    \item{\code{Holcus_lanatus}}{a numeric vector}
    \item{\code{Holcus_mollis}}{a numeric vector}
    \item{\code{Homalothecium_lutescens}}{a numeric vector}
    \item{\code{Homogyne_alpina}}{a numeric vector}
    \item{\code{Hordelymus_europaeus}}{a numeric vector}
    \item{\code{Hordeum_vulgare}}{a numeric vector}
    \item{\code{Humulus_lupulus}}{a numeric vector}
    \item{\code{Huperzia_selago}}{a numeric vector}
    \item{\code{Hydrocharis_morsus.ranae}}{a numeric vector}
    \item{\code{Hydrocotyle_vulgaris}}{a numeric vector}
    \item{\code{Hygrohypnum_luridum}}{a numeric vector}
    \item{\code{Hylocomium_brevirostre}}{a numeric vector}
    \item{\code{Hylocomium_pyrenaicum}}{a numeric vector}
    \item{\code{Hylocomium_splendens}}{a numeric vector}
    \item{\code{Hylocomium_umbratum}}{a numeric vector}
    \item{\code{Hypericum_maculatum}}{a numeric vector}
    \item{\code{Hypericum_perforatum}}{a numeric vector}
    \item{\code{Hypericum_tetrapterum}}{a numeric vector}
    \item{\code{Hypericum_x}}{a numeric vector}
    \item{\code{Hypnum_callichroum}}{a numeric vector}
    \item{\code{Hypnum_cupressiforme}}{a numeric vector}
    \item{\code{Hypnum_lindbergii}}{a numeric vector}
    \item{\code{Hypnum_pratense}}{a numeric vector}
    \item{\code{Hypochaeris_radicata}}{a numeric vector}
    \item{\code{Hypochaeris_uniflora}}{a numeric vector}
    \item{\code{Impatiens_parviflora}}{a numeric vector}
    \item{\code{Inula_helvetica}}{a numeric vector}
    \item{\code{Inula_salicina}}{a numeric vector}
    \item{\code{Iris_pseudacorus}}{a numeric vector}
    \item{\code{Isothecium_alopecuroides}}{a numeric vector}
    \item{\code{Juglans_regia}}{a numeric vector}
    \item{\code{Juncus_acutiflorus}}{a numeric vector}
    \item{\code{Juncus_alpinoarticulatus}}{a numeric vector}
    \item{\code{Juncus_arcticus}}{a numeric vector}
    \item{\code{Juncus_articulatus}}{a numeric vector}
    \item{\code{Juncus_bufonius}}{a numeric vector}
    \item{\code{Juncus_bulbosus}}{a numeric vector}
    \item{\code{Juncus_compressus}}{a numeric vector}
    \item{\code{Juncus_conglomeratus}}{a numeric vector}
    \item{\code{Juncus_effusus}}{a numeric vector}
    \item{\code{Juncus_filiformis}}{a numeric vector}
    \item{\code{Juncus_inflexus}}{a numeric vector}
    \item{\code{Juncus_subnodulosus}}{a numeric vector}
    \item{\code{Juncus_tenuis}}{a numeric vector}
    \item{\code{Juncus_triglumis}}{a numeric vector}
    \item{\code{Jungermannia_atrovirens}}{a numeric vector}
    \item{\code{Juniperus_communis_subsp._communis}}{a numeric vector}
    \item{\code{Juniperus_communis_subsp._nana}}{a numeric vector}
    \item{\code{Knautia_arvensis}}{a numeric vector}
    \item{\code{Knautia_dipsacifolia_subsp._dipsacifolia}}{a numeric vector}
    \item{\code{Knautia_godetii}}{a numeric vector}
    \item{\code{Kobresia_simpliciuscula}}{a numeric vector}
    \item{\code{Koeleria_pyramidata}}{a numeric vector}
    \item{\code{Lactuca_serriola}}{a numeric vector}
    \item{\code{Lamium_album}}{a numeric vector}
    \item{\code{Lamium_galeobdolon}}{a numeric vector}
    \item{\code{Lamiastrum_galeobdolon}}{a numeric vector}
    \item{\code{Lamium_maculatum}}{a numeric vector}
    \item{\code{Larix_decidua}}{a numeric vector}
    \item{\code{Laserpitium_gaudinii}}{a numeric vector}
    \item{\code{Laserpitium_halleri}}{a numeric vector}
    \item{\code{Laserpitium_latifolium}}{a numeric vector}
    \item{\code{Lathyrus_pratensis}}{a numeric vector}
    \item{\code{Lathyrus_vernus_subsp._vernus}}{a numeric vector}
    \item{\code{Lemna_minor}}{a numeric vector}
    \item{\code{Lemna_trisulca}}{a numeric vector}
    \item{\code{Leontodon_autumnalis}}{a numeric vector}
    \item{\code{Leontodon_helveticus}}{a numeric vector}
    \item{\code{Leontodon_hispidus}}{a numeric vector}
    \item{\code{Lepidozia_reptans}}{a numeric vector}
    \item{\code{Lescuraea_saxicola}}{a numeric vector}
    \item{\code{Leucanthemopsis_alpina}}{a numeric vector}
    \item{\code{Leucanthemum_halleri}}{a numeric vector}
    \item{\code{Leucobryum_glaucum}}{a numeric vector}
    \item{\code{Ligusticum_mutellina}}{a numeric vector}
    \item{\code{Ligusticum_mutellinoides}}{a numeric vector}
    \item{\code{Ligustrum_vulgare}}{a numeric vector}
    \item{\code{Lilium_martagon}}{a numeric vector}
    \item{\code{Linaria_vulgaris}}{a numeric vector}
    \item{\code{Linnaea_borealis}}{a numeric vector}
    \item{\code{Linum_catharticum}}{a numeric vector}
    \item{\code{Listera_cordata}}{a numeric vector}
    \item{\code{Listera_ovata}}{a numeric vector}
    \item{\code{Loiseleuria_procumbens}}{a numeric vector}
    \item{\code{Lolium_multiflorum}}{a numeric vector}
    \item{\code{Lolium_perenne}}{a numeric vector}
    \item{\code{Lonicera_alpigena}}{a numeric vector}
    \item{\code{Lonicera_caerulea}}{a numeric vector}
    \item{\code{Lonicera_nigra}}{a numeric vector}
    \item{\code{Lonicera_xylosteum}}{a numeric vector}
    \item{\code{Lophocolea_bidentata}}{a numeric vector}
    \item{\code{Lophocolea_heterophylla}}{a numeric vector}
    \item{\code{Lophozia_bantriensis}}{a numeric vector}
    \item{\code{Lophozia_collaris}}{a numeric vector}
    \item{\code{Lophozia_incisa}}{a numeric vector}
    \item{\code{Lophozia_ventricosa}}{a numeric vector}
    \item{\code{Lophozia_obtusa}}{a numeric vector}
    \item{\code{Lophozia_wenzelii}}{a numeric vector}
    \item{\code{Lotus_corniculatus}}{a numeric vector}
    \item{\code{Lotus_pedunculatus}}{a numeric vector}
    \item{\code{Lunaria_annua}}{a numeric vector}
    \item{\code{Lunularia_cruciata}}{a numeric vector}
    \item{\code{Luzula_alpinopilosa}}{a numeric vector}
    \item{\code{Luzula_campestris}}{a numeric vector}
    \item{\code{Luzula_lutea}}{a numeric vector}
    \item{\code{Luzula_luzulina}}{a numeric vector}
    \item{\code{Luzula_luzuloides}}{a numeric vector}
    \item{\code{Luzula_multiflora}}{a numeric vector}
    \item{\code{Luzula_nemorosa}}{a numeric vector}
    \item{\code{Luzula_nivea}}{a numeric vector}
    \item{\code{Luzula_pilosa}}{a numeric vector}
    \item{\code{Luzula_spicata}}{a numeric vector}
    \item{\code{Luzula_sudetica}}{a numeric vector}
    \item{\code{Luzula_sylvatica}}{a numeric vector}
    \item{\code{Lycopodiella_inundata}}{a numeric vector}
    \item{\code{Lycopodium_annotinum}}{a numeric vector}
    \item{\code{Lycopodium_clavatum}}{a numeric vector}
    \item{\code{Lycopus_europaeus}}{a numeric vector}
    \item{\code{Lysimachia_nemorum}}{a numeric vector}
    \item{\code{Lysimachia_nummularia}}{a numeric vector}
    \item{\code{Lysimachia_vulgaris}}{a numeric vector}
    \item{\code{Lythrum_salicaria}}{a numeric vector}
    \item{\code{Maianthemum_bifolium}}{a numeric vector}
    \item{\code{Malaxis_monophyllos}}{a numeric vector}
    \item{\code{Malus_sylvestris}}{a numeric vector}
    \item{\code{Marchantia_polymorpha}}{a numeric vector}
    \item{\code{Marsupella_funckii}}{a numeric vector}
    \item{\code{Marsupella_sphacelata}}{a numeric vector}
    \item{\code{Matricaria_discoidea}}{a numeric vector}
    \item{\code{Medicago_lupulina}}{a numeric vector}
    \item{\code{Medicago_sativa}}{a numeric vector}
    \item{\code{Meesia_uliginosa}}{a numeric vector}
    \item{\code{Melampyrum_pratense}}{a numeric vector}
    \item{\code{Melampyrum_sylvaticum}}{a numeric vector}
    \item{\code{Melica_nutans}}{a numeric vector}
    \item{\code{Mentha_aquatica}}{a numeric vector}
    \item{\code{Mentha_arvensis}}{a numeric vector}
    \item{\code{Mentha_longifolia}}{a numeric vector}
    \item{\code{Menyanthes_trifoliata}}{a numeric vector}
    \item{\code{Minuartia_verna}}{a numeric vector}
    \item{\code{Mnium_hornum}}{a numeric vector}
    \item{\code{Mnium_marginatum}}{a numeric vector}
    \item{\code{Mnium_spinosum}}{a numeric vector}
    \item{\code{Mnium_spinulosum}}{a numeric vector}
    \item{\code{Moehringia_trinervia}}{a numeric vector}
    \item{\code{Molinia_caerulea}}{a numeric vector}
    \item{\code{Moneses_uniflora}}{a numeric vector}
    \item{\code{Mycelis_muralis}}{a numeric vector}
    \item{\code{Mylia_anomala}}{a numeric vector}
    \item{\code{Myosotis_sylvatica}}{a numeric vector}
    \item{\code{Myosotis_arvensis}}{a numeric vector}
    \item{\code{Myosotis_palustris}}{a numeric vector}
    \item{\code{Myosotis_decumbens}}{a numeric vector}
    \item{\code{Myosoton_aquaticum}}{a numeric vector}
    \item{\code{Myriophyllum_verticillatum}}{a numeric vector}
    \item{\code{Nardia_scalaris}}{a numeric vector}
    \item{\code{Nardus_stricta}}{a numeric vector}
    \item{\code{Nasturtium_officinale}}{a numeric vector}
    \item{\code{Neottia_nidus.avis}}{a numeric vector}
    \item{\code{Nigritella_rhellicani}}{a numeric vector}
    \item{\code{Nuphar_lutea}}{a numeric vector}
    \item{\code{Nymphaea_alba}}{a numeric vector}
    \item{\code{Nymphoides_peltata}}{a numeric vector}
    \item{\code{Odontoschisma_denudatum}}{a numeric vector}
    \item{\code{Oligotrichum_hercynicum}}{a numeric vector}
    \item{\code{Oncophorus_virens}}{a numeric vector}
    \item{\code{Ononis_repens}}{a numeric vector}
    \item{\code{Ononis_spinosa}}{a numeric vector}
    \item{\code{Ophioglossum_vulgatum}}{a numeric vector}
    \item{\code{Orchis_morio}}{a numeric vector}
    \item{\code{Orchis_palustris}}{a numeric vector}
    \item{\code{Orchis_ustulata}}{a numeric vector}
    \item{\code{Oreopteris_limbosperma}}{a numeric vector}
    \item{\code{Origanum_vulgare}}{a numeric vector}
    \item{\code{Orthilia_secunda}}{a numeric vector}
    \item{\code{Oxalis_acetosella}}{a numeric vector}
    \item{\code{Oxystegus_tenuirostris}}{a numeric vector}
    \item{\code{Oxytropis_campestris}}{a numeric vector}
    \item{\code{Paludella_squarrosa}}{a numeric vector}
    \item{\code{Paraleucobryum_albicans}}{a numeric vector}
    \item{\code{Paris_quadrifolia}}{a numeric vector}
    \item{\code{Parnassia_palustris}}{a numeric vector}
    \item{\code{Pastinaca_sativa}}{a numeric vector}
    \item{\code{Pedicularis_foliosa}}{a numeric vector}
    \item{\code{Pedicularis_oederi}}{a numeric vector}
    \item{\code{Pedicularis_palustris}}{a numeric vector}
    \item{\code{Pedicularis_recutita}}{a numeric vector}
    \item{\code{Pedicularis_rostratocapitata}}{a numeric vector}
    \item{\code{Pedicularis_rostratospicata}}{a numeric vector}
    \item{\code{Pedicularis_sylvatica}}{a numeric vector}
    \item{\code{Pedicularis_tuberosa}}{a numeric vector}
    \item{\code{Pedicularis_verticillata}}{a numeric vector}
    \item{\code{Pellia_endiviifolia}}{a numeric vector}
    \item{\code{Pellia_epiphylla}}{a numeric vector}
    \item{\code{Pellia_neesiana}}{a numeric vector}
    \item{\code{Petasites_albus}}{a numeric vector}
    \item{\code{Petasites_hybridus}}{a numeric vector}
    \item{\code{Petasites_paradoxus}}{a numeric vector}
    \item{\code{Peucedanum_ostruthium}}{a numeric vector}
    \item{\code{Peucedanum_palustre}}{a numeric vector}
    \item{\code{Phalaris_arundinacea}}{a numeric vector}
    \item{\code{Phegopteris_connectilis}}{a numeric vector}
    \item{\code{Philonotis_calcarea}}{a numeric vector}
    \item{\code{Philonotis_fontana}}{a numeric vector}
    \item{\code{Philonotis_seriata}}{a numeric vector}
    \item{\code{Philonotis_tomentella}}{a numeric vector}
    \item{\code{Phleum_alpinum}}{a numeric vector}
    \item{\code{Phleum_pratense}}{a numeric vector}
    \item{\code{Phleum_hirsutum}}{a numeric vector}
    \item{\code{Phragmites_australis}}{a numeric vector}
    \item{\code{Phyteuma_betonicifolium}}{a numeric vector}
    \item{\code{Phyteuma_hemisphaericum}}{a numeric vector}
    \item{\code{Phyteuma_orbiculare}}{a numeric vector}
    \item{\code{Phyteuma_ovatum}}{a numeric vector}
    \item{\code{Phyteuma_spicatum}}{a numeric vector}
    \item{\code{Picea_abies}}{a numeric vector}
    \item{\code{Pimpinella_major}}{a numeric vector}
    \item{\code{Pimpinella_saxifraga}}{a numeric vector}
    \item{\code{Pinguicula_alpina}}{a numeric vector}
    \item{\code{Pinguicula_leptoceras}}{a numeric vector}
    \item{\code{Pinguicula_vulgaris}}{a numeric vector}
    \item{\code{Pinus_cembra}}{a numeric vector}
    \item{\code{Pinus_mugo}}{a numeric vector}
    \item{\code{Pinus_nigra}}{a numeric vector}
    \item{\code{Pinus_strobus}}{a numeric vector}
    \item{\code{Pinus_sylvestris}}{a numeric vector}
    \item{\code{Plagiochila_asplenioides}}{a numeric vector}
    \item{\code{Plagiomnium_affine}}{a numeric vector}
    \item{\code{Plagiomnium_undulatum}}{a numeric vector}
    \item{\code{Plagiothecium_laetum}}{a numeric vector}
    \item{\code{Plagiothecium_denticulatum}}{a numeric vector}
    \item{\code{Plagiothecium_latebricola}}{a numeric vector}
    \item{\code{Plagiothecium_nemorale}}{a numeric vector}
    \item{\code{Plagiothecium_platyphyllum}}{a numeric vector}
    \item{\code{Plagiothecium_undulatum}}{a numeric vector}
    \item{\code{Plantago_alpina}}{a numeric vector}
    \item{\code{Plantago_atrata}}{a numeric vector}
    \item{\code{Plantago_lanceolata}}{a numeric vector}
    \item{\code{Plantago_major}}{a numeric vector}
    \item{\code{Plantago_media}}{a numeric vector}
    \item{\code{Plantago_serpentina}}{a numeric vector}
    \item{\code{Platanthera_bifolia}}{a numeric vector}
    \item{\code{Platanthera_chlorantha}}{a numeric vector}
    \item{\code{Pleurozium_schreberi}}{a numeric vector}
    \item{\code{Poa_alpina}}{a numeric vector}
    \item{\code{Poa_annua}}{a numeric vector}
    \item{\code{Poa_chaixii}}{a numeric vector}
    \item{\code{Poa_nemoralis}}{a numeric vector}
    \item{\code{Poa_hybrida}}{a numeric vector}
    \item{\code{Poa_palustris}}{a numeric vector}
    \item{\code{Poa_pratensis}}{a numeric vector}
    \item{\code{Poa_trivialis_subsp._trivialis}}{a numeric vector}
    \item{\code{Poa_violacea}}{a numeric vector}
    \item{\code{Pogonatum_urnigerum}}{a numeric vector}
    \item{\code{Pohlia_camptotrachela}}{a numeric vector}
    \item{\code{Pohlia_elongata}}{a numeric vector}
    \item{\code{Pohlia_nutans}}{a numeric vector}
    \item{\code{Pohlia_wahlenbergii}}{a numeric vector}
    \item{\code{Polemonium_caeruleum}}{a numeric vector}
    \item{\code{Polygala_alpestris}}{a numeric vector}
    \item{\code{Polygala_alpina}}{a numeric vector}
    \item{\code{Polygala_amarella}}{a numeric vector}
    \item{\code{Polygala_chamaebuxus}}{a numeric vector}
    \item{\code{Polygala_serpyllifolia}}{a numeric vector}
    \item{\code{Polygala_vulgaris}}{a numeric vector}
    \item{\code{Polygonatum_verticillatum}}{a numeric vector}
    \item{\code{Polygonum_amphibium}}{a numeric vector}
    \item{\code{Polygonum_aviculare}}{a numeric vector}
    \item{\code{Polygonum_bistorta}}{a numeric vector}
    \item{\code{Polygonum_persicaria}}{a numeric vector}
    \item{\code{Polygonum_viviparum}}{a numeric vector}
    \item{\code{Polypodium_vulgare}}{a numeric vector}
    \item{\code{Polystichum_aculeatum}}{a numeric vector}
    \item{\code{Polystichum_lonchitis}}{a numeric vector}
    \item{\code{Polytrichum_alpinum}}{a numeric vector}
    \item{\code{Polytrichum_commune}}{a numeric vector}
    \item{\code{Polytrichum_formosum}}{a numeric vector}
    \item{\code{Polytrichum_juniperinum}}{a numeric vector}
    \item{\code{Polytrichum_longisetum}}{a numeric vector}
    \item{\code{Polytrichum_piliferum}}{a numeric vector}
    \item{\code{Populus_alba}}{a numeric vector}
    \item{\code{Populus_nigra}}{a numeric vector}
    \item{\code{Populus_tremula}}{a numeric vector}
    \item{\code{Potamogeton_gramineus}}{a numeric vector}
    \item{\code{Potamogeton_helveticus}}{a numeric vector}
    \item{\code{Potamogeton_natans}}{a numeric vector}
    \item{\code{Potamogeton_polygonifolius}}{a numeric vector}
    \item{\code{Potentilla_anserina}}{a numeric vector}
    \item{\code{Potentilla_aurea}}{a numeric vector}
    \item{\code{Potentilla_crantzii}}{a numeric vector}
    \item{\code{Potentilla_erecta}}{a numeric vector}
    \item{\code{Potentilla_grandiflora}}{a numeric vector}
    \item{\code{Potentilla_verna}}{a numeric vector}
    \item{\code{Potentilla_palustris}}{a numeric vector}
    \item{\code{Potentilla_reptans}}{a numeric vector}
    \item{\code{Potentilla_sterilis}}{a numeric vector}
    \item{\code{Preissia_quadrata}}{a numeric vector}
    \item{\code{Prenanthes_purpurea}}{a numeric vector}
    \item{\code{Primula_elatior}}{a numeric vector}
    \item{\code{Primula_farinosa}}{a numeric vector}
    \item{\code{Primula_integrifolia}}{a numeric vector}
    \item{\code{Primula_veris}}{a numeric vector}
    \item{\code{Prunella_grandiflora}}{a numeric vector}
    \item{\code{Prunella_vulgaris}}{a numeric vector}
    \item{\code{Prunus_avium}}{a numeric vector}
    \item{\code{Prunus_padus}}{a numeric vector}
    \item{\code{Prunus_spinosa}}{a numeric vector}
    \item{\code{Pseudephemerum_nitidum}}{a numeric vector}
    \item{\code{Pseudoleskea_incurvata}}{a numeric vector}
    \item{\code{Pseudoleskea_plicata}}{a numeric vector}
    \item{\code{Pseudoleskea_radicosa}}{a numeric vector}
    \item{\code{Pseudorchis_albida}}{a numeric vector}
    \item{\code{Pteridium_aquilinum}}{a numeric vector}
    \item{\code{Pterigynandrum_filiforme}}{a numeric vector}
    \item{\code{Ptilidium_ciliare}}{a numeric vector}
    \item{\code{Ptilium_crista.castrensis}}{a numeric vector}
    \item{\code{Pulicaria_dysenterica}}{a numeric vector}
    \item{\code{Pulmonaria_mollis}}{a numeric vector}
    \item{\code{Pulsatilla_alpina}}{a numeric vector}
    \item{\code{Pulsatilla_vernalis}}{a numeric vector}
    \item{\code{Pyrola_media}}{a numeric vector}
    \item{\code{Pyrola_minor}}{a numeric vector}
    \item{\code{Pyrola_rotundifolia}}{a numeric vector}
    \item{\code{Quercus_robur}}{a numeric vector}
    \item{\code{Racomitrium_canescens}}{a numeric vector}
    \item{\code{Racomitrium_heterostichum}}{a numeric vector}
    \item{\code{Radula_complanata}}{a numeric vector}
    \item{\code{Ranunculus_aconitifolius}}{a numeric vector}
    \item{\code{Ranunculus_acris}}{a numeric vector}
    \item{\code{Ranunculus_alpestris}}{a numeric vector}
    \item{\code{Ranunculus_auricomus}}{a numeric vector}
    \item{\code{Ranunculus_bulbosus}}{a numeric vector}
    \item{\code{Ranunculus_ficaria}}{a numeric vector}
    \item{\code{Ranunculus_flammula}}{a numeric vector}
    \item{\code{Ranunculus_lanuginosus}}{a numeric vector}
    \item{\code{Ranunculus_montanus}}{a numeric vector}
    \item{\code{Ranunculus_nemorosus}}{a numeric vector}
    \item{\code{Ranunculus_repens}}{a numeric vector}
    \item{\code{Ranunculus_reptans}}{a numeric vector}
    \item{\code{Ranunculus_sardous}}{a numeric vector}
    \item{\code{Ranunculus_sceleratus}}{a numeric vector}
    \item{\code{Ranunculus_trichophyllus}}{a numeric vector}
    \item{\code{Reynoutria_japonica}}{a numeric vector}
    \item{\code{Rhamnus_cathartica}}{a numeric vector}
    \item{\code{Rhinanthus_alectorolophus}}{a numeric vector}
    \item{\code{Rhinanthus_glaber}}{a numeric vector}
    \item{\code{Rhinanthus_glacialis}}{a numeric vector}
    \item{\code{Rhinanthus_minor}}{a numeric vector}
    \item{\code{Rhizomnium_magnifolium}}{a numeric vector}
    \item{\code{Rhizomnium_pseudopunctatum}}{a numeric vector}
    \item{\code{Rhizomnium_punctatum}}{a numeric vector}
    \item{\code{Rhodobryum_roseum}}{a numeric vector}
    \item{\code{Rhododendron_ferrugineum}}{a numeric vector}
    \item{\code{Rhododendron_hirsutum}}{a numeric vector}
    \item{\code{Rhynchospora_alba}}{a numeric vector}
    \item{\code{Rhynchospora_fusca}}{a numeric vector}
    \item{\code{Rhynchostegium_megapolitanum}}{a numeric vector}
    \item{\code{Rhynchostegium_riparioides}}{a numeric vector}
    \item{\code{Rhytidiadelphus_loreus}}{a numeric vector}
    \item{\code{Rhytidiadelphus_squarrosus}}{a numeric vector}
    \item{\code{Rhytidiadelphus_triquetrus}}{a numeric vector}
    \item{\code{Rhytidium_rugosum}}{a numeric vector}
    \item{\code{Ribes_alpinum}}{a numeric vector}
    \item{\code{Ribes_nigrum}}{a numeric vector}
    \item{\code{Ribes_rubrum}}{a numeric vector}
    \item{\code{Riccardia_chamaedryfolia}}{a numeric vector}
    \item{\code{Riccardia_multifida}}{a numeric vector}
    \item{\code{Riccardia_palmata}}{a numeric vector}
    \item{\code{Riccia_fluitans}}{a numeric vector}
    \item{\code{Rorippa_islandica}}{a numeric vector}
    \item{\code{Rorippa_sylvestris}}{a numeric vector}
    \item{\code{Rorippa_x}}{a numeric vector}
    \item{\code{Rosa_arvensis}}{a numeric vector}
    \item{\code{Rosa_canina}}{a numeric vector}
    \item{\code{Rosa_majalis}}{a numeric vector}
    \item{\code{Rosa_pendulina}}{a numeric vector}
    \item{\code{Rosa_rubiginosa}}{a numeric vector}
    \item{\code{Rubus_caesius}}{a numeric vector}
    \item{\code{Rubus_fruticosus}}{a numeric vector}
    \item{\code{Rubus_idaeus}}{a numeric vector}
    \item{\code{Rubus_saxatilis}}{a numeric vector}
    \item{\code{Rumex_acetosa}}{a numeric vector}
    \item{\code{Rumex_acetosella}}{a numeric vector}
    \item{\code{Rumex_alpestris}}{a numeric vector}
    \item{\code{Rumex_alpinus}}{a numeric vector}
    \item{\code{Rumex_crispus}}{a numeric vector}
    \item{\code{Rumex_obtusifolius}}{a numeric vector}
    \item{\code{Rumex_scutatus}}{a numeric vector}
    \item{\code{Sagina_procumbens}}{a numeric vector}
    \item{\code{Sagina_saginoides}}{a numeric vector}
    \item{\code{Salix_alba}}{a numeric vector}
    \item{\code{Salix_alba.2}}{a numeric vector}
    \item{\code{Salix_appendiculata}}{a numeric vector}
    \item{\code{Salix_aurita}}{a numeric vector}
    \item{\code{Salix_caprea}}{a numeric vector}
    \item{\code{Salix_caprea.1}}{a numeric vector}
    \item{\code{Salix_caprea.2}}{a numeric vector}
    \item{\code{Salix_cinerea}}{a numeric vector}
    \item{\code{Salix_elaeagnos}}{a numeric vector}
    \item{\code{Salix_foetida}}{a numeric vector}
    \item{\code{Salix_fragilis}}{a numeric vector}
    \item{\code{Salix_glaucosericea}}{a numeric vector}
    \item{\code{Salix_hastata}}{a numeric vector}
    \item{\code{Salix_helvetica}}{a numeric vector}
    \item{\code{Salix_herbacea}}{a numeric vector}
    \item{\code{Salix_myrsinifolia}}{a numeric vector}
    \item{\code{Salix_pentandra}}{a numeric vector}
    \item{\code{Salix_purpurea}}{a numeric vector}
    \item{\code{Salix_repens}}{a numeric vector}
    \item{\code{Salix_reticulata}}{a numeric vector}
    \item{\code{Salix_retusa}}{a numeric vector}
    \item{\code{Salix_rosmarinifolia}}{a numeric vector}
    \item{\code{Salix_triandra}}{a numeric vector}
    \item{\code{Salix_viminalis}}{a numeric vector}
    \item{\code{Salix_waldsteiniana}}{a numeric vector}
    \item{\code{Salvia_glutinosa}}{a numeric vector}
    \item{\code{Salvia_pratensis}}{a numeric vector}
    \item{\code{Sambucus_nigra}}{a numeric vector}
    \item{\code{Sambucus_racemosa}}{a numeric vector}
    \item{\code{Sanguisorba_minor}}{a numeric vector}
    \item{\code{Sanguisorba_officinalis}}{a numeric vector}
    \item{\code{Sanicula_europaea}}{a numeric vector}
    \item{\code{Satureja_montana}}{a numeric vector}
    \item{\code{Saussurea_alpina}}{a numeric vector}
    \item{\code{Saxifraga_aizoides}}{a numeric vector}
    \item{\code{Saxifraga_cuneifolia}}{a numeric vector}
    \item{\code{Saxifraga_oppositifolia}}{a numeric vector}
    \item{\code{Saxifraga_paniculata}}{a numeric vector}
    \item{\code{Saxifraga_rotundifolia}}{a numeric vector}
    \item{\code{Saxifraga_stellaris}}{a numeric vector}
    \item{\code{Scabiosa_columbaria}}{a numeric vector}
    \item{\code{Scapania_aequiloba}}{a numeric vector}
    \item{\code{Scapania_irrigua}}{a numeric vector}
    \item{\code{Scapania_paludicola}}{a numeric vector}
    \item{\code{Scapania_uliginosa}}{a numeric vector}
    \item{\code{Scapania_subalpina}}{a numeric vector}
    \item{\code{Scapania_umbrosa}}{a numeric vector}
    \item{\code{Scapania_undulata}}{a numeric vector}
    \item{\code{Scheuchzeria_palustris}}{a numeric vector}
    \item{\code{Schistidium_apocarpum}}{a numeric vector}
    \item{\code{Schoenoplectus_lacustris}}{a numeric vector}
    \item{\code{Schoenoplectus_tabernaemontani}}{a numeric vector}
    \item{\code{Schoenus_ferrugineus}}{a numeric vector}
    \item{\code{Schoenus_nigricans}}{a numeric vector}
    \item{\code{Scirpus_sylvaticus}}{a numeric vector}
    \item{\code{Scleropodium_purum}}{a numeric vector}
    \item{\code{Scorpidium_scorpioides}}{a numeric vector}
    \item{\code{Scrophularia_nodosa}}{a numeric vector}
    \item{\code{Scrophularia_umbrosa}}{a numeric vector}
    \item{\code{Scutellaria_galericulata}}{a numeric vector}
    \item{\code{Sedum_acre}}{a numeric vector}
    \item{\code{Sedum_album}}{a numeric vector}
    \item{\code{Sedum_alpestre}}{a numeric vector}
    \item{\code{Sedum_annuum}}{a numeric vector}
    \item{\code{Sedum_atratum}}{a numeric vector}
    \item{\code{Sedum_telephium}}{a numeric vector}
    \item{\code{Selaginella_selaginoides}}{a numeric vector}
    \item{\code{Seligeria_trifaria}}{a numeric vector}
    \item{\code{Sempervivum_arachnoideum}}{a numeric vector}
    \item{\code{Sempervivum_montanum}}{a numeric vector}
    \item{\code{Senecio_alpinus}}{a numeric vector}
    \item{\code{Senecio_aquaticus}}{a numeric vector}
    \item{\code{Senecio_doronicum}}{a numeric vector}
    \item{\code{Senecio_hercynicus}}{a numeric vector}
    \item{\code{Senecio_ovatus}}{a numeric vector}
    \item{\code{Senecio_paludosus}}{a numeric vector}
    \item{\code{Serratula_tinctoria}}{a numeric vector}
    \item{\code{Sesleria_caerulea}}{a numeric vector}
    \item{\code{Setaria_viridis}}{a numeric vector}
    \item{\code{Sibbaldia_procumbens}}{a numeric vector}
    \item{\code{Silaum_silaus}}{a numeric vector}
    \item{\code{Silene_acaulis}}{a numeric vector}
    \item{\code{Silene_dioica}}{a numeric vector}
    \item{\code{Silene_exscapa}}{a numeric vector}
    \item{\code{Silene_flos.cuculi}}{a numeric vector}
    \item{\code{Silene_nutans_subsp._nutans}}{a numeric vector}
    \item{\code{Silene_pratensis}}{a numeric vector}
    \item{\code{Silene_rupestris}}{a numeric vector}
    \item{\code{Silene_vulgaris}}{a numeric vector}
    \item{\code{Sinapis_arvensis}}{a numeric vector}
    \item{\code{Solanum_dulcamara}}{a numeric vector}
    \item{\code{Soldanella_pusilla}}{a numeric vector}
    \item{\code{Soldanella_alpina}}{a numeric vector}
    \item{\code{Solidago_canadensis}}{a numeric vector}
    \item{\code{Solidago_gigantea}}{a numeric vector}
    \item{\code{Solidago_virgaurea}}{a numeric vector}
    \item{\code{Sonchus_arvensis}}{a numeric vector}
    \item{\code{Sonchus_oleraceus}}{a numeric vector}
    \item{\code{Sorbus_aria}}{a numeric vector}
    \item{\code{Sorbus_aucuparia}}{a numeric vector}
    \item{\code{Sorbus_chamaemespilus}}{a numeric vector}
    \item{\code{Sorbus_mougeotii}}{a numeric vector}
    \item{\code{Sparganium_erectum}}{a numeric vector}
    \item{\code{Sphagnum_affine}}{a numeric vector}
    \item{\code{Sphagnum_capillifolium}}{a numeric vector}
    \item{\code{Sphagnum_palustre}}{a numeric vector}
    \item{\code{Sphagnum_compactum}}{a numeric vector}
    \item{\code{Sphagnum_contortum}}{a numeric vector}
    \item{\code{Sphagnum_cuspidatum}}{a numeric vector}
    \item{\code{Sphagnum_denticulatum}}{a numeric vector}
    \item{\code{Sphagnum_fimbriatum}}{a numeric vector}
    \item{\code{Sphagnum_fuscum}}{a numeric vector}
    \item{\code{Sphagnum_girgensohnii}}{a numeric vector}
    \item{\code{Sphagnum_magellanicum}}{a numeric vector}
    \item{\code{Sphagnum_majus}}{a numeric vector}
    \item{\code{Sphagnum_papillosum}}{a numeric vector}
    \item{\code{Sphagnum_platyphyllum}}{a numeric vector}
    \item{\code{Sphagnum_quinquefarium}}{a numeric vector}
    \item{\code{Sphagnum_recurvum}}{a numeric vector}
    \item{\code{Sphagnum_rubellum}}{a numeric vector}
    \item{\code{Sphagnum_russowii}}{a numeric vector}
    \item{\code{Sphagnum_squarrosum}}{a numeric vector}
    \item{\code{Sphagnum_subnitens}}{a numeric vector}
    \item{\code{Sphagnum_subsecundum}}{a numeric vector}
    \item{\code{Sphagnum_tenellum}}{a numeric vector}
    \item{\code{Sphagnum_teres}}{a numeric vector}
    \item{\code{Sphagnum_warnstorfii}}{a numeric vector}
    \item{\code{Spirodela_polyrhiza}}{a numeric vector}
    \item{\code{Splachnum_ampullaceum}}{a numeric vector}
    \item{\code{Splachnum_sphaericum}}{a numeric vector}
    \item{\code{Stachys_officinalis_subsp._officinalis}}{a numeric vector}
    \item{\code{Stachys_palustris}}{a numeric vector}
    \item{\code{Stachys_sylvatica}}{a numeric vector}
    \item{\code{Stellaria_alsine}}{a numeric vector}
    \item{\code{Stellaria_graminea}}{a numeric vector}
    \item{\code{Stellaria_media}}{a numeric vector}
    \item{\code{Stellaria_nemorum}}{a numeric vector}
    \item{\code{Streptopus_amplexifolius}}{a numeric vector}
    \item{\code{Succisa_pratensis}}{a numeric vector}
    \item{\code{Swertia_perennis}}{a numeric vector}
    \item{\code{Symphytum_officinale}}{a numeric vector}
    \item{\code{Tamus_communis}}{a numeric vector}
    \item{\code{Tanacetum_vulgare}}{a numeric vector}
    \item{\code{Taraxacum_officinale}}{a numeric vector}
    \item{\code{Taraxacum_palustre}}{a numeric vector}
    \item{\code{Taraxacum_schroeterianum}}{a numeric vector}
    \item{\code{Tayloria_serrata}}{a numeric vector}
    \item{\code{Tayloria_splachnoides}}{a numeric vector}
    \item{\code{Tephroseris_helenitis}}{a numeric vector}
    \item{\code{Tetraphis_pellucida}}{a numeric vector}
    \item{\code{Tetraplodon_angustatus}}{a numeric vector}
    \item{\code{Thalictrum_aquilegiifolium}}{a numeric vector}
    \item{\code{Thalictrum_flavum}}{a numeric vector}
    \item{\code{Thelypteris_palustris}}{a numeric vector}
    \item{\code{Thesium_alpinum}}{a numeric vector}
    \item{\code{Thesium_pyrenaicum}}{a numeric vector}
    \item{\code{Thlaspi_perfoliatum}}{a numeric vector}
    \item{\code{Thuidium_abietinum}}{a numeric vector}
    \item{\code{Thuidium_recognitum}}{a numeric vector}
    \item{\code{Thuidium_tamariscinum}}{a numeric vector}
    \item{\code{Thymus_serpyllum}}{a numeric vector}
    \item{\code{Tilia_platyphyllos}}{a numeric vector}
    \item{\code{Tofieldia_calyculata}}{a numeric vector}
    \item{\code{Tomentypnum_nitens}}{a numeric vector}
    \item{\code{Tortella_tortuosa}}{a numeric vector}
    \item{\code{Tortella_fragilis}}{a numeric vector}
    \item{\code{Tortella_inclinata}}{a numeric vector}
    \item{\code{Tortula_ruralis}}{a numeric vector}
    \item{\code{Tozzia_alpina}}{a numeric vector}
    \item{\code{Tragopogon_dubius}}{a numeric vector}
    \item{\code{Tragopogon_pratensis}}{a numeric vector}
    \item{\code{Traunsteinera_globosa}}{a numeric vector}
    \item{\code{Trichophorum_alpinum}}{a numeric vector}
    \item{\code{Trichophorum_cespitosum}}{a numeric vector}
    \item{\code{Trifolium_alpestre}}{a numeric vector}
    \item{\code{Trifolium_alpinum}}{a numeric vector}
    \item{\code{Trifolium_badium}}{a numeric vector}
    \item{\code{Trifolium_dubium}}{a numeric vector}
    \item{\code{Trifolium_hybridum}}{a numeric vector}
    \item{\code{Trifolium_medium}}{a numeric vector}
    \item{\code{Trifolium_montanum}}{a numeric vector}
    \item{\code{Trifolium_pallescens}}{a numeric vector}
    \item{\code{Trifolium_pratense_subsp._nivale}}{a numeric vector}
    \item{\code{Trifolium_pratense_subsp._pratense}}{a numeric vector}
    \item{\code{Trifolium_repens_subsp._repens}}{a numeric vector}
    \item{\code{Trifolium_spadiceum}}{a numeric vector}
    \item{\code{Trifolium_thalii}}{a numeric vector}
    \item{\code{Triglochin_palustris}}{a numeric vector}
    \item{\code{Trisetum_flavescens}}{a numeric vector}
    \item{\code{Tritomaria_polita}}{a numeric vector}
    \item{\code{Tritomaria_quinquedentata}}{a numeric vector}
    \item{\code{Trollius_europaeus}}{a numeric vector}
    \item{\code{Tussilago_farfara}}{a numeric vector}
    \item{\code{Typha_angustifolia}}{a numeric vector}
    \item{\code{Typha_latifolia}}{a numeric vector}
    \item{\code{Ulmus_glabra}}{a numeric vector}
    \item{\code{Urtica_dioica}}{a numeric vector}
    \item{\code{Utricularia_australis}}{a numeric vector}
    \item{\code{Utricularia_intermedia}}{a numeric vector}
    \item{\code{Utricularia_minor}}{a numeric vector}
    \item{\code{Utricularia_vulgaris}}{a numeric vector}
    \item{\code{Vaccinium_macrocarpon}}{a numeric vector}
    \item{\code{Vaccinium_microcarpum}}{a numeric vector}
    \item{\code{Vaccinium_myrtillus}}{a numeric vector}
    \item{\code{Vaccinium_oxycoccos}}{a numeric vector}
    \item{\code{Vaccinium_uliginosum}}{a numeric vector}
    \item{\code{Vaccinium_vitis.idaea}}{a numeric vector}
    \item{\code{Valeriana_dioica}}{a numeric vector}
    \item{\code{Valeriana_montana}}{a numeric vector}
    \item{\code{Valeriana_officinalis}}{a numeric vector}
    \item{\code{Valeriana_tripteris}}{a numeric vector}
    \item{\code{Veratrum_album}}{a numeric vector}
    \item{\code{Veronica_alpina}}{a numeric vector}
    \item{\code{Veronica_anagallis.aquatica}}{a numeric vector}
    \item{\code{Veronica_aphylla}}{a numeric vector}
    \item{\code{Veronica_arvensis}}{a numeric vector}
    \item{\code{Veronica_beccabunga}}{a numeric vector}
    \item{\code{Veronica_bellidioides}}{a numeric vector}
    \item{\code{Veronica_chamaedrys}}{a numeric vector}
    \item{\code{Veronica_filiformis}}{a numeric vector}
    \item{\code{Veronica_fruticans}}{a numeric vector}
    \item{\code{Veronica_fruticulosa}}{a numeric vector}
    \item{\code{Veronica_officinalis}}{a numeric vector}
    \item{\code{Veronica_persica}}{a numeric vector}
    \item{\code{Veronica_scutellata}}{a numeric vector}
    \item{\code{Veronica_serpyllifolia}}{a numeric vector}
    \item{\code{Veronica_urticifolia}}{a numeric vector}
    \item{\code{Viburnum_lantana}}{a numeric vector}
    \item{\code{Viburnum_opulus}}{a numeric vector}
    \item{\code{Vicia_cracca}}{a numeric vector}
    \item{\code{Vicia_sativa}}{a numeric vector}
    \item{\code{Vicia_sepium}}{a numeric vector}
    \item{\code{Vicia_sylvatica}}{a numeric vector}
    \item{\code{Viola_biflora}}{a numeric vector}
    \item{\code{Viola_palustris}}{a numeric vector}
    \item{\code{Viola_reichenbachiana}}{a numeric vector}
    \item{\code{Viola_riviniana}}{a numeric vector}
    \item{\code{Viola_tricolor}}{a numeric vector}
    \item{\code{Weissia_controversa}}{a numeric vector}
    \item{\code{Weissia_rostellat}}{a numeric vector}
  }
}
\source{
Graf, U., Wildi, O., Feldmeyer-Christe, E. &
Kuechler, M. 2010. A phytosociological classification of Swiss mire
vegetation. Botanica Helvetica 120: 1-13.
}
\references{
Wildi, O. 2013. Data Analysis in Vegetation Ecology. 2nd ed. Wiley-Blackwell,
Chichester.
}
\examples{
summary(wetveg)
}
\keyword{datasets}
