\name{wetveg}
\alias{wetveg}
\docType{data}
\title{
Random sample of Swiss wetland vegetation, vegetation information.
}
\description{
Random sample of Swiss wetland vegetation, vegetation information. Site data is in \code{\link{wetsit}}.
}
\usage{data(wetveg)}
\format{
A data frame with 1500 observations on 1164 variables, the species. A 0 to 4 step scales is used.
}
\source{
Graf, U., Wildi, O., Feldmeyer-Christe, E. &
Kuechler, M. 2010. A phytosociological classification of Swiss mire
vegetation. Botanica Helvetica 120: 1-13.
}
\references{
Wildi, O. 2013. Data Analysis in Vegetation Ecology. 2nd ed. Wiley-Blackwell,
Chichester.
}
\examples{
summary(wetveg)
}
\keyword{datasets}
