\name{ws500}
\alias{ws500}
\docType{data}
\title{
Four kilometre grid forest data of Switzerland, 500m2 plots
}
\description{
Four kilometre grid forest data of Switzerland, 500m2 plots. Also see ws30, ws200. Site information is in \code{\link{wssit}}.
}
\usage{data(ws500)}
\format{
  A data frame with 726 observations on 1262 variables, the species.
}
\details{
See object wssit for corresponding site information
}
\source{
Wohlgemuth, T., Moser B., Braendli, U.-B., Kull, P. & Schuetz, M. 2008. 
Diversity of forest plant species at the community and landscape scales in Switzerland. 
Plant Biosystems 142: 604-613.
}
\references{
Wildi, O. 2013. Data Analysis in Vegetation Ecology. 2nd ed. Wiley-Blackwell,
Chichester.
}
\examples{
summary(ws500)
}
\keyword{datasets}
